% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildrawDatadt.R
\name{BuildrawDatadt}
\alias{BuildrawDatadt}
\title{Constructor of objects of class \linkS4class{rawDatadt}.}
\usage{
BuildrawDatadt(key, value)
}
\arguments{
\item{key}{object of class \linkS4class{rawKey}.}

\item{value}{Character vector with the values of the variables in key parameter.}
}
\value{
An object of class \linkS4class{rawDatadt}.
}
\description{
This constructor returns objects of class \linkS4class{rawDatadt}.
The input parameters are an object of class \linkS4class{rawKey} and a character \code{vector}.
}
\examples{

library(data.table)
key <- new(Class = 'rawKey', 
          data.table(IDDDKey = c('Employees', 'Employees', 'RemEmployees', 'Turnover'),
                     QualKey = c('25641378SS2.1.1.', '25641378SS1.    ', '25641378SS    1.', 
                                 '25641378SS')))
value <- c('625', '954', '122', '105124')
rawData <- BuildrawDatadt(key, value)

}

