% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildrawStQList.R
\name{BuildrawStQList}
\alias{BuildrawStQList}
\title{Constructor of objects of class \linkS4class{rawStQList}.}
\usage{
BuildrawStQList(Data)
}
\arguments{
\item{Data}{\code{List} of objects of class \linkS4class{rawStQ}.}
}
\value{
An object of class \linkS4class{rawStQList}.
}
\description{
This constructor returns objects of class \linkS4class{rawStQList}.
The input parameter is a \code{list} of objects of class \linkS4class{rawStQ}.
}
\examples{
mm <- c(paste0('0', 1:9), 10:12)
TimePer <- paste0('MM', mm, '2015')
EmptyrawQ <- new(Class = 'rawStQ')
EmptyrawQList <- vector('list', 12)
EmptyrawQList <- lapply(EmptyrawQList, function(x) EmptyrawQ)
names(EmptyrawQList) <- TimePer

rawQList <- BuildrawStQList(EmptyrawQList)
rawQList
#Notice that it is indeed an object with complex structure:
str(rawQList)

}

