<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildStQList.R
||||||| merged common ancestors
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StQList.constructor.R
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StQList.constructor.R
>>>>>>> e3ac9308c1978eb314de9c1fda9971de40a9bce7
\name{BuildStQList}
\alias{BuildStQList}
\title{Constructor of objects of class \linkS4class{StQList}}
\usage{
BuildStQList(Data)
}
\arguments{
\item{Data}{\code{List} of objects of class \linkS4class{StQ}.}
}
\value{
An object of class \linkS4class{StQList}.
}
\description{
This constructor returns objects of class \linkS4class{StQList}.
THe input parameter is a \code{list} of objects of class \linkS4class{StQ}.
}
\examples{
mm <- c(paste0('0', 1:9), 10:12)
TimePer <- paste0('MM', mm, '2015')
EmptyQ <- new(Class = 'StQ')
EmptyQList <- vector('list', 12)
EmptyQList <- lapply(EmptyQList, function(x) EmptyQ)
names(EmptyQList) <- TimePer

QList <- BuildStQList(EmptyQList)
QList
#Notice that it is indeed an object with complex structure:
str(QList)

}

