% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatadtToDT.R
\name{DatadtToDT}
\alias{DatadtToDT}
\title{Transforms a \linkS4class{Datadt} into a \linkS4class{data.table}}
\usage{
DatadtToDT(object)
}
\arguments{
\item{object}{\linkS4class{Datadt} to be transformed into a \linkS4class{data.table}.}
}
\value{
\linkS4class{data.table}
}
\description{
\code{DatadtToDT} returns a \linkS4class{data.table} with the data and values of the
 input \linkS4class{Datadt} object.
}
\details{
This function is included because of the issues regarding the inheritance of class 
\linkS4class{data.table} (see e.g. this \href{http://stackoverflow.com/questions/32206351/how-to-use-data-table-as-super-class-in-s4}{SO entry}).
}
\examples{
library(data.table)
Ddt <- new(Class = 'Datadt', 
            data.table(ID = c('001', '001', '001', '001'), 
                       Market = c('0.', '1.', '', ''),
                       EsRemuner = c('2.2.','2.1.','',''),
                       IDDD = c('Turnover', 'Turnover', 'Province', 'NACE09'),
                       Value = c('625000', '23154', '04', '0512')))
class(Ddt)                        
DT <- DatadtToDT(Ddt)                        
class(DT)

data(ExampleDatadt)
class(ExampleDatadt)
DT <- DatadtToDT(ExampleDatadt)
class(DT)

}

