% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcast_Datadt.R
\name{dcast_Datadt}
\alias{dcast_Datadt}
\title{Convert an \linkS4class{Datadt} object into a dcasted data.table}
\usage{
dcast_Datadt(object, VarNames)
}
\arguments{
\item{object}{Objecto of class \linkS4class{Datadt}.}

\item{VarNames}{Character vector with names of the output variables.}
}
\value{
\linkS4class{data.table} with data of the input
\linkS4class{Datadt} object with statistical units by rows and variables by
 columns. Only variables in \code{VarNames} will be output. If no variable
 name is specified, all variables in the input object will be output.
}
\description{
\code{dcast_Datadt} returns a \linkS4class{data.table} in dcasted
form (observations by row and variables by columns) with data from the input
\linkS4class{Datadt} object.

This methods converts the input \code{Datadt} object
 with key-value pair structure into a \linkS4class{data.table} with
 statistical units by row and variables specified in the input parameter
 \code{VarNames} by columns.
}
\examples{
Ddt <- new(Class = 'Datadt', 
            data.table(ID = c('001', '001', '001', '001'), 
                       IsNatMarket = c('0', '1', '', ''),
                       EsRemuner = c('1','0','',''),
                       IDDD = c('Turnover', 'Turnover', 'Province', 'NACE09'),
                       Value = c('625000', '23154', '04', '0512')))
                       
VarNames <- c('Turnover_0_1')
dcast_Datadt(Ddt,VarNames)

VarNames <- c('Turnover')
dcast_Datadt(Ddt,VarNames)

dcast_Datadt(Ddt)

}

