% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD.R
\name{DD}
\alias{DD}
\title{S4 class with the data dictionary (variable specifications)}
\usage{
DD(VNC = BuildVNC(), ID = data.table(Variable = character(0), Sort =
  character(0), Class = character(0), Length = character(0), Qual1 =
  character(0), ValueRegExp = character(0)), MicroData = data.table(Variable =
  character(0), Sort = character(0), Class = character(0), Length =
  character(0), Qual1 = character(0), ValueRegExp = character(0)),
  ParaData = data.table(Variable = character(0), Sort = character(0), Class =
  character(0), Length = character(0), Qual1 = character(0), ValueRegExp =
  character(0)), Aggregates = data.table(Variable = character(0), Sort =
  character(0), Class = character(0), Length = character(0), Qual1 =
  character(0), ValueRegExp = character(0)), AggWeights = data.table(Variable
  = character(0), Sort = character(0), Class = character(0), Length =
  character(0), Qual1 = character(0), ValueRegExp = character(0)),
  Other = data.table(Variable = character(0), Sort = character(0), Class =
  character(0), Length = character(0), Qual1 = character(0), ValueRegExp =
  character(0)))
}
\description{
Definition of an S4 class named \code{DD} with the specification of each variable.

The class \code{DD} comprises a slot of class \linkS4class{VarNameCorresp} and slots of class 
\linkS4class{data.table} of names \code{ID}, \code{MicroData}, \code{ParaData}, 
\code{Aggregates}, \code{AggWeights}, \code{Other}.
}
\examples{
library(data.table)
### We build the VNC object
VarList <- list(ID = data.table(IDQual = c('NumIdEst', rep('', 4)),
                                NonIDQual = c('','','','',''),
                                IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                                NumIdEst = c('', rep('.', 4)),
                                UnitName = c('numidest', 'nombre', 'apellidos', 'direccion', 'telefono'),
                                InFiles = rep('FI', 5)),
MicroData = data.table(IDQual = c('NumIdEst', rep('', 2)),
                       NonIDQual = c('', 'Market', ''),
                       IDDD = c(rep('', 2), 'NewOrders'),
                       NumIdEst = c(rep('', 2), '.'),
                       Market = c(rep('', 2), '1'),
                       UnitName = c('numidest', '', 'cp09'),
                       InFiles = rep('FF, FD, FG', 3)),
ParaData = data.table(IDQual = c('NumIdEst', rep('', 2)),
                      NonIDQual = c('', 'Action', ''),
                      IDDD = c(rep('', 2), 'Date'),
                      NumIdEst = c(rep('', 2), '.'),
                      Action = c(rep('', 2), 'Imputation'),
                      UnitName = c('numidest', '', 'FechaImput'),
                      InFiles = rep('FP', 3)))

VNC <- BuildVNC(VarList)

IDdt <- data.table(Variable = c('NumIdEst', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                   Sort = c('IDQual', rep('IDDD', 4)),
                   Class = rep('character', 5),
                   Length = c('11', '15', '15', '20','9'),
                   Qual1 = c('', rep('NumIdEst', 4)),
                   ValueRegExp = c('[0-9]{9}PP', '.+', '.+', '.+', '(6|9)[0-9]{8}'))
Microdt <- data.table(Variable = c('NumIdEst', 'Market', 'NewOrders'),
                      Sort = c('IDQual', 'NonIDQual', 'IDDD'),
                      Class = c(rep('character', 2), 'numeric'),
                      Length = c('11', '2', '7'),
                      Qual1 = c(rep('', 2), 'NumIdEst'),
                      ValueRegExp = c('[0-9]{9}PP', '.+', '([0-9]{1, 10}| )'))
Paradt <- data.table(Variable = c('NumIdEst', 'Action', 'Date'),
                     Sort = c('IDQual', 'NonIDQual', 'IDDD'),
                     Class = rep('character', 3),
                     Length = c('11', '10', '10'),
                     Qual1 = c(rep('', 2), 'NumIdEst'),
                     Qual2 = c(rep('', 2), 'Action'),
                     ValueRegExp = c('[0-9]{9}PP', 'Collection|Editing|Imputation',
                                     '(([0-9]{2}-(0[1-9]|1(0-2))-[0-9]{4})| )'))

DD <- DD(VNC = VNC, ID = IDdt, MicroData = Microdt, ParaData = Paradt)
DD

}

