% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDdt-class.R
\docType{class}
\name{DDdt-class}
\alias{DDdt-class}
\title{S4 class for \linkS4class{data.table}s with columns containing the variable specifications}
\description{
Definition of an S4 class named \code{DDdt} with the properties of each statistical 
variable. This is a subclass of \linkS4class{data.table} used as slots in a \linkS4class{DD}
object.

The class \code{DDdt} is a \linkS4class{data.table} with the following columns (in this order):

\itemize{
 \item \code{Variable}: name of each variable.
 \item \code{Sort}: role of each variable (IDQual, NonIDQual, IDDD).
 \item \code{Class}: class of each variable (character, numeric, ...).
 \item \code{Length}: highest length for each variable (in terms of number of characters).
 \item One column of name [\emph{Qual}\code{j}] per each name of the qualifiers needed for each 
 variable.
 \item \code{ValueRegExp}: regexp for the value of each variable (not active yet).
}
}
\examples{
library(data.table)
new(Class = 'DDdt', 
    data.table(Variable = 'NOrden', 
               Sort = 'IDQual', 
               Class = 'character',
               Length = '11', 
               Qual1 = 'NOrden', 
               ValueRegExp = '[0-9]{9}SS'))
               
data(ExampleDDdt)
ExampleDDdt

}

