% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDdtToVNC.R
\name{DDdtToVNC}
\alias{DDdtToVNC}
<<<<<<< HEAD
\title{Returns the corresponding object of class \linkS4class{VNC} according to the input 
\linkS4class{data.table}}
=======
\title{Returns the corresponding object of class \linkS4class{VarNameCorresp} according to the
 DDdt input object}
>>>>>>> 5034523f22c62817420f2f5687369d62b4523cd8
\usage{
DDdtToVNC(DDdt, NameVNC, InFiles = rep("", dim(DDdt)[1]))
}
\arguments{
<<<<<<< HEAD
\item{DDdt}{Object of class \linkS4class{data.table}.}
=======
\item{DDdt}{Object of class \linkS4class{DDdt}.}
>>>>>>> 5034523f22c62817420f2f5687369d62b4523cd8

\item{NameVNC}{character vector with the name of the element in VNC which is being build.}

\item{InFiles}{character vector with as many components as the rows of \code{DDdt} specifying the 
files where to include each variable.}
}
\value{
Returns the corresponding object of class \linkS4class{VarNameCorresp}.
}
\description{
\code{DDdtToVNC} returns the corresponding object of class 
<<<<<<< HEAD
\linkS4class{VNC} according to the input \linkS4class{data.table}.
=======
\linkS4class{VarNameCorresp} according to the \linkS4class{DDdt} input object.
>>>>>>> 5034523f22c62817420f2f5687369d62b4523cd8
}
\examples{
library(data.table)
AggWeights <- data.table(Variable = c('ID', 'Pond1'), Sort = c('IDQual', 'IDDD'),
                         Class = c('character', 'character'),
                         Length = c('11', '7'),
                         Qual1 = c('', 'ID'), ValueRegExp = c('', ''))
<<<<<<< HEAD
=======
AggWeights <- new(Class = 'DDdt', AggWeights)
>>>>>>> 5034523f22c62817420f2f5687369d62b4523cd8
DDdtToVNC(AggWeights, 'AggWeights', rep('FA', 2))

}

