% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.StQ.R
\name{dim.StQ}
\alias{dim.StQ}
\title{Dimensions of an Object}
\usage{
\method{dim}{StQ}(x)
}
\arguments{
\item{x}{Object of class \linkS4class{rawStQ} or \link{StQ}.}
}
\value{
dim retrieves the dim attribute of the slot \code{Data} of the input object.
}
\description{
Retrieve or set the dimension of an object.

It is indeed the method \code{dim} for the classes \linkS4class{rawStQ} and \link{StQ}. 
This method returns the dimensions of the slot \code{Data} from the input object.
}
\examples{
data(ExampleStQ)
dim(ExampleStQ)

}
