% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DTToKey.R
\docType{methods}
\name{DTToKey}
\alias{DTToKey}
\alias{DTToKey,data.table-method}
\title{Return a character vector of keys out of a \linkS4class{data.table}.}
\usage{
DTToKey(DT)

\S4method{DTToKey}{data.table}(DT)
}
\arguments{
\item{DT}{\linkS4class{data.table} with the key components.}
}
\value{
Object of class \linkS4class{rawKey} with the keys of the input
}
\description{
\code{DTToKey} returns a character vector of keys from the input
\linkS4class{data.table} whose columns are the different qualifiers with as
many rows as components (length) of this input key.
}
\examples{
key <- new(Class = 'rawKey', c('IDDD:IASSEmpleo_Norden:391092SS_EsRemuner:1_TipoRemuner:1',
                               'IDDD:IASSCifraNeg_Norden:asd2SS',
                               'IDDD:IASSLPCifraNeg_Norden:1231_CCAA:01'))
keyDT <- DTToKey(KeyToDT(key))
}

