% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDatadt.R
\docType{data}
\name{ExampleDatadt}
\alias{ExampleDatadt}
\title{Object of class \linkS4class{data.table}}
\format{
Object of class \code{data.table}. Its slot \code{Data} has 354458 rows and 5 columns.
}
\usage{
data(ExampleDatadt)
}
\description{
\code{ExampleDatadt} is an object of class \linkS4class{data.table} containing severely 
perturbed anonymised data from a real official statistical operation.
}
