% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDD.R
\docType{data}
\name{ExampleDD}
\alias{ExampleDD}
\title{Object of class \linkS4class{DD}}
\format{Object of class \linkS4class{DD}. Its slot \code{Data} has 52 rows
and 6 columns.}
\usage{
data(ExampleDD)
}
\description{
\code{ExampleDD} is an object of class \linkS4class{DD}.

The columns of the slot \code{Data} are:
\itemize{
  \item Variable: statistical variable name.
  \item Sort: semantic sort of the variable (IDQual, NonIDQual, IDDD).
  \item Class: class of the variable (\code{numeric}, \code{character},...).
  \item Qual1: first variable qualifier.
  \item Qual2: second variable qualifier.
  \item Qual3: third variable qualifier.
}
}

