% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDDdt.R
\docType{data}
\name{ExampleDDdt}
\alias{ExampleDDdt}
\title{Object of class \linkS4class{DDdt}}
\format{Object of class \linkS4class{DDdt} with 14 rows and 8 columns.}
\usage{
data(ExampleDDdt)
}
\description{
\code{ExampleDDdt} is an object of class \linkS4class{DDdt}.

This class is a subclass of \linkS4class{data.table} with columns \code{Variable}, \code{Sort},
\code{Class}, \code{Length},\code{Qual}\emph{1}-\emph{j}, and \code{ValueRegExp}.

The meaning of these columns are:
\itemize{
  \item \code{Variable}: statistical variable name.
  \item \code{Sort}: semantic sort of the variable (IDQual, NonIDQual, IDDD).
  \item \code{Class}: class of the variable (\code{numeric}, \code{character},...).
  \item \code{Length}: highest length for each variable.
  \item \code{Qual1} to \code{Qual3}: 1st to 3rd variable qualifier.
  \item \code{ValueRegExp}: regexp for the variable values.
}
}

