<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> e9efa7c0eccfd9de1f1e25ef8428b085f47e8788
% Please edit documentation in R/ExampleDM.R
\docType{data}
\name{ExampleDM}
\alias{ExampleDM}
<<<<<<< HEAD
\title{\linkS4class{data.table} with a rectangular data set}
\format{\code{data.table} with 29499 rows and 59 columns.}
=======
\title{Conjunto de datos con estructura matricial.}
\format{\code{data.table} con 29499 filas y 59 columnas.}
>>>>>>> e9efa7c0eccfd9de1f1e25ef8428b085f47e8788
\usage{
data(ExampleDM)
}
\description{
<<<<<<< HEAD
\code{ExampleDM} is a \linkS4class{data.table} in data matrix
structure with anonymised data from an undisclosed period of the Sector
Service Indicators Survey.
=======
\code{ExampleDM} es una \link{data.table} con estructura de datos matricial,
con los datos correspondientes a un periodo no especificado de la encuesta de
 IASS.
}
\details{
Los nombres de las columnas son aquellos de las variables en el repositorio,
anexando los calificadores con un guión bajo cuando sea necesario (ver
\code{DD} para más información acerca de sus significados).
>>>>>>> e9efa7c0eccfd9de1f1e25ef8428b085f47e8788
}

