% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDTkey.R
\docType{data}
\name{ExampleDTkey}
\alias{ExampleDTkey}
\title{\linkS4class{data.table} with the key component of a \linkS4class{StQ} object}
\format{Object of class \linkS4class{data.table} for 354458 values.}
\usage{
data(ExampleDTkey)
}
\description{
\code{ExampleDTkey} is an object of class \linkS4class{data.table} with the compound
key for each value of data set.
}

