% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleStQ.R, R/ExampleVNC.R
\docType{data}
\name{ExampleQ}
\alias{ExampleQ}
\title{Object of class \linkS4class{StQ}}
\format{Object of class \code{StQ}. Its slot \code{Data} has 1710942 rows and
 6 columns.}
\usage{
data(ExampleQ)

data(ExampleQ)
}
\description{
\code{ExampleQ} is an object of class \linkS4class{StQ} whose
slot \code{Data} is a \link{data.table} with key-value pair structure with
anonymised data of an undisclosed time period.

The columns of slot \code{Data} of this object is:
\itemize{
  \item NOrden: First qualifier (statistical unit qualifier).
  \item CCAA: Second qualifier (variable name qualifier).
  \item EsRemuner: Third qualifier (variable name qualifier).
  \item TipoRem: Fourth qualifier (variable name qualifier).
  \item IDDD: Variable names.
  \item Value: Variable values (default class: \code{character}).
}

\code{ExampleQ} is an object of class \linkS4class{StQ} whose
slot \code{Data} is a \link{data.table} with key-value pair structure with
anonymised data of an undisclosed time period.

The columns of slot \code{Data} of this object is:
\itemize{
  \item NOrden: First qualifier (statistical unit qualifier).
  \item CCAA: Second qualifier (variable name qualifier).
  \item EsRemuner: Third qualifier (variable name qualifier).
  \item TipoRem: Fourth qualifier (variable name qualifier).
  \item IDDD: Variable names.
  \item Value: Variable values (default class: \code{character}).
}
}

