% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExamplerawKey.R
\docType{data}
\name{ExamplerawKey}
\alias{ExamplerawKey}
\title{Object of class \linkS4class{rawKey}}
\format{Object of class \linkS4class{rawKey} for 354456 values.}
\usage{
data(ExamplerawKey)
}
\description{
\code{ExamplerawKey} is an object of class \linkS4class{rawKey} with the compound
key for each value of data set.
}

