% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExamplerawStQ.R
\docType{data}
\name{ExamplerawStQ}
\alias{ExamplerawStQ}
\title{Object of class \linkS4class{rawStQ}}
\format{Object of class \code{rawStQ}. Its slot \code{Data} has 354458 rows and
 2 columns.}
\usage{
data(ExamplerawStQ)
}
\description{
\code{ExamplerawStQ} is an object of class \linkS4class{rawStQ}
containing severely perturbed anonymised data of real data set.
}

