% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleStQ.R
\docType{data}
\name{ExampleStQ}
\alias{ExampleStQ}
\title{Object of class \linkS4class{StQ}}
\format{Object of class \code{StQ}. Its slot \code{Data} has 354458 rows and 4 columns.}
\usage{
data(ExampleStQ)
}
\description{
\code{ExampleStQ} is an object of class \linkS4class{StQ} containing severely
perturbed anonymised data from a real official statistical operation.
}

