% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractNames.R
\name{ExtractNames}
\alias{ExtractNames}
\title{Extract the root name of compound variable names}
\usage{
ExtractNames(NamesVector)
}
\arguments{
\item{NamesVector}{\code{Character} vector with the original compound variable names.}
}
\value{
\code{Character} vector of the same length as the input parameter \code{NamesVector} with
 the corresponding root names of each variable name.
}
\description{
\code{ExtractNames} returns a \code{character} vector with the root names of the 
input parameter.

This function is thought to be applied on those variable names with suffixes containing 
underscores (_) and qualifier values. It takes as input parameter \code{NamesVector} a character 
vector with the names of possibly several variables and extracts their root names dropping out 
every suffix. The returned variable root names must have an exact correspondence with the 
variable names (column "Variable") of class \link{DD}.
}
\examples{
ExtractNames(c('ICMCifraNeg', 'ICMEmpleo_1.', 'ICMEmpleo_2.1', 'ICMEmpleo_2.2'))
ExtractNames(c('IASSCifraNeg.x', 'IASSCifraNeg.y'))
ExtractNames(c('ICMCifraNeg.x', 'ICMEmpleo_1..x', 'ICMEmpleo_2.1.x', 'ICMEmpleo_2.2.x',
               'ICMCifraNeg.y', 'ICMEmpleo_1..y', 'ICMEmpleo_2.1.y', 'ICMEmpleo_2.2.y'))
               
}
