% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelativeN-class.R
\docType{class}
\name{FileName-class}
\alias{FileName-class}
\title{S4 class \code{CorrelativeN} for a sequence of integer numbers}
\description{
An S4 class called \code{CorrelativeN} for a sequence of integer numbers, mostly
correlative to be used as suffixes in the number of variables.
}
\examples{
# Empty object
x <- new(Class = 'FileName')
x
str(x)

# Example with one time interval
FF.MM032014 <- new(Class = 'FileName', c('E30183.FF_V1.MM032014.D_1'))
FF.MM032014
str(FF.MM032014)

}

