% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FirstLine.R
\docType{methods}
\name{FirstLine}
\alias{FirstLine}
\alias{FirstLine,StQ-method}
\alias{FirstLine,data.table-method}
\title{Generación de la primera línea de un fichero del repositorio}
\usage{
FirstLine(object)

\S4method{FirstLine}{data.table}(object)

\S4method{FirstLine}{StQ}(object)
}
\arguments{
\item{object}{Objeto de clase \code{\linkS4class{StQ}}.}

\item{NIV}{\code{\link{vector}} de longitud 1 con el número de niveles en el fichero.
Corresponde al número total de calificadores que aparecen en el fichero.Por
defecto toma el valor \code{NIV} = \code{1}.}
}
\value{
\code{Vector} de tipo \code{character} de longitud 1 con las variables
que aparecen en el objeto \code{StQ} especificado y sus formatos correspondientes
es decir, con el diseño de registro del fichero. Incluye además como primera
variable NIV, con el valor especificado en el input, y una última, \code{M},
que corresponde a la máxima longitud de registro en el fichero.
}
\description{
\code{FirstLine} devuelve un \code{vector} alfanumérico de longitud 1
con la información que aparece en la primera línea de los ficheros ubicados
en el repositorio.
}
\details{
Esta función toma un objeto de clase \code{\linkS4class{StQ}} como parámetro
de entrada y devuelve la primera línea que el objeto debe contener cuando es
incluido en el repositorio de microdatos de la encuesta. La información que
contiene hace referencia al diseño de registro del fichero.

Además del objeto cuyos datos vamos a incluir en el fichero, es necesario
especificar como argumento de entrada el número de calificadores
utilizados en el fichero (\code{NIV}).
}
\examples{
data(ExampleQ)
FirstLine(ExampleQ)
}
\seealso{
\link{ReadRepoFile}, \link{WriteRepoFile}
}

