% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAggr.R
\docType{methods}
\name{getAggr}
\alias{getAggr}
\alias{getAggr,DD-method}
\title{Return slot Aggregates from a DD object}
\usage{
getAggr(object)

\S4method{getAggr}{DD}(object)
}
\arguments{
\item{object}{a DD Object whose slot \code{Aggregates} is queried.}
}
\value{
\linkS4class{data.table} with data from slot \code{Aggregates} of the
input \linkS4class{DD} object.
}
\description{
\code{getAggr} returns slot \code{Aggregates} of the input
\linkS4class{DD} object.
}
\examples{
# An example with data created previosly:
data(ExampleDD)
getAggr(ExampleDD)

}

