% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getAggWeights.R, R/getOtherDD.R
\docType{methods}
\name{getAggWeights}
\alias{getAggWeights}
\alias{getAggWeights,DD-method}
\alias{getOtherDD,DD-method}
\title{Return slot AggWeights from a DD object}
\usage{
getAggWeights(object)

\S4method{getAggWeights}{DD}(object)

\S4method{getOtherDD}{DD}(object)
}
\arguments{
\item{object}{a DD Object whose slot \code{AggWeights} is queried.}
}
\value{
\linkS4class{data.table} with data from slot \code{AggWeights} of the
input \linkS4class{DD} object.
}
\description{
\code{getAggWeights} returns slot \code{AggWeights} of the input
\linkS4class{DD} object.
}
\examples{
# An example with data created previosly:
data(ExampleDD)
getAggWeights(ExampleDD)
}

