% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClass.R
\docType{methods}
\name{getClass}
\alias{getClass}
\alias{getClass,DD-method}
\alias{getClass,StQ-method}
\title{Return class for each input variable from the dictionary of the input object}
\usage{
getClass(object, variables = NULL)

\S4method{getClass}{DD}(object, variables = NULL)

\S4method{getClass}{StQ}(object, variables = NULL)
}
\arguments{
\item{object}{Object whose class for each input variable in the slot \linkS4class{DD} are 
required.}

\item{variables}{\code{Character} vector with the names of the variables whose class is required.}
}
\value{
Returns a character vector with the class of each input variables. If \code{variables} is 
\code{NULL} all variables in the dictionary are considered.
}
\description{
\code{getClass} returns the class for each input variable from the dictionary of the 
input object.
}
\examples{
getClass(ExampleStQ)


}
