% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDD.R
\docType{methods}
\name{getDD}
\alias{getDD}
\alias{getDD,StQ-method}
\alias{getDD,StQList-method}
\alias{getDD,rawStQ-method}
\title{Return slot \code{DD} from an object of class \linkS4class{StQ}}
\usage{
getDD(object)

\S4method{getDD}{StQ}(object)

\S4method{getDD}{StQList}(object)

\S4method{getDD}{rawStQ}(object)
}
\arguments{
\item{Object}{of class \linkS4class{StQ}.}
}
\value{
Object of class \linkS4class{DD} corresponding to the slot \code{DD}
of the input parameter.
}
\description{
\code{getDD} extracts slot \code{DD} from an object of class
 \linkS4class{StQ}.

This method returns the \linkS4class{data.table} corresponding to slot
\code{DD} from an object of class \linkS4class{StQ} specified as input
argument.
}
\examples{
data(Q)
DD <- getDD(Q)
DD
str(DD)

}

