% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDotQual.R
\docType{methods}
\name{getDotQual}
\alias{getDotQual}
\alias{getDotQual,DD-method}
\alias{getDotQual,StQ-method}
\alias{getDotQual,VNC-method}
\title{Return non-unit qualifiers not used in compound variable names}
\usage{
getDotQual(object, Component = names(object))

\S4method{getDotQual}{VNC}(object, Component = names(object))

\S4method{getDotQual}{DD}(object, Component = names(getVNC(object)))

\S4method{getDotQual}{StQ}(object, Component = names(getVNC(object)))
}
\arguments{
\item{object}{Object containing the dot qualifiers.}

\item{Component}{\code{Character} vector with the components or slots from which dot qualifiers 
are required.}
}
\value{
\code{Character} vector with all the dot qualifier names.
}
\description{
\code{getDotQual} returns a character vector with all qualifier names specified with
a wildcard dot in the VNC component of the input object.
}
\examples{
library(data.table)
VarList <- list(ID = data.table(IDQual = c('NumIdEst', rep('', 4)),
                                NonIDQual = c('','','','',''),
                                IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                                NumIdEst = c('', rep('.', 4)),
                                UnitName = c('numidest', 'nombre', 'apellidos', 'direccion', 
                                             'telefono'),
                                InFiles = rep('FI', 5)),
MicroData = data.table(IDQual = c('NumIdEst', rep('', 2)),
                       NonIDQual = c('', 'Market', ''),
                       IDDD = c(rep('', 2), 'NewOrders'),
                       NumIdEst = c(rep('', 2), '.'),
                       Market = c(rep('', 2), '1.'),
                       UnitName = c('numidest', '', 'cp09'),
                       InFiles = rep('FF, FD, FG', 3)),
ParaData = data.table(IDQual = c('NumIdEst', rep('', 2)),
                      NonIDQual = c('', 'Action', ''),
                      IDDD = c(rep('', 2), 'Date'),
                      NumIdEst = c(rep('', 2), '.'),
                      Action = c(rep('', 2), 'Imputation'),
                      UnitName = c('numidest', '', 'FechaImput'),
                      InFiles = rep('FP', 3)),
AggWeights = data.table(IDQual = c('CCAA', 'NACE09', ''),
                        NonIDQual = rep('', 3),
                        IDDD = c('', '', 'Ponderacion'),
                        CCAA = c('', '', '.'),
                        NACE09 = c('', '', '.'),
                        UnitName = c('Provincia', '', ''),
                        InFiles = rep('FA', 3)))
VNC <- BuildVNC(VarList)
getDotQual(VNC)

}

