% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEditNames.R
\docType{methods}
\name{getEditNames}
\alias{getEditNames}
\alias{getEditNames,data.table-method}
\title{Extract edit names from an object (THIS FUNCTION MUST BE REFACTORED)}
\usage{
getEditNames(object)

\S4method{getEditNames}{data.table}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{StQ}.}
}
\value{
It returns a \code{data.table} with the statistical units in the input object with their
corresponding edit names.
}
\description{
This method identifies the IDQual qualifiers in the input object and returns a
\linkS4class{data.table} with the values of these qualifiers for each statistical unit and their
corresponding edit names.
}
\examples{
\dontrun{
data(ExampleStQ)
getEditNames(ExampleStQ)
}

}

