% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getID.R
\docType{methods}
\name{getID}
\alias{getID}
\alias{getID,DD-method}
\title{Return slot ID from a DD object}
\usage{
getID(object)

\S4method{getID}{DD}(object)
}
\arguments{
\item{object}{a DD Object whose slot \code{ID} is queried.}
}
\value{
\linkS4class{data.table} with data from slot \code{ID} of the
input \linkS4class{DD} object.
}
\description{
\code{getID} returns slot \code{ID} of the input
\linkS4class{DD} object.
}
\examples{
# An example with data created previosly:
data(ExampleDD)
getID(ExampleDD)

}

