% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDDD.R
\docType{methods}
\name{getIDDD}
\alias{getIDDD}
\alias{getIDDD,DD-method}
\alias{getIDDD,DDdt-method}
\alias{getIDDD,Datadt-method}
\alias{getIDDD,StQ-method}
\alias{getIDDD,VNCdt-method}
\alias{getIDDD,VarNameCorresp-method}
\title{Return IDDD identifiers from an object}
\usage{
getIDDD(object, CompNames)

\S4method{getIDDD}{VNCdt}(object)

\S4method{getIDDD}{VarNameCorresp}(object, CompNames)

\S4method{getIDDD}{DDdt}(object)

\S4method{getIDDD}{DD}(object, CompNames)

\S4method{getIDDD}{Datadt}(object)

\S4method{getIDDD}{StQ}(object)
}
\arguments{
\item{object}{Object whose IDDD identifiers are required.}

\item{Namesdt}{Character vector with the names of components or slots of the 
object from which the IDDD identifiers are requested.}
}
\value{
In the case of \linkS4class{VarNameCorresp} it returns a character
vector with the IDDD identifiers from the components specified in Namesdt. 
If no Namesdt is specified, it returns the IDDD identifier from all the 
components of the object.
}
\description{
\code{getIDDD} returns the IDDD identifiers from the input 
object.
}
\examples{
library(data.table)
VarList <- list(
  ID = new(Class = 'VNCdt', 
           .Data = data.table(
                 IDQual = c('NumIdEst', rep('', 4)),
                 NonIDQual = rep('', 5),
                 IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                 NumIdEst = c('', rep('.', 4)),
                 Unit1 = c('numidest', 'nombre', 'apellidos', 'direccion', 'telefono'))),
  MicroData = new(Class = 'VNCdt', 
                  .Data = data.table(
                      IDQual = c('NumIdEst', rep('', 4)),
                      NonIDQual = c('', 'IsNatMarket', 'IsEuroMarket', 'IsRWMarket', ''),
                      IDDD = c(rep('', 4), 'NewOrders'),
                      NumIdEst = c(rep('', 4), '.'),
                      IsNatMarket = c(rep('', 4), '0'),
                      IsEuroMarket = c(rep('', 4), '0'),
                      IsRWMarket = c(rep('', 4), '1'),
                      Unit1 = c('numidest', rep('', 3), 'cp09'))),
  ParaData = new(Class = 'VNCdt'),
  Aggregates = new(Class = 'VNCdt', 
                   .Data = data.table(
                      IDQual = c('Province', 'NACE', 'IsNatMarket', ''),
                      NonIDQual = rep('', 4),
                      IDDD = c('', '', '', 'TotalTurnover'),
                      Province = c('', '', '', '.'),
                      NACE = c('', '', '', '.'),
                      IsNatMarket = c('', '', '', '1'),
                      Unit1 = c('provincia', 'actividad', '', 'cn01'))))
VNC <- new(Class = 'VarNameCorresp', .Data = VarList)
getIDDD(VNC)


}

