% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNonIDQual.R
\docType{methods}
\name{getNonIDQual}
\alias{getNonIDQual}
\alias{getNonIDQual,DD-method}
\alias{getNonIDQual,DDdt-method}
\alias{getNonIDQual,StQ-method}
\alias{getNonIDQual,VNCdt-method}
\alias{getNonIDQual,VarNameCorresp-method}
\title{Return the variable names included in columns 'NonIDQual" of the input object}
\usage{
getNonIDQual(object, Namesdt)

\S4method{getNonIDQual}{VNCdt}(object)

\S4method{getNonIDQual}{VarNameCorresp}(object, Namesdt)

\S4method{getNonIDQual}{DDdt}(object)

\S4method{getNonIDQual}{DD}(object, Namesdt)

\S4method{getNonIDQual}{StQ}(object)
}
\arguments{
\item{object}{Object with the NonIDQual variable qualifiers.}

\item{Namesdt}{Character vector with the components or slots of the object
from which NonIDQuals are required.}
}
\value{
Character vector with the NonIDQual variable qualifiers names.
}
\description{
\code{getNonIDQual} returns a character vector with all variable 
qualifiers names (NonIDQual) included in the input object.
}
\examples{
library(data.table)
VarList <- list(
  ID = new(Class = 'VNCdt', 
           .Data = data.table(
                 IDQual = c('NumIdEst', rep('', 4)),
                 NonIDQual = rep('', 5),
                 IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                 NumIdEst = c('', rep('.', 4)),
                 Unit1 = c('numidest', 'nombre', 'apellidos', 'direccion', 'telefono'))),
  MicroData = new(Class = 'VNCdt', 
                  .Data = data.table(
                      IDQual = c('NumIdEst', rep('', 4)),
                      NonIDQual = c('', 'IsNatMarket', 'IsEuroMarket', 'IsRWMarket', ''),
                      IDDD = c(rep('', 4), 'NewOrders'),
                      NumIdEst = c(rep('', 4), '.'),
                      IsNatMarket = c(rep('', 4), '0'),
                      IsEuroMarket = c(rep('', 4), '0'),
                      IsRWMarket = c(rep('', 4), '1'),
                      Unit1 = c('numidest', rep('', 3), 'cp09'))),
 ParaData = new(Class = 'VNCdt'),
 Aggregates = new(Class = 'VNCdt', 
                  .Data = data.table(
                     IDQual = c('Province', 'NACE', 'IsNatMarket', ''),
                     NonIDQual = rep('', 4),
                     IDDD = c('', '', '', 'TotalTurnover'),
                     Province = c('', '', '', '.'),
                     NACE = c('', '', '', '.'),
                     IsNatMarket = c('', '', '', '1'),
                     Unit1 = c('provincia', 'actividad', '', 'cn01'))))
VNC <- new(Class = 'VarNameCorresp', .Data = VarList)
getNonIDQual(VNC)


}

