% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMicroData.R, R/getParaData.R
\docType{methods}
\name{getMicroData,DD-method}
\alias{getMicroData,DD-method}
\alias{getParaData}
\alias{getParaData,VNC-method}
\alias{getParaData,DD-method}
\title{Return slot ParaData from an object}
\usage{
\S4method{getMicroData}{DD}(object)

getParaData(object)

\S4method{getParaData}{VNC}(object)

\S4method{getParaData}{DD}(object)
}
\arguments{
\item{object}{a DD Object whose slot \code{ParaData} is queried.}
}
\value{
Returns a \linkS4class{DDdt} object containing the slot \code{ParaData} of the input object.
}
\description{
\code{getParaData} returns slot \code{ParaData} of the input object.
}
\examples{
# An example:
library(data.table)
### We build the VNC object
VarList <- list(ID = data.table(IDQual = c('NumIdEst', rep('', 4)),
                                NonIDQual = c('','','','',''),
                                IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                                NumIdEst = c('', rep('.', 4)),
                                UnitName = c('numidest', 'nombre', 'apellidos', 'direccion', 'telefono'),
                                InFiles = rep('FI', 5)),
MicroData = data.table(IDQual = c('NumIdEst', rep('', 2)),
                       NonIDQual = c('', 'Market', ''),
                       IDDD = c(rep('', 2), 'NewOrders'),
                       NumIdEst = c(rep('', 2), '.'),
                       Market = c(rep('', 2), '1.'),
                       UnitName = c('numidest', '', 'cp09'),
                       InFiles = rep('FF', 3)),
ParaData = data.table(IDQual = c('NumIdEst', rep('', 2)),
                      NonIDQual = c('', 'Action', ''),
                      IDDD = c(rep('', 2), 'Date'),
                      NumIdEst = c(rep('', 2), '.'),
                      Action = c(rep('', 2), 'Imputation'),
                      UnitName = c('numidest', '', 'FechaImput'),
                      InFiles = rep('FP', 3)),
AggWeights = data.table(IDQual = c('CCAA', 'NACE09', ''),
                        NonIDQual = rep('', 3),
                        IDDD = c('', '', 'Ponderacion'),
                        CCAA = c('', '', '.'),
                        NACE09 = c('', '', '.'),
                        UnitName = c('Provincia', '', ''),
                        InFiles = rep('FF', 3)))

VNC <- BuildVNC(VarList)
getParaData(VNC)

}
