% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPeriods.rawStQList.R
\docType{methods}
\name{getPeriods.rawStQList}
\alias{getPeriods.rawStQList}
\alias{getPeriods.rawStQList,rawStQList-method}
\title{Return a vector with the names of the time intervals.}
\usage{
getPeriods.rawStQList(object)

\S4method{getPeriods.rawStQList}{rawStQList}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{rawStQList}.}
}
\value{
Character vector with the names of the time periods according to the notation of package
\linkS4class{RepoTimeInt}.
}
\description{
\code{getPeriods} returns a vector with the names of the time periods from the slot
\code{Periods} of the input \linkS4class{rawStQList} object.
}
\examples{
library(RepoTime)
mm <- c(paste0('0', 1:9), 10:12)
TimePer <- paste0('MM', mm, '2015')
PeriodList <- newRepoTime(TimePer)
EmptyrawQ <- new(Class = 'rawStQ')
EmptyrawQList <- vector('list', 12)
EmptyrawQList <- lapply(EmptyrawQList, function(x) EmptyrawQ)
rawQList <- new(Class = 'rawStQList', Data = EmptyrawQList, Periods = PeriodList)
getPeriods.rawStQList(rawQList)

}

