% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPeriods.R
\docType{methods}
\name{getPeriods}
\alias{getPeriods}
\alias{getPeriods,StQList-method}
\alias{getPeriods,rawStQList-method}
\title{Return a vector with the names of the time period intervals}
\usage{
getPeriods(object)

\S4method{getPeriods}{rawStQList}(object)

\S4method{getPeriods}{StQList}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{rawStQList} or \linkS4class{StQList}.}
}
\value{
Character vector with the names of the time periods according to the notation of package
\linkS4class{RepoTimeInt}.
}
\description{
\code{getPeriods} returns a vector with the names of the time periods from the input
 object.
}
\examples{
library(RepoTime)
mm <- c(paste0('0', 1:9), 10:12)
TimePer <- paste0('MM', mm, '2015')
PeriodList <- newRepoTime(TimePer)

<<<<<<< HEAD
EmptyrawQ <- rawStQ())
EmptyrawQList <- vector('list', 12)
EmptyrawQList <- lapply(EmptyrawQList, function(x) EmptyrawQ)
rawQList <- rawStQList(Data = EmptyrawQList, Periods = PeriodList)
=======
EmptyrawQ <- new(Class = 'rawStQ')
EmptyrawQList <- vector('list', 12)
EmptyrawQList <- lapply(EmptyrawQList, function(x) EmptyrawQ)
rawQList <- new(Class = 'rawStQList', Data = EmptyrawQList, Periods = PeriodList)
>>>>>>> 5034523f22c62817420f2f5687369d62b4523cd8
getPeriods(rawQList)

EmptyQ <- new(Class = 'StQ')
EmptyQList <- vector('list', 12)
EmptyQList <- lapply(EmptyQList, function(x) EmptyQ)
<<<<<<< HEAD
QList <- StQList(Data = EmptyQList, Periods = PeriodList)
=======
QList <- new(Class = 'StQList', Data = EmptyQList, Periods = PeriodList)
>>>>>>> 5034523f22c62817420f2f5687369d62b4523cd8
getPeriods(QList)

}

