% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPeriods.StQList.R
\docType{methods}
\name{getPeriods.StQList}
\alias{getPeriods.StQList}
\alias{getPeriods.StQList,StQList-method}
\title{Return a vector with the names of the time intervals}
\usage{
getPeriods.StQList(object)

\S4method{getPeriods.StQList}{StQList}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{StQList}.}
}
\value{
Character vector with the names of the time periods according to the
notation of package \linkS4class{RepoTimeInt}.
}
\description{
\code{getPeriods} returns a vector with the names of the time
periods from the slot \code{Periods} of the input \linkS4class{StQList}
object.
}
\examples{
library(RepoTime)
mm <- c(paste0('0', 1:9), 10:12)
TimePer <- paste0('MM', mm, '2015')
PeriodList <- newRepoTime(TimePer)
EmptyQ <- new(Class = 'StQ')
EmptyQList <- vector('list', 12)
EmptyQList <- lapply(EmptyQList, function(x) EmptyQ)
QList <- new(Class = 'StQList', Data = EmptyQList, Periods = PeriodList)
getPeriods.StQList(QList)

}

