% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUnitName.R
\name{getUnitName}
\alias{getUnitName}
\alias{getUnitName,VNC-method}
\alias{getUnitName,DD-method}
\alias{getUnitName,StQ-method}
\title{Return UnitNames from the input object}
\usage{
getUnitName(object, CompNames = names(object))

\S4method{getUnitName}{VNC}(object, CompNames = names(object))

\S4method{getUnitName}{DD}(object, CompNames = names(object))

\S4method{getUnitName}{StQ}(object, CompNames = names(object))
}
\arguments{
\item{object}{Object whose UnitNames in the slot \link{DD} are required.}

\item{CompNames}{\code{Character} vector with the names of components or slots of the object from
which the UnitNames are requested.}
}
\value{
Returns a character vector with the UnitNames from the components specified in
CompNames. If no CompNames parameter is specified, it returns the UnitNames from all the
components of the input object.
}
\description{
\code{getUnitName} returns the UnitNames from input object.
}
\examples{
# An example:
library(data.table)
### We build the VNC object
VarList <- list(ID = data.table(IDQual = c('NumIdEst', rep('', 4)),
                                NonIDQual = c('','','','',''),
                                IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                                NumIdEst = c('', rep('.', 4)),
                                UnitName = c('numidest', 'nombre', 'apellidos', 'direccion', 'telefono'),
                                InFiles = rep('FI', 5)),
MicroData = data.table(IDQual = c('NumIdEst', rep('', 2)),
                       NonIDQual = c('', 'Market', ''),
                       IDDD = c(rep('', 2), 'NewOrders'),
                       NumIdEst = c(rep('', 2), '.'),
                       Market = c(rep('', 2), '1.'),
                       UnitName = c('numidest', '', 'cp09'),
                       InFiles = rep('FF', 3)),
ParaData = data.table(IDQual = c('NumIdEst', rep('', 2)),
                      NonIDQual = c('', 'Action', ''),
                      IDDD = c(rep('', 2), 'Date'),
                      NumIdEst = c(rep('', 2), '.'),
                      Action = c(rep('', 2), 'Imputation'),
                      UnitName = c('numidest', '', 'FechaImput'),
                      InFiles = rep('FP', 3)),
AggWeights = data.table(IDQual = c('CCAA', 'NACE09', ''),
                        NonIDQual = rep('', 3),
                        IDDD = c('', '', 'Ponderacion'),
                        CCAA = c('', '', '.'),
                        NACE09 = c('', '', '.'),
                        UnitName = c('Provincia', '', ''),
                        InFiles = rep('FF', 3)))

VNC <- BuildVNC(VarList)

getUnitName(VNC)


}
