% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUnits.R
\docType{methods}
\name{getUnits}
\alias{getUnits}
\alias{getUnits,StQ-method}
\alias{getUnits,data.table-method}
\title{Extract statistical units from an object}
\usage{
getUnits(object)

\S4method{getUnits}{StQ}(object)

\S4method{getUnits}{data.table}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{StQ} or class 
\linkS4class{data.table}.}
}
\value{
It returns a \code{data.table} with the statistical units in the
input object.
}
\description{
This method identifies the IDQual qualifiers in the input object
and returns a \linkS4class{data.table} with the values of these qualifiers
for each statistical unit.
}
\examples{
<<<<<<< HEAD
data(Q)
getUnits(Q)

||||||| merged common ancestors
data(ExampleQ)
getUnits(ExampleQ)
=======
data(ExampleQ)
getUnits(ExampleQ)

>>>>>>> e3ac9308c1978eb314de9c1fda9971de40a9bce7
}

