% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValues.R
\docType{methods}
\name{getValues}
\alias{getValues}
\alias{getValues,StQ,character-method}
\alias{getValues,StQList,character-method}
\title{Return the values of a variable of specified statistical units}
\usage{
getValues(object, VarName, Units = getUnits(object, "MicroData"))

\S4method{getValues}{StQ,character}(object, VarName, Units = getUnits(object,
  "MicroData"))

\S4method{getValues}{StQList,character}(object, VarName,
  Units = getUnits(object, "MicroData"))
}
\arguments{
\item{object}{Object of class \linkS4class{StQ} or \linkS4class{StQList}.}

\item{VarName}{Character vector with the name of the variable.}

\item{Units}{\linkS4class{data.table} with the qualifier values identifying each statistical unit
in the input object.}
}
\value{
Returns a \linkS4class{data.table} with the unit qualifier and the corresponding value 
for each statistical unit.
}
\description{
\code{getValues} returns the values of the variable specified as input parameter of
those statistical units also specified as an input parameter of the input object.
}
\examples{
library(data.table)
getValues(ExampleStQ, 'Employees_1.')
getValues(ExampleStQ, 'Turnover')

}

