% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVNC.R
\docType{methods}
\name{getVNC}
\alias{getVNC}
\alias{getVNC,DD-method}
\alias{getVNC,StQ-method}
\title{Return slot \code{VarNameCorresp} from an object}
\usage{
getVNC(object)

\S4method{getVNC}{DD}(object)

\S4method{getVNC}{StQ}(object)
}
\arguments{
\item{Object}{of class \linkS4class{DD} or \linkS4class{StQ}.}
}
\value{
Object of class \linkS4class{VarNameCorresp} of the input parameter.
}
\description{
\code{getVNC} extracts slot \code{VarNameCorresp} from an object
 of class \linkS4class{DD} or class \linkS4class{StQ}.

This method returns an object of class \linkS4class{VarNameCorresp} from an 
object of class \linkS4class{DD} or \linkS4class{StQ} specified as input 
argument.
}
\examples{
# From an object of class DD
data(ExampleDD)
VNC <- getVNC(ExampleDD)
VNC
str(VNC)

# From an object of class StQ
data(ExampleQ)
VNC <- getVNC(ExampleQ)
VNC
str(VNC)

}

