% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDDDToUnitNames.R
\docType{methods}
\name{IDDDToUnitNames}
\alias{IDDDToUnitNames}
\alias{IDDDToUnitNames,character,VarNameCorresp-method}
\title{Convert IDDD variable names into their corresponding unit\emph{j} names}
\usage{
IDDDToUnitNames(IDDDNames, VNC, Unit = "Unit1")

\S4method{IDDDToUnitNames}{character,VarNameCorresp}(IDDDNames, VNC,
  Unit = "Unit1")
}
\arguments{
\item{object}{Object with the IDDD variable identifiers.}
}
\value{
Character vector with all the variable names.
}
\description{
\code{IDDDToUnitNames} returns a character vector with the unit
variable name for each IDDD variable name.
}
\examples{
IDDDToUnitNames(ExampleQ)

}

