% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/melt_StQ.R
\name{melt_StQ}
\alias{melt_StQ}
\title{Convert a dcasted \linkS4class{data.table} into an object of class
\linkS4class{StQ}}
\usage{
melt_StQ(DataMatrix, DD, DDslot = "MicroData")
}
\arguments{
\item{DataMatrix}{of class \linkS4class{data.table} with dcasted form
(statistical units by rows and variables by columns).}

\item{DD}{Object of class \linkS4class{DD}.}

\item{DDslot}{Character vector of length 1 with the name of DD slot used
to make the transformation of the input object. Its default value is
\code{MicroData}.}
}
\value{
Object of class \linkS4class{StQ}.
}
\description{
\code{melt_StQ} returns an object of class \linkS4class{StQ}
from a \linkS4class{data.table} in dcasted form.

This method builds an object of class \linkS4class{StQ} with the slot
\code{Data} constructed from the input \linkS4class{data.table} and the slot
\code{DD} given as an input parameter.

This function can be considered as a constructor for the class \code{StQ}. It
is indeed a wrapper for the function \code{\link[data.table]{melt.data.table}}
from the package \linkS4class{data.table} adapted to the structure of the slot
\code{Data} of object \linkS4class{StQ}.

The input parameter \code{DD} must be of class \linkS4class{DD}.
}
\examples{
data(ExampleDM)
data(ExampleDD)
Q <- melt_StQ(ExampleDM, ExampleDD)
str(Q)
}
\seealso{
\code{\link{dcast_StQ}}, \code{\link[data.table]{dcast.data.table}},
\code{\link[data.table]{melt.data.table}}, \code{\link[reshape2]{melt}},
\code{\link[reshape2]{dcast}}
}

