% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.VNC.R
\name{+.VNC}
\alias{+.VNC}
\title{Method \code{+} for the class \linkS4class{VarNameCorresp}}
\usage{
\method{+}{VNC}(e1, e2)
}
\arguments{
\item{e1}{Object of class \linkS4class{VarNameCorresp}.}

\item{e2}{Object of class \linkS4class{VarNameCorresp}.}
}
\value{
Object of class \linkS4class{VarNameCorresp} resulting from integrating both 
\linkS4class{VarNameCorresp} objects in a single \linkS4class{VarNameCorresp} object.
}
\description{
\code{+} sums two objects of class \linkS4class{VarNameCorresp}. This method 
overloads the operator \link{+} and returns a new object of class \linkS4class{VarNameCorresp}.

The integration is carried out according to the names of the components.
}
\examples{
library(data.table)
VarList1 <- list(
    ID = data.table(IDQual = c('NumIdEst', rep('', 4)),
                    NonIDQual = rep('', 5),
                    IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                    NumIdEst = c('', rep('.', 4)),
                    UnitName = c('numidest', 'nombre', 'apellidos', 'direccion', 'telefono'),
                    InFiles = rep('FI', 5)),
    MicroData = data.table(IDQual = c('NumIdEst', '', '', '', '', ''),
                           NonIDQual = c('', 'Market', 'Cod','Var1', 'Var2', ''),
                           IDDD = c('', '', '', '', '','IEPEntradaPed'),
                           NumIdEst = c('', '', '', '', '', '.'),
                           Market = c('', '', '', '', '','1.'),
                           Cod = c('', '', '', '', '', ''),
                           Var1 = c('', '', '', '', '', ''),
                           Var2 = c('', '', '', '', '', ''),
                           UnitName = c('', '', '', '', '','cp09'),
                           InFiles = rep('FF', 6)),
    Aggregates = data.table(IDQual = c('Province', 'NACE', 'Market', ''),
                            NonIDQual = rep('', 4),
                            IDDD = c('', '', '', 'TotalTurnover'),
                            Province = c('', '', '', '.'),
                            NACE = c('', '', '', '.'),
                            Market = c('', '', '', '2.'),
                            UnitName = c('provincia', 'actividad', '', 'cn01'),
                            InFiles = rep('FP', 4)))
VNC1 <- BuildVNC(VarList1)

VarList2 <- list(
    ID = data.table(IDQual = c('NumIdEst', rep('', 4)),
                    NonIDQual = rep('', 5),
                    IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                    NumIdEst = c('', rep('.', 4)),
                    UnitName = c('numidest', 'nombre', 'apellidos', 'direccion', 'telefono'),
                    InFiles = rep('FI', 5)),
    MicroData = data.table(IDQual = c('NumIdEst', '', '', '', '', ''),
                           NonIDQual = c('', 'Market', 'Cod','Var1', 'Var2', ''),
                           IDDD = c('', '', '', '', '','IEPEntradaPed'),
                           NumIdEst = c('', '', '', '', '', '.'),
                           Market = c('', '', '', '', '','3.'),
                           Cod = c('', '', '', '', '', ''),
                           Var1 = c('', '', '', '', '', ''),
                           Var2 = c('', '', '', '', '', ''),
                           UnitName = c('', '', '', '', '','cp09'),
                           InFiles = rep('FF', 6)),
    Aggregates = data.table(IDQual = c('Province', 'NACE', 'Market', ''),
                            NonIDQual = rep('', 4),
                            IDDD = c('', '', '', 'TotalTurnover'),
                            Province = c('', '', '', '.'),
                            NACE = c('', '', '', '.'),
                            Market = c('', '', '', '2.'),
                            UnitName = c('provincia', 'actividad', '', 'cn01'),
                            InFiles = rep('FA', 4)))

VNC2 <- BuildVNC(VarList2)

VNC1 + VNC2

}
