% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.Datadt.R
\docType{methods}
\name{+,Datadt,Datadt-method}
\alias{+,Datadt,Datadt-method}
\title{Method \code{+} for the class \linkS4class{Datadt}}
\usage{
\S4method{+}{Datadt,Datadt}(e1, e2)
}
\arguments{
\item{e1}{Object of class \code{Datadt}.}

\item{e2}{Object of class \code{Datadt}.}
}
\value{
Object of class \code{Datadt}.
}
\description{
\code{+} combines two objects of class \linkS4class{Datadt} in a single object of 
the same class.

This method overloads the operator \code{\link{+}} and builds a new \linkS4class{Datadt} object 
combining both input objects.
}
\examples{
# We build two trivial data sets:
library(data.table)
Ddt1 <- new(Class = 'Datadt', 
        data.table(ID = c('001', '001', '001', '001'), 
                   Market = c('1.', '2.', '', ''),
                   IDDD = c('Turnover', 'Turnover', 'Province', 'NACE09'),
                   Value = c('625000', '23154', '04', '0512')))

Ddt2 <- new(Class = 'Datadt', 
        data.table(ID = c('001', '002', '002'), 
                   IDDD = c( 'Turnover', 'CCAA', 'IASSCifraNeg'),
                   Value = c('648154', '25', '45645')))

Ddt1 + Ddt2

}

