% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.DD.R
\docType{methods}
\name{+,DD,DD-method}
\alias{+,DD,DD-method}
\title{Method \code{+} for the class \linkS4class{DD}}
\usage{
\S4method{+}{DD,DD}(e1, e2)
}
\arguments{
\item{e1}{Object of class \linkS4class{DD}.}

\item{e2}{Object of class \linkS4class{DD}.}
}
\value{
Object of class \linkS4class{DD} resulting from integrating both
\linkS4class{DD} objects in a single \linkS4class{DD} object.
}
\description{
\code{+} sums two objects of class \linkS4class{DD}. This method
overloads the operator \link{+} and returns a new object of class
\linkS4class{DD}.
}
\examples{
library(data.table)
DD1 <- new(Class = 'DD',
           Data = data.table(Variable = c('NOrden', 'CCAA', 'CifraNeg'),
                             Sort = c('IDQual', 'NonIDQual', 'IDDD'),
                             Class = c('character', 'character', 'numeric'),
                             Qual1 = c('', '', 'NOrden')))
DD2 <- new(Class = 'DD',
           Data = data.table(Variable = c('logCifraNeg'),
                             Sort = c('IDDD'),
                             Class = c('numeric'),
                             Qual1 = c('NOrden')))
DD1 + DD2

}

