% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.DDdt.R
\docType{methods}
\name{+,DDdt,DDdt-method}
\alias{+,DDdt,DDdt-method}
\title{Method \code{+} for the class \linkS4class{DDdt}}
\usage{
\S4method{+}{DDdt,DDdt}(e1, e2)
}
\arguments{
\item{e1}{Object of class \linkS4class{DDdt}.}

\item{e2}{Object of class \linkS4class{DDdt}.}
}
\value{
Object of class \linkS4class{DDdt} resulting from integrating both
\linkS4class{DDdt} objects in a single \linkS4class{DDdt} object.
}
\description{
\code{+} sums two objects of class \linkS4class{DDdt}. This method
overloads the operator \link{+} and returns a new object of class
\linkS4class{DDdt}.
}

