% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.VNCdt.R
\docType{methods}
\name{+,VNCdt,VNCdt-method}
\alias{+,VNCdt,VNCdt-method}
\title{Method \code{+} for the class \linkS4class{VNCdt}}
\usage{
\S4method{+}{VNCdt,VNCdt}(e1, e2)
}
\arguments{
\item{e1}{Object of class \linkS4class{VNCdt}.}

\item{e2}{Object of class \linkS4class{VNCdt}.}
}
\value{
Object of class \linkS4class{VNCdt} resulting from 
integrating both \linkS4class{VNCdt} objects in a single 
\linkS4class{VNCdt} object.
}
\description{
\code{+} sums two objects of class \linkS4class{VNCdt}.
 This method overloads the operator \link{+} and returns a new object of 
 class \linkS4class{VNCdt}.

The integration is carried out according to the names of the components.
}
\examples{
library(data.table)

ID1 <- new(Class = "VNCdt", 
          .Data = data.table(IDQual = c('NumIdEst', rep('', 4)),
                             NonIDQual = rep('', 5),
                             IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                             NumIdEst = c('', rep('.', 4)),
                             UnitName = c('numidest', 'nombre', 'apellidos', 'direccion', 
                                       'telefono'),
                             InFiles = rep('FI', 5)))
                                           
ID2 <- new(Class = 'VNCdt',
          .Data = data.table(IDQual = c('NumIdEst', rep('', 4)),
                             NonIDQual = c(rep('', 5)),
                             IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                             NumIdEst = c('', rep('.', 4)),
                             UnitName = c('numidest', 'nombre', 'apellidos', 'direccion', 
                                       'telefono'),
                             InFiles = rep('FI', 5)))
                                           
ID1 + ID2                                            
                                           
MicroData1 <- new(Class = "VNCdt",
                  .Data = data.table(IDQual = c('NumIdEst', rep('', 2)),
                                     NonIDQual = c('', 'Market', ''),
                                     IDDD = c(rep('', 2), 'Turnover'),
                                     NumIdEst = c(rep('', 2), '.'),
                                     Market = c(rep('', 2), '1.'),
                                     UnitName = c('numidest', '', 'cn05'),
                                     InFiles = rep('FF, FD, FG', 3)))
                                     
MicroData2 <- new(Class ='VNCdt',
                  .Data = data.table(IDQual = c('NumIdEst', rep('', 2)),
                                     NonIDQual = c('', 'Market', ''),
                                     IDDD = c(rep('', 2), 'NewOrders'),
                                     NumIdEst = c(rep('', 2), '.'),
                                     Market = c(rep('', 2), '3.'),
                                     UnitName = c('numidest', '', 'cp09'),
                                     InFiles = rep('FF, FD, FG', 3)))
                                     
MicroData1 + MicroData2                                      
                                     
Aggregates1 <- new(Class = "VNCdt",
                   .Data = data.table(IDQual = c('Province', 'NACE', 'Market', ''),
                                      NonIDQual = c(rep('', 4)),
                                      IDDD = c('', '', '', 'Turnover'),
                                      Province = c('', '', '', '.'),
                                      NACE = c('', '', '', '.'),
                                      Market = c('', '', '', '2.'),
                                      UnitName = c('provincia', 'actividad', '', 'cn01'),
                                      InFiles = rep('FA', 4)))


Aggregates2 <- new(Class = 'VNCdt',
                  .Data = data.table(IDQual = c('Province', 'NACE', 'Market', ''),
                                     NonIDQual = c('', '', '', ''),              
                                     IDDD = c('', '', '', 'NewOrders'),
                                     Province = c('', '', '', '.'),
                                     NACE = c('', '', '', '.'),
                                     Market = c('', '', '', '1.'),
                                     UnitName = c('provincia', 'actividad', '', 'cp02'),
                                     InFiles = rep('FA', 4))) 

Aggregates1 + Aggregates2                                                   


}

