% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawStQList.R
\name{rawStQList}
\alias{rawStQList}
\title{S4 class to implement a named list of objects of class \linkS4class{rawStQ}}
\usage{
rawStQList(Data = list(), Periods = new(Class = "RepoTimeInt"))
}
\description{
Definition of an S4 class of name \code{rawStQList} with a list of objects of class
\linkS4class{rawStQ}. Instead of an atomic vector as a name, we associate with it an object of
class \linkS4class{RepoTimeInt} containing the names of the time periods corresponding to each
\linkS4class{rawStQ} object.

The structure of this S4 class \code{rawStQList} comprises 2 attributes:

\itemize{

\item the attribute \code{Data}, which is a \code{list} of objects of class \linkS4class{rawStQ};

\item  the attribute \code{Periods}, which is an object of class \linkS4class{RepoTimeInt}.
}
}
\section{Slots}{

\describe{
\item{\code{Data}}{\code{List} of objects of class \linkS4class{rawStQ};}

\item{\code{Periods}}{An object of class \linkS4class{RepoTimeInt};}
}}
\examples{
library(RepoTime)
mm <- c(paste0('0', 1:9), 10:12)
TimePer <- paste0('MM', mm, '2015')
PeriodList <- newRepoTime(TimePer)
EmptyrawQ <- rawStQ()
EmptyrawQList <- vector('list', 12)
EmptyrawQList <- lapply(EmptyrawQList, function(x) EmptyrawQ)
rawQList <- rawStQList(Data = EmptyrawQList, Periods = PeriodList)
rawQList
str(rawQList)

}

