% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawStQToStQ.R
\docType{methods}
\name{rawStQToStQ}
\alias{rawStQToStQ}
\alias{rawStQToStQ,rawStQ-method}
\title{Generates an object of class \linkS4class{StQ} from the input
\linkS4class{rawStQ} object.}
\usage{
rawStQToStQ(rawQ)

\S4method{rawStQToStQ}{rawStQ}(rawQ)
}
\arguments{
\item{rawQ}{Object of class \linkS4class{rawStQ} whose slot Data has the
key-value pair structure.}
}
\value{
Object of class \linkS4class{StQ} whose slot Data has at least the 
columns \code{IDDD} and \code{Value}.
}
\description{
\code{rawStQToStQ} returns an object of class \linkS4class{StQ}
from the input\linkS4class{rawStQ} object.
}
\examples{
library(data.table)
data(ExampleDD)
key <- new(Class = 'rawKey', 
           c('Turnover@001@@', 
             'Employees@001@1@0', 
             'Employees@001@0@', 
             'Employees@001@1@1'))
rawData <- new(Class = 'rawDatadt', 
               data.table(Key = key, Value = c('625000', '23154', '25004', '10512')))
rawQ <- new(Class = 'rawStQ', Data = rawData, DD = ExampleDD)

newQ <- rawStQToStQ(rawQ)
}

