% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ReadRepoFile.R
\name{ReadRepoFile}
\alias{ReadRepoFile}
\title{Lectura de un fichero del repositorio}
\usage{
ReadRepoFile(FileName)
}
\arguments{
\item{FileName}{\code{\link{vector}} de tipo \code{character} de longitud 1
con el nombre del fichero que se desea leer. El nombre debe incluir la ruta
completa de ubicación del fichero (directorio en el que se encuentra el
fichero y su nombre).}
}
\value{
\code{data.table} con los datos del fichero leído.
}
\description{
\code{ReadRepoFile} devuelve un \code{\link{data.table}} con el
contenido del fichero del repositorio especificado.
}
\details{
Esta función permite la lectura del ficheros ASCII ubicados en el repositorio
de la encuesta.
}
\examples{
# Se asume que el fichero con estructura par clave-valor,
# ASCII \\code{E30183.FF_V1.MM032014.D_1}, se encuentra en el escritorio del
# administrador:
RepoName <- 'C:/Users/Administrador/Desktop/E30183.FF_V1.MM032014.D_1'
Example.kv <- ReadRepoFile(RepoName)
str(Example.kv)
}
\seealso{
\link{ReadSASFile}, \link{WriteRepoFile}
}

