% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ReadSASFile.R
\name{ReadSASFile}
\alias{ReadSASFile}
\title{Lectura de un fichero SAS de microdatos}
\usage{
ReadSASFile(SASFileName, Exceldf, DD)
}
\arguments{
\item{SASFileName}{\code{\link{vector}} de tipo \code{character} de longitud 1 con
la ruta completa de ubicación del fichero SAS que se desea leer.}

\item{Exceldf}{\code{\link{data.frame}} con el contenido del fichero Excel
de equivalencias entre los nombres de las variables de la encuesta asignados
por las distintas unidades del INE implicadas.}

\item{DD}{\code{\link{data.frame}} con el contenido del fichero \code{DD} de
definición de datos de la encuesta, en su versión para R.}
}
\value{
\code{\link{data.table}} con los datos del fichero SAS leído, con las
unidades en filas y las variables en columnas.
}
\description{
\code{ReadSASFile} devuelve un \code{\link{data.table}} con los datos del
fichero SAS especificado.
}
\details{
Esta función lee el fichero SAS especificado en \code{SASFileName} y devuelve
su contenido en un \code{data.table} con estructura matricial (unidades en
filas y variables en columnas). Los nombres de las variables se asignan de
acuerdo a los especificados en el fichero Excel que se encuentra en el
repositorio de la encuesta considerada.
Para llevar a cabo esta asignación de nombres es necesario especificar también
los argumentos de entrada \code{Exceldf}, que es un \code{data.frame} con el
contenido de la correspondiente hoja del fichero Excel \code{Ennnnn.NombresVariables.xls}
del repositorio de la encuesta, y \code{DD} que es el fichero \code{DD} de
definición de datos de la encuesta en su formato para R.

El formato del fichero \code{DD} para R lo obtendremos a partir del fichero ASCII
\code{Ennnnn.DD_Vn} incluido en el repositorio de la encuesta haciendo uso de la
función \code{\link{RepoDDToDD}} del paquete \code{\linkS4class{StQ}}.
}
\examples{
# Se asume que el fichero \\code{SAS MM032014.sas7bdat} se encuentra en el escritorio
# del administrador:
SASName <- 'C:/Users/Administrador/Desktop/MM032014.sas7bdat'
data(XLS)
data(RepoDD)
DD <- RepoDDToDD(RepoDD)
Example.DM <- ReadSASFile(SASName, XLS, DD)
}
\seealso{
\link{RepoDDToDD}, \link{ReadRepoFile}, \link{WriteRepoFile}
}

