% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/renVar.StQ.R
\docType{methods}
\name{renVar}
\alias{renVar}
\alias{renVar,StQ,character,character-method}
\title{Rename variables in an object}
\usage{
renVar(object, VarNames, NewVarNames)

\S4method{renVar}{StQ,character,character}(object, VarNames, NewVarNames)
}
\arguments{
\item{object}{Input object.}

\item{VarNames}{Character vector with the variable names to be renamed.}

\item{NewVarNames}{Character vector with the new variable names.}
}
\value{
Object with the same class as the input object and all specified
variable names duly renamed in all slots.
}
\description{
\code{renVar} rename variables in a object keeping the rest of
the object intact.

This method returns the input object with the specified variable names in the
input parameter \code{VarNames} renamed with the names in the input parameter
\code{NewVarNames}. Variables are renamed in all slots.
}
\examples{
data(ExampleStQ)
renVar(ExampleStQ, 'Stocks', 'NewStocks')
}

