% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RepoDDToDD.R
\name{RepoDDToDD}
\alias{RepoDDToDD}
\title{Transforma el fichero DD del repositorio a su formato para R}
\usage{
RepoDDToDD(RepoDD)
}
\arguments{
\item{RepoDD}{\code{data.frame} con el contenido del fichero
\code{DD} de definición de datos ubicado en el repositorio de microdatos de
la encuesta en su formato para R.}
}
\value{
\code{data.table} con los datos del fichero \code{DD}
en el formato adecuado para su inclusión como slot \code{DD} de un objeto de
clase \code{StQ}.
}
\description{
Esta función es un constructor para la clase \linkS4class{DD} a partir del
contenido del fichero DD correspondiente en el repositorio.
}
\details{
\code{RepoDDToDD} transforma un \code{\linkS4class{data.frame}} con el contenido
del fichero DD de definición de datos ubicado en el repositorio de microdatos
de la encuesta, en un \code{\linkS4class{data.table}} con el formato requerido para
su inclusión como slot \code{DD} de un objeto de clase \code{\linkS4class{StQ}}.

Esta función toma el contenido del \code{data.frame} de entrada y
construye un \code{data.table} con las columnas \code{Variable},
\code{Sort}, \code{Class} and \code{Qual\code{1}} to \code{Qual\code{q}}.

La columna \code{Variable} contiene los nombres de todas las variables, tanto
de las del cuestionario como de las creadas durante el proceso.

La columna \code{Sort} toma los valores \code{'IDQual'}, \code{'NonIDQual'},
e \code{'IDDD'}, para calificadores de identificación de las unidades
estadísticas, para calificadores de nombres de variables y para nombres de
variables otras variables en el cuestionario.

La columna \code{Class} especifica la clase de la variable, y toma los valores
\code{numeric} o \code{character}.

Las columnas \code{Qual\code{1}} a \code{Qual\code{q}} contienen los nombres de los
calificadores de cada variable (fila).

Se transforma el fichero de definición de datos para que su tratamiento desde
el punto de vista de la programación sea más eficiente.
}
\examples{
data(RepoDD)
RepoDDToDD(RepoDD)
}

