% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RepoFileVersion.R
\name{RepoFileVersion}
\alias{RepoFileVersion}
\title{Número de la última versión de un fichero en un directorio}
\usage{
RepoFileVersion(Path, FileNameString)
}
\arguments{
\item{Path}{\code{Vector} de tipo \code{character} de longitud uno con la
ruta del directorio en el que queremos buscar el fichero.}

\item{FileNameString}{\code{Vector} de tipo \code{character} de longitud uno con
el nombre del fichero para el que queremos determinar el número de la última
versión disponible en el directorio indicado en \code{Path}.}
}
\value{
\code{Vector} de tipo \code{integer} de longitud uno con el número de
la última versión del fichero especificado que existe en el directorio de
entrada.
}
\description{
\code{RepoTopn} obtiene el número de la última versión que existe de un fichero
en un determinado directorio.
}
\details{
Esta función busca en el directorio especificado en el parámetro de entrada
\code{Path}, la última versión que existe del fichero indicado en \code{FileNameString}
y devuelve el número correspondiente a dicha versión.

Esta función está pensada para ficheros que tiene anexada la versión a la
que corresponden sus datos al final de su nombre (versión definitiva o parcial).
}
\examples{
\dontrun{
RepoTopn('R:/E30183', 'FF_V1.MM122014')
}
}

