% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RepoNextFileVerNumber.R
\name{RepoNextFileVerNumber}
\alias{RepoNextFileVerNumber}
\title{Creación del nombre de una nueva versión de un fichero}
\usage{
RepoNextFileVerNumber(Periods, Path, FileType)
}
\arguments{
\item{Periods}{\code{Vector} de tipo \code{character} con los periodos para
los que queremos generar los nombres de la última versión de los ficheros.}

\item{Path}{\code{Vector} de tipo \code{character} con la ruta donde queremos
buscar los ficheros.}

\item{FileType}{\code{Vector} de tipo \code{character} con el tipo de fichero
que queremos buscar.}
}
\value{
Devuelve un \code{Vector} de tipo \code{character} de longitud igual
al número de periodos especificados en \code{Periods} con los nombres completos
de los ficheros de tipo \code{FileType} que corresponde a las siguiente versión
que corresponde escribir para cada uno de periodos.
}
\description{
\code{RepoNextFileVerNumber} devuelve el nombre que debe tener la siguiente
versión de un fichero.
}
\details{
Esta función toma los ficheros ubicados en el directorio indicado en el
parámetro \code{Path} correspondientes al tipo especificado en el input \code{FileType}
y para los periodos en \code{Periods} y genera para cada uno de ellos los
nombres de las siguientes versiones de cada uno de ellos que corresponde
generar.
}
\examples{
\dontrun{
RepoTopn(Path, Type)
}
}

