% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RepoPeriodRange.R
\name{RepoPeriodRange}
\alias{RepoPeriodRange}
\title{Obtención de los períodos de un tipo dado de ficheros en un directorio}
\usage{
RepoPeriodRange(Path, Type, Last = out[length(out)])
}
\arguments{
\item{Path}{\code{Vector} de tipo \code{character} de longitud uno, con la
ruta del directorio en el que queremos buscar los ficheros del tipo \code{Type}.}

\item{Type}{\code{Vector} de tipo \code{character} de longitud uno con el
tipo del fichero del que queremos saber los periodos en el directorio indicado
en \code{Path}. Puede tomar los valores \code{FF}, \code{FG}, \code{FD},
\code{FL} o \code{FT}.}

\item{Last}{\code{Vector} de tipo \code{character} de longitud uno con el último
periodo que queremos obtener del tipo de fichero especificado.}
}
\value{
\code{Vector} de tipo \code{character} con los distintos periodos que
existen en el directorio de entrada del tipo de fichero especificado.
}
\description{
\code{RepoPeriodRange} obtiene los periodos que existen de un tipo concreto de
ficheros en un determinado directorio.
}
\details{
Esta función obtiene para el tipo de ficheros indicado en el input \code{Type},
los distintos periodos que existen en el directorio especificado en el parámetro
de entrada \code{Path}.

Por defecto devuelve todos los periodos del fichero que haya ubicados en el
directorio. El parámetro \code{Last} nos permite solicitar los periodos hasta
uno concreto de ellos.
}
\examples{
\dontrun{
RepoPeriodRange('R:/E30183', 'FF', 'MM102014')
}
}

