% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmVar.R
\docType{methods}
\name{rmVar}
\alias{rmVar}
\alias{rmVar,StQ-method}
\title{Remove variables from an object}
\usage{
rmVar(object, VarNames = character(0))

\S4method{rmVar}{StQ}(object, VarNames = character(0))
}
\arguments{
\item{object}{Input object to remove the variables from.}

\item{VarNames}{Character vector with the variable names to remove.}
}
\value{
Object with the same class as the input object but where the
variables in \code{VarNames} have been removed. Variables are removed only
from slot \code{Data}.
}
\description{
\code{rmVar} removes from the input object those values for variable names in
the input parameter \code{VarNames}.
}
\details{
This method returns the same input object with those variables specified in
the input parameter \code{VarNames} removed from the data set.
}
\examples{
data(ExampleQ)
rmVar(ExampleQ)
rmVar(ExampleQ, 'IASSTame')

}

