% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAggr.R
\docType{methods}
\name{setAggr<-}
\alias{setAggr<-}
\alias{setAggr<-,DD,data.table-method}
\title{Set value of slot \code{Aggregates}}
\usage{
setAggr(object) <- value

\S4method{setAggr}{DD,data.table}(object) <- value
}
\arguments{
\item{object}{Object whose slot \code{Aggregates} is to be assigned.}

\item{value}{\linkS4class{data.table} to be assigned to the slot \code{Aggregates}.}
}
\value{
Object with slot Aggregates updated.
}
\description{
\code{setAggr} assigns a \linkS4class{data.table} to the slot
\code{Aggregates} of the input object.
}
\examples{
# We build an empty data.table:
library(data.table)
Aggregates <- data.table(IDDD = character(0), Value = numeric(0))

# We assign this data.table to the slot Aggregates of object NewExampleQ:
data(ExampleQ)
NewExampleQ <- ExampleQ
setAggr(NewExampleQ) <- Aggregates
NewExampleQ

}

