% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAggregates.R
\docType{methods}
\name{setAggregates<-}
\alias{setAggregates<-}
<<<<<<< HEAD
\alias{setAggregates<-,DD,data.table-method}
\alias{setAggregates<-,VNC,data.table-method}
=======
\alias{setAggregates<-,DD,DDdt-method}
\alias{setAggregates<-,StQ,DDdt-method}
>>>>>>> 5034523f22c62817420f2f5687369d62b4523cd8
\title{Set value of slot \code{Aggregates}}
\usage{
setAggregates(object) <- value

<<<<<<< HEAD
\S4method{setAggregates}{VNC,data.table}(object) <- value

\S4method{setAggregates}{DD,data.table}(object) <- value
=======
\S4method{setAggregates}{DD,DDdt}(object) <- value

\S4method{setAggregates}{StQ,DDdt}(object) <- value
>>>>>>> 5034523f22c62817420f2f5687369d62b4523cd8
}
\arguments{
\item{object}{Object whose slot \code{Aggregates} is to be assigned.}

<<<<<<< HEAD
\item{value}{\linkS4class{data.table} to be assigned to the slot \code{Aggregates}.}
=======
\item{value}{\linkS4class{DDdt} to be assigned to the slot \code{Aggregates}.}
>>>>>>> 5034523f22c62817420f2f5687369d62b4523cd8
}
\value{
Object with slot Aggregates updated.
}
\description{
<<<<<<< HEAD
\code{setAggregates} assigns a \linkS4class{data.table} to the slot \code{Aggregates} of 
=======
\code{setAggregates} assigns a \linkS4class{DDdt} to the slot \code{Aggregates} of 
>>>>>>> 5034523f22c62817420f2f5687369d62b4523cd8
the input object.
}
\examples{
library(data.table)
MicroDataDD <- data.table(Variable = 'NewOrders', Sort = 'IDDD', Class = 'numeric',
                          Length = '8',
                          Qual1 = 'NumIdEst', Qual2 = 'Market', ValueRegExp = '')
<<<<<<< HEAD
VarList <- list(MicroData = data.table(IDQual = c('NumIdEst','','',''),
                                       NonIDQual = c('', 'Market', 'Cod', ''),
                                       IDDD = c('', '', '' ,'NewOrders'),
                                       NumIdEst = c('', '', '', '.'),
                                       Market = c('', '', '', '1.'),
                                       Cod = rep('', 4),
                                       UnitName = c('', '', '', 'cp02'),
                                       InFiles = rep('FF', 4)))
VNC <- BuildVNC(VarList)
DD <- BuildDD(list(VNC = VNC, MicroData = MicroDataDD))
=======
MicroDataDD <- new(Class = 'DDdt', MicroDataDD)
VarList <- list(MicroData = new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst','','',''),
                                                          NonIDQual = c('', 'Market', 'Cod', ''),
                                                          IDDD = c('', '', '' ,'NewOrders'),
                                                          NumIdEst = c('', '', '', '.'),
                                                          Market = c('', '', '', '1.'),
                                                          Cod = rep('', 4),
                                                          UnitName = c('', '', '', 'cp02'),
                                                          InFiles = rep('FF', 4))))
VNC <- BuildVNC(VarList)
DD <- new(Class = 'DD', VarNameCorresp = VNC, MicroData = MicroDataDD)
>>>>>>> 5034523f22c62817420f2f5687369d62b4523cd8
          
Aggdt <- data.table(Variable = c('NACE09', 'Turnover'),
                    Sort = c('IDQual', 'IDDD'),
                    Class = c('character', 'character'),
                    Length = c('4', '10'),
                    Qual1 = c('', 'NACE09'),
                    ValueRegExp = c('([0-4][0-9])|(5[0-2])', '([0-9]{1, 15}| )'))  
<<<<<<< HEAD
=======
Aggdt <- new(Class = 'DDdt', Aggdt)          
>>>>>>> 5034523f22c62817420f2f5687369d62b4523cd8
setAggregates(DD) <- Aggdt
DD

}

