% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAggrWeights.R
\docType{methods}
\name{setAggrWeights<-}
\alias{setAggrWeights<-}
\alias{setAggrWeights<-,DD,DDdt-method}
\title{Set value of slot \code{AggrWeights}}
\usage{
setAggrWeights(object) <- value

\S4method{setAggrWeights}{DD,DDdt}(object) <- value
}
\arguments{
\item{object}{Object whose slot \code{AggrWeights} is to be assigned.}

\item{value}{\linkS4class{DDdt} to be assigned to the slot \code{AggrWeights}.}
}
\value{
Object with slot AggrWeights updated.
}
\description{
\code{setAggrWeights} assigns a \linkS4class{DDdt} to the slot
\code{AggrWeights} of the input object.
}
\examples{
# We build an empty data.table:
library(data.table)
AggrWeights <- data.table(IDDD = character(0), Value = numeric(0))

# We assign this data.table to the slot AggrWeights of object NewExampleQ:
data(ExampleQ)
NewExampleQ <- ExampleQ
setAggrWeights(NewExampleQ) <- AggrWeights
NewExampleQ

}

