% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setData.R
\docType{methods}
\name{setData<-}
\alias{setData<-}
<<<<<<< HEAD
\alias{setData<-,StQ,data.table-method}
\alias{setData<-,rawStQ,data.table-method}
=======
\alias{setData<-,StQ,Datadt-method}
>>>>>>> 5034523f22c62817420f2f5687369d62b4523cd8
\title{Set value of slot \code{Data}}
\usage{
setData(object) <- value

<<<<<<< HEAD
\S4method{setData}{StQ,data.table}(object) <- value

\S4method{setData}{rawStQ,data.table}(object) <- value
=======
\S4method{setData}{StQ,Datadt}(object) <- value
>>>>>>> 5034523f22c62817420f2f5687369d62b4523cd8
}
\arguments{
\item{object}{Object whose slot \code{Data} is to be assigned.}

\item{value}{\linkS4class{Datadt} to be assigned to the slot \code{Data}.}
}
\value{
Object with slot Data updated.
}
\description{
\code{setData} assigns a \linkS4class{Datadt} to the slot \code{Data} of the input
object.
}
\examples{
# We build an empty data.table:
library(data.table)
Data <- new(Class = 'Datadt', data.table(IDDD = character(0), Value = numeric(0)))

# We assign this data.table to the slot Data of object NewExampleStQ:
data(ExampleStQ)
NewExampleStQ <- ExampleStQ
setData(NewExampleStQ) <- Data
NewExampleStQ

}

