% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setDD.R
\docType{methods}
\name{setDD<-}
\alias{setDD<-}
\alias{setDD<-,StQ,DD-method}
\title{Set value of slot \code{DD}.}
\usage{
setDD(object) <- value

\S4method{setDD}{StQ,DD}(object) <- value
}
\arguments{
\item{object}{Object whose slot \code{DD} is to be assigned.}

\item{value}{Object of class \linkS4class{DD} to be assigned to the input
object.}
}
\value{
Object of class \linkS4class{StQ} with slot DD updated.
}
\description{
\code{setDD} assigns an object of class \linkS4class{DD} to the
 slot \code{DD} of the input object.
}
\examples{
# We build an empty StQ object:
library(data.table)
Q <- new(Class = 'StQ')

# We use an DD object previously created:
data(ExampleDD)
# We assign this DD object to slot DD of the preceding StQ object:
setDD(Q) <- ExampleDD
str(Q)
}

