% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setID.R
\docType{methods}
\name{setID<-}
\alias{setID<-}
\alias{setID<-,DD,data.table-method}
\title{Set value of slot \code{ID} of an object \linkS4class{DD}}
\usage{
setID(object) <- value

\S4method{setID}{DD,data.table}(object) <- value
}
\arguments{
\item{object}{Object containing slot \code{ID} to be assigned.}

\item{value}{\linkS4class{data.table} to be assigned to the slot \code{ID}.}
}
\value{
Object \linkS4class{DD} with slot \code{ID} updated.
}
\description{
\code{setID} assigns a \linkS4class{data.table} to the
 slot \code{ID} of the input object.
}
\examples{
# On an empty object of class VarNameCorresp:
VNC <- new(Class = 'VarNameCorresp')
VarList <- list(MicroData = data.table(IDQual = c('NumIdEst','','','',''),
                                       NonIDQual = c('EsMercNac', 
                                                     'EsMercEuro', 
                                                     'EsMercRM',
                                                     'Cod',
                                                     ''),
                                       IDDD = c('', '', '' ,'' ,
                                                'IEPEntradaPed'),
                                       Unit1 = c('', '', '', '', 'cp02')))
VNC[['MicroData']] <- VarList
str(VNC)

# On an object of class DD:
MicroDataDD <- data.table(Variable = 'IEPEntradaPed', 
                          Sort = 'IDDD', 
                          Class = 'numeric',
                          Qual1 = 'NOrden')
DD <- new(Class = 'DD', MicroData = MicroDataDD)
VarList <- list(MicroData = data.table(IDQual = c('NumIdEst','','','',''),
                                       NonIDQual = c('EsMercNac', 
                                                     'EsMercEuro', 
                                                     'EsMercRM',
                                                     'Cod',
                                                     ''),
                                       IDDD = c('', '', '' ,'' ,
                                                'IEPEntradaPed'),
                                       Unit1 = c('', '', '', '', 'cp02')))
setID(DD) <- VarList
DD
str(DD)

# On an object of class StQ:
MicroDataDD <- data.table(Variable = 'IEPEntradaPed', 
                          Sort = 'IDDD', 
                          Class = 'numeric',
                          Qual1 = 'NOrden')
DD <- new(Class = 'DD', MicroData = MicroDataDD)
StQ <- new(Class = 'StQ', 
           Data = data.table(IDDD = character(0), Value = character(0)), 
           DD = DD)
VarList <- list(MicroData = data.table(IDQual = c('NumIdEst','','','',''),
                                       NonIDQual = c('EsMercNac', 
                                                     'EsMercEuro', 
                                                     'EsMercRM',
                                                     'Cod',
                                                     ''),
                                       IDDD = c('', '', '' ,'' ,
                                                'IEPEntradaPed'),
                                       Unit1 = c('', '', '', '', 'cp02')))
setID(StQ) <- VarList
StQ
str(StQ)

}

