% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMicroData.R
\docType{methods}
\name{setMicroData<-}
\alias{setMicroData<-}
\alias{setMicroData<-,DD,DDdt-method}
\alias{setMicroData<-,StQ,DDdt-method}
\title{Set value of slot \code{MicroData} of an object \linkS4class{DD}}
\usage{
setMicroData(object) <- value

\S4method{setMicroData}{DD,DDdt}(object) <- value

\S4method{setMicroData}{StQ,DDdt}(object) <- value
}
\arguments{
\item{object}{Object containing slot \code{ID} to be assigned.}

\item{value}{\linkS4class{DDdt} to be assigned to the slot \code{ID}.}
}
\value{
Object \linkS4class{DD} with slot \code{ID} updated.
}
\description{
\code{setMicroData} assigns a \linkS4class{DDdt} to the
 slot \code{ID} of the input object.
}
\examples{
# An example:
library(data.table)
MicroDataDD <- data.table(Variable = 'IEPEntradaPed', Sort = 'IDDD', Class = 'numeric',
                          Length = '8',
                          Qual1 = 'NumIdEst', Qual2 = 'Market', ValueRegExp = '')
MicroDataDD <- new(Class = 'DDdt', MicroDataDD)
VarList <- list(MicroData = new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst','','',''),
                                                          NonIDQual = c('', 'Market', 'Cod', ''),
                                                          IDDD = c('', '', '' ,'IEPEntradaPed'),
                                                          NumIdEst = c('', '', '', '.'),
                                                          Market = c('', '', '', '1.'),
                                                          Cod = rep('', 4),
                                                          Unit1 = c('', '', '', 'cp02'))))
VNC <- BuildVNC(VarList)
DD <- new(Class = 'DD', VarNameCorresp = VNC, MicroData = MicroDataDD)

Microdt <- data.table(Variable = c('NumIdEst', 'NewOrders'),
                      Sort = c('IDQual', 'IDDD'),
                      Class = c('character', 'character'),
                      Length = c('11', '8'),
                      Qual1 = c('', 'NumIdEst'),
                      ValueRegExp = c('[0-9]{9}PP', '([0-9]{1, 10}| )'))
Microdt <- new(Class = 'DDdt', Microdt)

setMicroData(DD) <- Microdt
DD

}

