% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOther.R
\docType{methods}
\name{setOther<-}
\alias{setOther<-}
\alias{setOther<-,VNC,data.table-method}
\alias{setOther<-,DD,data.table-method}
\title{Set value of slot \code{Other}}
\usage{
setOther(object) <- value

\S4method{setOther}{VNC,data.table}(object) <- value

\S4method{setOther}{DD,data.table}(object) <- value
}
\arguments{
\item{object}{Object whose slot \code{Other} is to be assigned.}

\item{value}{\linkS4class{data.table} to be assigned to the slot \code{Other}.}
}
\value{
Object with slot \code{Other} updated.
}
\description{
\code{setOtherDD} assigns a \linkS4class{data.table} to the slot \code{Other} of the
input object.
}
\examples{
\dontrun{
library(data.table)
MicroDataDD <- data.table(Variable = 'NewOrders', Sort = 'IDDD', Class = 'numeric',
                          Length = '8',
                          Qual1 = 'NumIdEst', Qual2 = 'Market', ValueRegExp = '')
VarList <- list(MicroData = data.table(IDQual = c('NumIdEst','','',''),
                                       NonIDQual = c('', 'Market', 'Cod', ''),
                                       IDDD = c('', '', '' ,'NewOrders'),
                                       NumIdEst = c('', '', '', '.'),
                                       Market = c('', '', '', '1.'),
                                       Cod = rep('', 4),
                                       UnitName = c('', '', '', 'cp02'),
                                       InFiles = rep('FF', 4)))
VNC <- BuildVNC(VarList)
DD <- DD(VNC = VNC, MicroData = MicroDataDD)
          
Otherdt <- data.table(Variable = c('NACE08'), Sort = c('NonIDQual'), Class = c('character'),
                      Length = '4',
                      Qual1 = c(''), ValueRegExp = c('([0-4][0-9])|(5[0-2])'))
setOther(DD) <- Otherdt
DD
}

}
