% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOtherDD.R
\docType{methods}
\name{setOtherDD<-}
\alias{setOtherDD<-}
\alias{setOtherDD<-,StQ,data.table-method}
\title{Set value of slot \code{Other}}
\usage{
setOtherDD(object) <- value

\S4method{setOtherDD}{StQ,data.table}(object) <- value
}
\arguments{
\item{object}{Object whose slot \code{Other} is to be assigned.}

\item{value}{\linkS4class{data.table} to be assigned to the slot \code{Other}.}
}
\value{
Object with slot Other updated.
}
\description{
\code{setOtherDD} assigns a \linkS4class{data.table} to the slot
\code{Other} of the input object.
}
\examples{
# We build an empty data.table:
library(data.table)
Other <- data.table(IDDD = character(0), Value = numeric(0))

# We assign this data.table to the slot Other of object NewExampleQ:
data(ExampleQ)
NewExampleQ <- ExampleQ
setOtherDD(NewExampleQ) <- Other
NewExampleQ

}

