% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setValues.R
\name{setValues}
\alias{setValues}
\alias{setValues,StQ,DD-method}
\alias{setValues,StQList,DD-method}
\title{Method to set values of a variable}
\usage{
setValues(object, newDD, DDslot = "MicroData", Value, lag = NULL, by = NULL)

\S4method{setValues}{StQ,DD}(object, newDD, DDslot = "MicroData", Value, lag = NULL, by = NULL)

\S4method{setValues}{StQList,DD}(object, newDD, DDslot = "MicroData", Value, lag = NULL, by = NULL)
}
\arguments{
\item{object}{Object in which to include the new variable.}

\item{newDD}{\link{DD} object with the information of the new variable needed for the slot
\code{DD} of the input object.}

\item{DDslot}{Character vector of length 1 with the name of the slot of newDD in which the new 
variable is defined. Its default value is \code{MicroData}.}

\item{Value}{Vector of length the number of statistical units in the input object with the values
of the new variable for each of these units or alternatively an object of class 
\code{\link{call}} with a mathematical formula for the computation of these values.}

\item{lag}{Integer vector of length 1 with the lag to use when the input object is of class 
\linkS4class{StQList}. Its default value is \code{NULL}. If its value is not \code{NULL}, it is
not possible to use an object of class \code{\link{call}} in "Value".}

\item{by}{Character vector with the names of the variables specifying those statistical units 
groups under which the mathematical formula will be applied. Its default value is \code{NULL}.}
}
\value{
Object of the same class as the input object with the new variable included.
}
\description{
\code{setValues} sets the values of a variable in the input object.

This method creates a variable with the name \code{VarName} with values specified in the input 
\code{Value}. If the variable is already present in the input variable, the method overwrites it.

It is also necessary to provide as input parameter an object of class \link{DD} named 
\code{newDD} with the row corresponding to the variable set with this method to be included in
the slot \code{DD}. This row is specified as a \linkS4class{data.table} with a single row and 
columns Variable, Sort, Class, Qual1-Qualn (as many as so many qualifiers are necessary) and
ValueRegExp.

Input parameters \code{by} and \code{lag} are optional and specify how to compute the values of 
the new variable in some circunstamces.
}
\examples{
\dontrun{
library(data.table)
data(ExampleStQ)
newVNCVar <- data.table(IDQual = '', NonIDQual = '', IDDD = 'lTurnover',
                            UnitName = '', InFiles = 'FF')
newVNC <- BuildVNC(list(MicroData = newVNCVar))
newDD <- DD(VNC = newVNC, 
            MicroData = data.table(Variable = 'lTurnover',
                                   Sort = 'IDDD',
                                   Class = 'numeric',
                                   Length = '10',
                                   Qual1 = 'ID',
                                   ValueRegExp = ''))
NewQ <- setValues(object = ExampleStQ,
               newDD = newDD,
               DDslot = 'MicroData',
               Value = log(1 + as.numeric(Turnover)))
getValues(NewQ, 'lTurnover')
}
}
