% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setVariable.R
\docType{methods}
\name{setVariable}
\alias{setVariable}
\alias{setVariable,StQ,Variable-method}
\alias{setVariable,StQList,Variable-method}
\title{Generate a new variable}
\usage{
setVariable(object, Var, lag = NULL, by = NULL)

\S4method{setVariable}{StQ,Variable}(object, Var, lag = NULL,
  by = NULL)

\S4method{setVariable}{StQList,Variable}(object, Var, lag = NULL,
  by = NULL)
}
\arguments{
\item{object}{Object of class \link{StQ} or \linkS4class{StQList}.}

\item{Var}{Object of class \linkS4class{Variable}.}
}
\value{
Input object including the new variable.
}
\description{
\code{setVariable} .
}
\examples{
data(ExampleStQ)
Var <- new(Class = 'Variable', UnitName = 'LRTEmp',
                               IDDD = 'Employees',
                               QualsValues = list(ID = '', EmplType = ''),
                               Length = '8',
                               ClassVar = 'numeric',
                               ValueRegExp = '[.]+',
                               Formula = as.call(list('log( 1 + (Turnover / (Employees_1. + Employees_2.1.)))')),
                               SlotName = 'MicroData',
                               Literal = '',
                               DDversion = '1')
ExampleStQ <- setVariable(ExampleStQ, Var)

ExampleStQList <- BuildStQList(list(MM012012 = ExampleStQ, MM022012 = ExampleStQ))
ExampleStQList <- setVariable(ExampleStQList, Var)

}
