% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setVNC.R
\docType{methods}
\name{setVNC<-}
\alias{setVNC<-}
\alias{setVNC<-,DD,list-method}
\alias{setVNC<-,StQ,list-method}
\title{Set value of slot \code{VarNameCorresp}.}
\usage{
setVNC(object) <- value

\S4method{setVNC}{DD,list}(object) <- value

\S4method{setVNC}{StQ,list}(object) <- value
}
\arguments{
\item{object}{Object containing slot \code{VarNameCorresp} to be assigned.}

<<<<<<< HEAD
\item{value}{List of \linkS4class{VNCdt}s to be assigned to the slot 
||||||| merged common ancestors
\item{value}{List of \linkS4class{data.table}s to be assigned to the slot
=======
\item{value}{List of \linkS4class{data.table}s to be assigned to the slot 
>>>>>>> e3ac9308c1978eb314de9c1fda9971de40a9bce7
\code{VarNameCorresp}.}
}
\value{
Object with slot \code{VarNameCorresp} updated.
}
\description{
\code{setVNC} assigns a list of \linkS4class{VNCdt}s to the
 slot \code{VarNameCorresp} of the input object.
}
\examples{
# On an empty object of class VarNameCorresp:
VNC <- new(Class = 'VarNameCorresp')
VarList <- list(MicroData = data.table(IDQual = c('NumIdEst','','','',''),
                                       NonIDQual = c('EsMercNac', 
                                                     'EsMercEuro', 
                                                     'EsMercRM',
                                                     'Cod',
                                                     ''),
                                       IDDD = c('', '', '' ,'' ,
                                                'IEPEntradaPed'),
                                       Unit1 = c('', '', '', '', 'cp02')))
VNC[['MicroData']] <- VarList
str(VNC)

# On an object of class DD:
MicroDataDD <- data.table(Variable = 'IEPEntradaPed', 
                          Sort = 'IDDD', 
                          Class = 'numeric',
                          Qual1 = 'NOrden')
DD <- new(Class = 'DD', MicroData = MicroDataDD)
VarList <- list(MicroData = data.table(IDQual = c('NumIdEst','','','',''),
                                       NonIDQual = c('EsMercNac', 
                                                     'EsMercEuro', 
                                                     'EsMercRM',
                                                     'Cod',
                                                     ''),
                                       IDDD = c('', '', '' ,'' ,
                                                'IEPEntradaPed'),
                                       Unit1 = c('', '', '', '', 'cp02')))
setVNC(DD) <- VarList
DD
str(DD)

# On an object of class StQ:
<<<<<<< HEAD
library(data.table)
data(Q)
VarList <- list(Aggregates = new(Class = 'VNCdt',
                                data.table(IDQual = c('Province','NACE09',
                                                      'IsNatMarket',''),
                                           NonIDQual = rep('', 4),
                                           IDDD = c('', '', '' ,'Turnover'),
                                           Province = c('', '', '', '.'),
                                           NACE09 = c('', '', '', '.'),
                                           IsNatMarket = c('', '', '', '1'),
                                           Unit1 = c('provincia', 'actividad',
                                                     '', 'cn01'))))
setVNC(Q) <- VarList
Q
str(Q)

||||||| merged common ancestors
MicroDataDD <- data.table(Variable = 'IEPEntradaPed',
                          Sort = 'IDDD',
                          Class = 'numeric',
                          Qual1 = 'NOrden')
DD <- new(Class = 'DD', MicroData = MicroDataDD)
StQ <- new(Class = 'StQ',
           Data = data.table(IDDD = character(0), Value = character(0)),
           DD = DD)
VarList <- list(MicroData = data.table(IDQual = c('NumIdEst','','','',''),
                                       NonIDQual = c('EsMercNac',
                                                     'EsMercEuro',
                                                     'EsMercRM',
                                                     'Cod',
                                                     ''),
                                       IDDD = c('', '', '' ,'' ,
                                                'IEPEntradaPed'),
                                       Unit1 = c('', '', '', '', 'cp02')))
setVNC(StQ) <- VarList
StQ
str(StQ)
=======
MicroDataDD <- data.table(Variable = 'IEPEntradaPed', 
                          Sort = 'IDDD', 
                          Class = 'numeric',
                          Qual1 = 'NOrden')
DD <- new(Class = 'DD', MicroData = MicroDataDD)
StQ <- new(Class = 'StQ', 
           Data = data.table(IDDD = character(0), Value = character(0)), 
           DD = DD)
VarList <- list(MicroData = data.table(IDQual = c('NumIdEst','','','',''),
                                       NonIDQual = c('EsMercNac', 
                                                     'EsMercEuro', 
                                                     'EsMercRM',
                                                     'Cod',
                                                     ''),
                                       IDDD = c('', '', '' ,'' ,
                                                'IEPEntradaPed'),
                                       Unit1 = c('', '', '', '', 'cp02')))
setVNC(StQ) <- VarList
StQ
str(StQ)

>>>>>>> e3ac9308c1978eb314de9c1fda9971de40a9bce7
}

