% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.Datadt.R
\docType{methods}
\name{show,Datadt-method}
\alias{show,Datadt-method}
\title{Show an object of class \linkS4class{Datadt}}
\usage{
\S4method{show}{Datadt}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Datadt}.}
}
\value{
Invisible object of class \code{\link{NULL}}.
}
\description{
The method \code{show} shows the slots of an object
\linkS4class{Datadt} limiting the number of columns on screen up to 8.

It is indeed the method \link[methods]{show} adapted to the class
\linkS4class{Datadt}.
}
\examples{
library(data.table)
# A trivial example
show(new(Class = 'Datadt'))

data(ExampleStQ)
show(getData(ExampleStQ))
                        
DDdt2 <- new(Class = 'DDdt', data.table(Variable = 'NOrden', 
                                        Sort = 'IDQual', 
                                        Class = 'character', 
                                        Qual1 = 'NOrden',
                                        Qual2 = 'Cal2',
                                        Qual3 = 'Cal4',
                                        Qual4 = 'Cal4',
                                        Qual5 = 'Cal5',
                                        Qual6 = 'Cal6',
                                        ValueRegExp = '[0-9]{9}SS'))
show(DDdt2)

}

