% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/show.DD.R
\docType{methods}
\name{show,DD-method}
\alias{show,DD-method}
\title{Show an object of class \linkS4class{DD}}
\usage{
\S4method{show}{DD}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{DD}.}
}
\value{
Invisible object of class \code{\link{NULL}}.
}
\description{
The method \code{show} shows the slots of an object
\linkS4class{DD} limiting the number of columns on screen up to 8.

It is indeed the method \link[methods]{show} adapted to the class
\linkS4class{DD}.
}
\examples{
# A trivial example
show(new(Class = 'DD'))

# A more elaborate example. See StQ-class and method getDD for details
data(ExampleQ)
show(getDD(ExampleQ))
}

