% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.DDdt.R
\docType{methods}
\name{show,DDdt-method}
\alias{show,DDdt-method}
\title{Show an object of class \linkS4class{DDdt}}
\usage{
\S4method{show}{DDdt}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{DDdt}.}
}
\value{
Invisible object of class \code{\link{NULL}}.
}
\description{
The method \code{show} shows the slots of an object
\linkS4class{DDdt} limiting the number of columns on screen up to 8.

It is indeed the method \link[methods]{show} adapted to the class
\linkS4class{DDdt}.
}
\examples{
# A trivial example
library(data.table)
show(new(Class = 'DDdt'))

DDdt1 <- new(Class = 'DDdt', data.table(Variable = 'NOrden', 
                                        Sort = 'IDQual', 
                                        Class = 'character',
                                        Length = '11',
                                        Qual1 = 'NOrden',
                                        ValueRegExp = '[0-9]{9}SS'))
show(DDdt1)
                        
DDdt2 <- new(Class = 'DDdt', data.table(Variable = 'NOrden', 
                                        Sort = 'IDQual', 
                                        Class = 'character',
                                        Length = '11',
                                        Qual1 = 'NOrden',
                                        Qual2 = 'Cal2',
                                        ValueRegExp = '[0-9]{9}SS'))
show(DDdt2)

}

