% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.rawDatadt.R
\docType{methods}
\name{show,rawDatadt-method}
\alias{show,rawDatadt-method}
\title{Show an object of class \linkS4class{rawDatadt}}
\usage{
\S4method{show}{rawDatadt}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{rawDatadt}.}
}
\value{
Invisible object of class \code{\link{NULL}}.
}
\description{
The method \code{show} shows the slot \code{Data} of an object 
\linkS4class{rawDatadt}.

It is indeed the method \link[methods]{show} adapted to the class \linkS4class{rawDatadt}.
}
\examples{
# A trivial example
show(new(Class = 'rawDatadt'))

show(getData(ExamplerawStQ))

}

