% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.rawKey.R
\docType{methods}
\name{show,rawKey-method}
\alias{show,rawKey-method}
\title{Show an object of class \linkS4class{rawKey}}
\usage{
\S4method{show}{rawKey}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{rawKey}.}
}
\value{
Invisible object of class \code{\link{NULL}}.
}
\description{
The method \code{show} displays an object \linkS4class{rawKey}.

It is indeed the method \link[methods]{show} adapted to the class \linkS4class{rawKey}.
}
\examples{
# A trivial example
show(new(Class = 'rawKey'))

show(getData(ExamplerawStQ))

}

