% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.rawStQ.R
\name{show,rawStQ-method}
\alias{show,rawStQ-method}
\title{Show an object of class \linkS4class{rawStQ}}
\usage{
\S4method{show}{rawStQ}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{rawStQ}.}
}
\value{
Invisible \code{\link{NULL}}.
}
\description{
\code{show} displays the slot \code{Data} of the input \linkS4class{rawStQ} object limiting the
number of columns on screen up to 8.
}
\details{
This method displays only the content of slot \code{Data} from the input \linkS4class{rawStQ} 
object. It is indeed the method \link[methods]{show} adapted to class \linkS4class{rawStQ}.
}
\examples{
data(ExamplerawStQ)
show(ExamplerawStQ)

}
