% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.StQ.R
\name{show,StQ-method}
\alias{show,StQ-method}
\title{Show an object of class \link{StQ}}
\usage{
\S4method{show}{StQ}(object)
}
\arguments{
\item{object}{Object of class \link{StQ}.}
}
\value{
Invisible \code{\link{NULL}}.
}
\description{
\code{show} displays the slot \code{Data} of the input \link{StQ} object limiting the
number of columns on screen up to 8.

This method displays only the content of slot \code{Data} from the input \link{StQ} 
object. It is indeed the method \link[methods]{show} adapted to class \link{StQ}.
}
\examples{
data(ExampleStQ)
show(ExampleStQ)
ExampleStQ


}
