% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.StQList.R
\name{show,StQList-method}
\alias{show,StQList-method}
\title{Show an object of class \linkS4class{StQList}}
\usage{
\S4method{show}{StQList}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{StQList}.}
}
\value{
Object of class \link{NULL}.
}
\description{
\code{show} displays the slot \code{Data} of the input \linkS4class{StQList} object
with the names of time periods from the slot \code{Periods}.

It is indeed the method \link[methods]{show} adapted to the class \linkS4class{StQList}.
}
\examples{
mm <- c(paste0('0', 1:9), 10:12)
TimePer <- paste0('MM', mm, '2015')
PeriodList <- RepoTime::newRepoTime(TimePer)
EmptyQ <- StQ()
EmptyQList <- vector('list', 12)
EmptyQList <- lapply(EmptyQList, function(x) EmptyQ)
QList <- StQList(Data = EmptyQList, Periods = PeriodList)
show(QList)
str(QList)

}
