% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.VNC.R
\name{show,VNC-method}
\alias{show,VNC-method}
\title{Show an object of class \link{VNC}}
\usage{
\S4method{show}{VNC}(object)
}
\arguments{
\item{object}{Object of class \link{VNC}.}
}
\value{
Invisible object of class \code{\link{NULL}}.
}
\description{
The method \code{show} shows the slots of an object \link{VNC}
limiting the number of columns on screen up to 8.

It is indeed the method \link[methods]{show} adapted to the class \link{VNC}.
}
\examples{
# A trivial example
show(VNC())

# A more elaborate example
library(data.table)
VarList <- list(
  ID = data.table(IDQual = c('NumIdEst', rep('', 4)),
                  NonIDQual = rep('', 5),
                  IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                  NumIdEst = c('', rep('.', 4)),
                  UnitName = c('numidest', 'nombre', 'apellidos', 'direccion', 'telefono'),
                  InFiles = rep('FI', 5)),
  MicroData = data.table(IDQual = c('NumIdEst', rep('', 2)),
                         NonIDQual = c('', 'Market', ''),
                         IDDD = c(rep('', 2), 'NewOrders'),
                         NumIdEst = c(rep('', 2), '.'),
                         Market = c(rep('', 2), '1.'),
                         UnitName = c('numidest', '', 'cp09'),
                         InFiles = rep('FF', 3)),
 ParaData = data.table(IDQual = character(0), 
                       NonIDQual = character(0), 
                       IDDD = character(0),
                       UnitName = character(0),
                       InFiles = character(0)))
VNC <- BuildVNC(VarList)
show(VNC)

}
