% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StQ-class.R
\docType{class}
\name{StQ-class}
\alias{StQ-class}
\title{S4 class for sets of \emph{St}andarized \emph{Q}uestionnaires}
\description{
Definition of an S4 class named \code{StQ} for sets of
standardized questionnaires.

The structure of the class \code{StQ} comprises 2 attribute:
\itemize{
\item The attribute \code{Data}, which is a \linkS4class{data.table} with
key-value pair structure containing all statistical variables,both from the
questionnaire and any resulting metadata from the data processing.

\item The attribute \code{DD}, which is an object of class \linkS4class{DD}.
It basically contains the definition and properties of each variable.
}

Every variable name in the attribute \code{Data} must be present in the
attribute \code{DD}.
}
\section{Slots}{

\describe{
\item{\code{Data}}{Object of class \linkS4class{data.table} with key-value pair
structure. It must have at least two columns: \code{IDDD} and \code{Value}.
It contains all statistical variables (including some metadata) together with
their corresponding values. If \code{Data} is not specified as an input
parameter, an empty \linkS4class{data.table} with columns \code{IDDD} and
\code{Value} will be initiated.}

\item{\code{DD}}{Object of class \linkS4class{DD} with the definition and properties
of all variables. If \code{DD} is not specified as an input parameter, an
empty \linkS4class{DD} object with columns \code{Variable}, \code{Sort},
\code{Class} and \code{Qual1} will be initiated.}
}}
\examples{
# An empty standardized questionnaire set:
new(Class = 'StQ')

# An empty data set:
library(data.table)
Data <- data.table(IDDD = character(0), Value = character(0))

# An object DD in two steps:
DDData <- data.table(Variable = c('NOrden', 'EsRemuner', 'Mes', 'Anno',
                                  'CCAA','CNAE2009', 'CifraNeg', 'Empleo'),
                     Sort = c('IDQual', 'NonIDQual', 'IDDD', 'IDDD',
                              'IDDD', 'IDDD', 'IDDD', 'IDDD'),
                     Class = c('character', 'integer', 'character', 'character',
                               'character', 'character', 'numeric', 'integer'),
                     Qual1 = c('', '', 'NOrden', 'NOrden',
                               'NOrden', 'NOrden', 'NOrden', 'NOrden'),
                     Qual2 = c('', '', '', '',
                               '', '', '', 'EsRemuner'))
DD <- new(Class = 'DD', MicroData = DDData)

# We create an object StQ with no data
Q <- new(Class = 'StQ', 
         Data = data.table(IDDD = character(0),
                           Value = character(0)),
         DD = DD)
Q
# Notice that only the slot Data appears on screen, but the object is not a
# data.table:
str(Q)

}

