% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StQ.R
\name{StQ}
\alias{StQ}
\title{S3 class for sets of \emph{St}andarized \emph{Q}uestionnaires}
\usage{
StQ(Data = data.table(IDDD = character(0), Value = character(0)),
  DD = BuildDD())
}
\arguments{
\item{Data}{Object of class \linkS4class{data.table}.}

\item{DD}{Object of class \link{DD}}
}
\value{
Object of class \code{\link{StQ}}.
}
\description{
Definition of an S3 class named \code{StQ} for sets of standardized questionnaires.

The structure of the class \code{StQ} comprises a \linkS4class{list} with 2 components:
\itemize{
\item the component \code{Data}, which is an object of class \linkS4class{data.table} with at
least two columns: \code{IDDD} and \code{Value}. It contains all statistical variables (including 
some metadata) together with their corresponding values. If \code{Data} is not specified as an 
input parameter, an empty \linkS4class{data.table} object with columns \code{IDDD} and 
\code{Value} will be initiated.

\item the component \code{DD}, which is an object of class \link{DD}. with the definition 
and properties of all variables. If \code{DD} is not specified as an input parameter, an empty 
\link{DD} object with columns \code{Variable}, \code{Sort}, \code{Class}, \code{Qual1} and
\code{ValueRegExp} will be initiated.
}

Every variable name in the component \code{Data} must be present in the component \code{DD}.
}
\examples{
library(data.table)
data(ExampleDD)
data(ExampleStQ)
ExampleData <- getData(ExampleStQ)
Q <- StQ(Data = ExampleData, DD = ExampleDD)
Q
# Notice that only the slot Data appears on screen, but the object is not a Datadt data.table:
str(Q)

}
