% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StQListToStQ.R
\docType{methods}
\name{StQListToStQ}
\alias{StQListToStQ}
\alias{StQListToStQ,StQList-method}
\title{Method to transform an \linkS4class{StQList} object into a list of StQs objects}
\usage{
StQListToStQ(object)

\S4method{StQListToStQ}{StQList}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{StQList} to be transformed.}
}
\value{
a list of objects of class \linkS4class{StQ}.
}
\description{
\code{StQListToStQ} transform an \linkS4class{StQList} object into a list of StQs 
objects with a new variable "Period" to take into account the interval period times related to 
\linkS4class{StQList} object.

This method creates a variable with the name \code{Period} in Data slots with the period related
to each Data and adds this variable to each DD slot of the new StQs objects.
}

