% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StQToStQList.R
\name{StQToStQList}
\alias{StQToStQList}
\alias{StQToStQList,StQ-method}
\title{Method to transform an \link{StQ} object into a \linkS4class{StQList} object}
\usage{
StQToStQList(object)

\S4method{StQToStQList}{StQ}(object)
}
\arguments{
\item{object}{Object of class \link{StQ} to be transformed.}
}
\value{
object of class \linkS4class{StQList}.
}
\description{
\code{StQToStQList} transform an \link{StQ} object (with a variable "Period"
taking into account the interval period times related to data) into the corresponding
 \linkS4class{StQList} object.

The variable \code{Period} with the period related to each data will be removed from the DD
components in the new \linkS4class{StQList} object.
}
