% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub.DD.R
\name{[.DD}
\alias{[.DD}
\title{Extract parts of an object of class \linkS4class{DD}}
\usage{
\method{[}{DD}(x, i, j, by, keyby, with = TRUE,
  nomatch = getOption("datatable.nomatch"), mult = "all", roll = FALSE,
  rollends = if (roll == "nearest") c(TRUE, TRUE) else if (roll >= 0) c(FALSE,
  TRUE) else c(TRUE, FALSE), which = FALSE, .SDcols,
  verbose = getOption("datatable.verbose"),
  allow.cartesian = getOption("datatable.allow.cartesian"), drop = NULL,
  on = NULL)
}
\arguments{
\item{x}{Object of class \linkS4class{DD}.}

\item{i, j, ...}{indices corresponding to elements to be extracted. The indices are numeric or
character vectors, \code{\link{missing}} or \code{\link{NULL}}. Numeric values are coerced to
\code{integer} with \code{\link{as.integer}} (thus truncated to zero).}

\item{drop}{Included by coherence.}
}
\value{
Object of class \linkS4class{DD} with the subsetted input object.
}
\description{
\code{[} extracts parts of an object of class \linkS4class{DD}. 

It is indeed the method \code{[} for the class \linkS4class{DD}. This method returns subsets from
an object of class \linkS4class{DD} specified as an input parameter. The output is an object of
the same class \linkS4class{DD} as the input parameter \code{x}.
}
\examples{
data(ExampleDD)
ExampleDD[Variable == 'Turnover']

}

