% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub.VNC.R
\name{[.VNC}
\alias{[.VNC}
\title{Extract parts of an object of class \linkS4class{VNC}}
\usage{
\method{[}{VNC}(x, i, j, by, keyby, with = TRUE,
  nomatch = getOption("datatable.nomatch"), mult = "all",
  roll = FALSE, rollends = if (roll == "nearest") c(TRUE, TRUE) else if
  (roll >= 0) c(FALSE, TRUE) else c(TRUE, FALSE), which = FALSE, .SDcols,
  verbose = getOption("datatable.verbose"),
  allow.cartesian = getOption("datatable.allow.cartesian"),
  drop = NULL, on = NULL)
}
\arguments{
\item{x}{Object of class \linkS4class{VNC}.}

\item{i, j, ...}{Indices corresponding to elements to be extracted. The indices are numeric or
character vectors, \code{\link{missing}} or \code{\link{NULL}}. Numeric values are coerced to
\code{integer} with \code{\link{as.integer}} (thus truncated to zero).}

\item{drop}{Included by coherence.}
}
\value{
Object of class \linkS4class{VNC} with the subsetted input object.
}
\description{
\code{[} extracts parts of an object of class \linkS4class{VNC}.

It is indeed the method \code{[} for the class \linkS4class{VNC}.
}
\examples{
library(data.table)
VNC <- VNC(ID = data.table(IDQual = c('NumIdEst', rep('', 4)),
                           NonIDQual = c('','','','',''),
                           IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                           NumIdEst = c('', rep('.', 4)),
                           UnitName = c('numidest', 'nombre', 'apellidos', 'direccion', 'telefono'),
                           InFiles = rep('FI', 5)),
           MicroData = data.table(IDQual = c('NumIdEst', rep('', 2)),
                                  NonIDQual = c('', 'Market', ''),
                                  IDDD = c(rep('', 2), 'NewOrders'),
                                  NumIdEst = c(rep('', 2), '.'),
                                  Market = c(rep('', 2), '1'),
                                  UnitName = c('numidest', '', 'cp09'),
                                  InFiles = rep('FF, FD, FG', 3)),
           ParaData = data.table(IDQual = c('NumIdEst', rep('', 2)),
                                 NonIDQual = c('', 'Action', ''),
                                 IDDD = c(rep('', 2), 'Date'),
                                 NumIdEst = c(rep('', 2), '.'),
                                 Action = c(rep('', 2), 'Imputation'),
                                 UnitName = c('numidest', '', 'FechaImput'),
                                 InFiles = rep('FP', 3)))
VNC[c('ID', 'MicroData')]


}
