% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.DDdt.R
\docType{methods}
\name{[,DDdt,ANY,ANY-method}
\alias{[,DDdt,ANY,ANY-method}
\title{Extract parts of an object of class \linkS4class{DDdt}}
\usage{
\S4method{[}{DDdt,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{Object of class \linkS4class{DDdt}.}

\item{i, j, ...}{Indices corresponding to elements to be extracted. The indices are numeric or
character vectors, \code{\link{missing}} or \code{\link{NULL}}. Numeric values are coerced to
\code{integer} with \code{\link{as.integer}} (thus truncated to zero).}

\item{drop}{Included by coherence.}
}
\value{
Object of class \linkS4class{DDdt} with the subsetted input object.
}
\description{
\code{[} extracts parts of an object of class \linkS4class{DDdt}.

It is indeed the method \code{[} for the class \linkS4class{DDdt}.
}
\examples{
data(ExampleDDdt)
ExampleDDdt[Sort == 'IDQual']

}

