% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.rawKey.R
\docType{methods}
\name{[,rawKey-method}
\alias{[,rawKey-method}
\title{Extract parts of an object of class \linkS4class{rawKey}}
\usage{
\S4method{[}{rawKey}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{Object of class \linkS4class{rawKey}.}

\item{i, j, ...}{Indices corresponding to elements to be extracted. The indices are numeric or
character vectors, \code{\link{missing}} or \code{\link{NULL}}. Numeric values are coerced to
\code{integer} with \code{\link{as.integer}} (thus truncated to zero).}

\item{drop}{Included by coherence.}
}
\value{
Object of class \linkS4class{rawKey} with the subsetted input object.
}
\description{
\code{[} extracts parts of an object of class \linkS4class{rawKey}.

It is indeed the method \code{[} for the class \linkS4class{rawKey}.
}
\examples{
data(ExamplerawKey)
ExamplerawKey[1]
ExamplerawKey[IDDDKey == 'Turnover']

}

