% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subset.rawStQList.R
\docType{methods}
\name{[,rawStQList,ANY,ANY-method}
\alias{[,rawStQList,ANY,ANY-method}
\title{Extract parts of an object of class \linkS4class{rawStQList}}
\usage{
\S4method{[}{rawStQList,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{object of class \linkS4class{rawStQList}.}

\item{i,j,...}{Indices corresponding to the elements to be extracted. These
indices are numeric vector or character vector or \link{missing} or
\link{NULL}. Numeric values are coerced internally to \code{integer} through
\code{\link{as.integer}} (and thus truncated to zero). Character vector
 correspond to names of the respective time period of each component object
of class \linkS4class{StQ}.}

\item{drop}{Included by coherence.}
}
\value{
Object of class \linkS4class{rawStQList} with the corresponding data
subset.
}
\description{
\code{[} extracts parts of an object of class
\linkS4class{rawStQList}.

It is indeed the method \code{[} for the class \linkS4class{rawStQList}. This
method returns a subset of an input object of class \linkS4class{rawStQList}.
Thus it returns an object of class \linkS4class{rawStQList}.
}
\examples{
library(RepoTime)
mm <- c(paste0('0', 1:9), 10:12)
TimePer <- paste0('MM', mm, '2015')
QList <- vector('list', 12)
QList <- lapply(QList, function(x) ExampleStQ)
names(QList) <- TimePer
QList <- new(Class = 'StQList', Data = QList, Periods = newRepoTime(TimePer))
QList[c('MM092015', 'MM102015')]
}

