% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subset.StQ.R
\docType{methods}
\name{[,StQ,ANY,ANY-method}
\alias{[,StQ,ANY,ANY-method}
\title{Extract parts of an object of class \linkS4class{StQ}}
\usage{
\S4method{[}{StQ,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{Object of class \linkS4class{StQ}.}

\item{i,j,...}{Indices corresponding to elements to be extracted. The
indices are numeric or character vectors, \code{\link{missing}} or
\code{\link{NULL}}. Numeric values are coerced to \code{integer} with
\code{\link{as.integer}} (thus truncated to zero).}

\item{drop}{Included by coherence.}
}
\value{
Object of class \linkS4class{StQ} with the subsetted input object.
}
\description{
\code{[} extracts parts of an object of class \linkS4class{StQ}.

It is indeed the method \code{[} for the class \linkS4class{StQ}. This method
returns subsets of the slot \code{Data} from an object of class
\linkS4class{StQ} specified as an input parameter. The output is an object of
 the same class \linkS4class{StQ} as the input parameter \code{x}.
}
\examples{
data(ExampleQ)
ExampleQ[IDDD == 'IASSCifraNeg']
}

