% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.StQList.R
\name{[[.StQList}
\alias{[[.StQList}
\title{Extract parts of an object of class \linkS4class{StQList}}
\usage{
\method{[[}{StQList}(x, i, j, ..., exact = TRUE)
}
\arguments{
\item{x}{object of class \linkS4class{StQList}.}

\item{i, j, ...}{Indices corresponding to the elements to be extracted. These indices are numeric
vector or character vector or \link{missing} or \link{NULL}. Numeric values are coerced
internally to \code{integer} through \code{\link{as.integer}} (and thus truncated to zero).
Character vector correspond to names of the respective time period of each component object of
class \linkS4class{StQ}.}

\item{exact}{Included by coherence.}
}
\value{
Object of class \linkS4class{StQ}.
}
\description{
\code{[[} extracts parts of an object of class \linkS4class{StQList}.

It is indeed the method \code{[[} for the class \linkS4class{StQList}. This method returns
objects of class \linkS4class{StQ} from the slot \code{Data} of the input object \code{x}.
}
\examples{
mm <- c(paste0('0', 1:9), 10:12)
TimePer <- paste0('MM', mm, '2015')
PeriodList <- RepoTime::newRepoTime(TimePer)
EmptyQ <- StQ()
EmptyQList <- vector('list', 12)
EmptyQList <- lapply(EmptyQList, function(x) EmptyQ)
QList <- StQList(Data = EmptyQList, Periods = PeriodList)
QList[['MM092015']]
QList[[2]]

}
