% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subPeriods.R
\docType{methods}
\name{subPeriods}
\alias{subPeriods}
\alias{subPeriods,StQList-method}
\title{Extract parts of an object of class \linkS4class{StQList}}
\usage{
subPeriods(x, i)

\S4method{subPeriods}{StQList}(x, i)
}
\arguments{
\item{x}{object of class \linkS4class{StQList}.}

\item{i, j, ...}{Indices corresponding to the elements to be extracted. These indices are numeric
vector or character vector or \link{missing} or \link{NULL}. Numeric values are coerced
internally to \code{integer} through \code{\link{as.integer}} (and thus truncated to zero).
Character vector correspond to names of the respective time period of each component object of
class \linkS4class{StQ}.}

\item{drop}{Included by coherence.}
}
\value{
Object of class \linkS4class{StQList} with the corresponding data
subset.
}
\description{
\code{[} extracts parts of an object of class \linkS4class{StQList}.

It is indeed the method \code{[} for the class \linkS4class{StQList}. This method returns a
subset of an input object of class \linkS4class{StQList}. Thus it returns an object of class
\linkS4class{StQList}.
}
\examples{
\dontrun{
mm <- c(paste0('0', 1:9), 10:12)
TimePer <- paste0('MM', mm, '2015')
QList <- vector('list', 12)
QList <- lapply(QList, function(x) ExampleStQ)
names(QList) <- TimePer
StQList <- StQList(Data = QList, Periods = RepoTime::newRepoTime(TimePer))
StQList[c('MM092015', 'MM102015')]
}

}

