% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Substitute.R
\docType{methods}
\name{Substitute}
\alias{Substitute}
\alias{Substitute,StQ,StQ,list,ANY-method}
\title{Substitute values of variables of some statistical units from an
object into another object}
\usage{
Substitute(In, From, VarNames, Units)

\S4method{Substitute}{StQ,StQ,list,ANY}(In, From, VarNames, Units)
}
\arguments{
\item{In}{Object which values will be sustituted into.}

\item{From}{Object which values will be taken from.}

\item{VarNames}{Character vector with the names of the variables whose values
will be substituted.}

\item{Units}{\linkS4class{data.table} with the values of the statistical unit
qualifiers whose values are to be substituted.}
}
\value{
This method returns an object with the same class as the input object
 \code{In} with the values of the input variable substituted. If the input
 parameter \code{Units} is missing, the substitution will be carried out
in every statistical unit of the object \code{From}.
}
\description{
This method implements the substitution of values of variables
specified as an input parameter of a set of statistical units also
specified as an input parameter of an input object taken the corresponding v
alues from another input object.
}
\examples{
# Creamos dos conjuntos de datos
library(data.table)
Data1 <- data.table(NOrden = c('000002896SS', '000003986SS'),
                    EsRemuner = c('', ''),
                    IDDD = c('CifraNeg', 'CifraNeg'),
                    Value = c(0, 187364))

Data2 <- data.table(NOrden = c('000002896SS', '000003986SS'),
                    EsRemuner = c('', ''),
                    IDDD = c('CifraNeg', 'CifraNeg'),
                    Value = c(10000, 188364))
                    
# Creamos el objeto VarNameCorresp
VNCList <- list(data.table(IDQual = c('Norden', '', '', '', '', '', '', ''),
                          NonIDQual = c('', 'EsRemuner', '', '', '', '', '',
                           ''), IDDD = c('', '', 'Mes', 'Anno', 'CCAA',
                          'CNAE2009', 'CifraNeg', 'Empleo'), Norden = c('',
                          '', '', '', '', '', '', ''), EsRemuner = c('',
                          '', '', '', '', '', '', '1'), Unit1 = c('', '', '',
                          '', '', '', '', '')))
VNC <- new(Class = 'VarNameCorresp', VarNameCorresp = VNCList)

# Creamos un slot DD:
DDData <- data.table(Variable = c('NOrden',
                                  'EsRemuner',
                                  'Mes',
                                  'Anno',
                                  'CCAA',
                                  'CNAE2009',
                                  'CifraNeg',
                                  'Empleo'),
                     Sort = c('IDQual',
                              'NonIDQual',
                              'IDDD',
                              'IDDD',
                              'IDDD',
                              'IDDD',
                              'IDDD',
                              'IDDD'),
                     Class = c('character',
                               'integer',
                               'character',
                               'character',
                               'character',
                               'character',
                               'numeric',
                               'integer'),
                     Qual1 = c('', '', 'NOrden', 'NOrden', 'NOrden', 
                               'NOrden', 'NOrden', 'NOrden'),
                     Qual2 = c('', '', '', '', '', '', '', 'EsRemuner'))
DD <- new(Class = 'DD', Data = DDData, VarNameCorresp = VNC)

# Finalmente creamos el objeto
Q1 <- new(Class = 'StQ', Data = Data1, DD = DD)
Q2 <- new(Class = 'StQ', Data = Data2, DD = DD)
Substitute(In = Q1, From = Q2, 'CifraNeg', data.table(NOrden = '000002896SS'))

}

