% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variable-class.R
\docType{class}
\name{Variable-class}
\alias{Variable-class}
\title{S4 class with the main attributes of a variable}
\description{
Definition of an S4 class named \code{Variable} with its main attributes
}
\section{Slots}{

\describe{
\item{\code{UnitName}}{\code{character} vector of lenght 1 with the UnitName of the new variable to 
create.}

\item{\code{IDDD}}{\code{character} vector of length 1 with the IDDD name of the new variable to create.}

\item{\code{QualsValues}}{Named list with the names of the qualifiers of the new variable and their
respective values.}

\item{\code{Length}}{\code{character} vector of length 1 with the maximum length for the values of the
variable.}

\item{\code{ClassVar}}{\code{character} vector of length 1 with the class of the variable (numeric o 
character)}

\item{\code{ValueRegExp}}{\code{character} vector of length 1 with the expression of right values for the
variable.}

\item{\code{Formula}}{with the character expression to calculate the new variable.}

\item{\code{SlotName}}{\code{character} vector of length 1 with the name of the DD slot in which the
new variable will be created.}

\item{\code{Literal}}{\code{character} vector of length 1 with a comment for the variable.}
}}
\examples{
Var <- new(Class = 'Variable', UnitName = 'LRTEmp',
                               IDDD = 'Employees',
                               QualsValues = list(ID = '', EmplType = ''),
                               Length = '8',
                               ClassVar = 'numeric',
                               ValueRegExp = '[.]+',
                               Formula = as.call(list('log( 1 + (Turnover / (Employees_1. + Employees_2.1.)))')),
                               SlotName = 'MicroData',
                               Literal = '')

}

