% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableIDDDNames.R
\name{VariableIDDDNames}
\alias{VariableIDDDNames}
\alias{VariableIDDDNames,Variable,DD-method}
\title{Convert production unit names of an object of class \linkS4class{Variable} into their
corresponding statistical variable names (IDDD)}
\usage{
VariableIDDDNames(Variable, Correspondence)

\S4method{VariableIDDDNames}{Variable,DD}(Variable, Correspondence)
}
\arguments{
\item{Variable}{Object with its main attributes}

\item{Correspondence}{Object in which Variable will be included.}
}
\value{
Returns a \code{character} vector with the corresponding IDDD variable name.
}
\description{
\code{VariableIDDDNames} returns a \linkS4class{character} vector with the
statistical variable name (IDDD + Qualifiers) corresponding to the production unit variable name
of the \linkS4class{Variable} object specified as input argument.
}
\details{
IDDD and qualifiers compose together the so-called IDDDname of the variable by pasting
the IDDD identifier and each consecutive qualifier with an underscore _.
}
\examples{
data(ExampleStQ)
DD <- getDD(ExampleStQ)
Var <- new(Class = 'Variable', UnitName = 'LRTEmp',
                               IDDD = 'Employees',
                               QualsValues = list(ID = '', EmplType = ''),
                               Length = '8',
                               ClassVar = 'numeric',
                               ValueRegExp = '[.]+',
                               Formula = as.call(list('log( 1 + (Turnover / (Employees_1. + Employees_2.1.)))')),
                               SlotName = 'MicroData',
                               Literal = '',
                               DDversion = '1')
IDDDName <- VariableIDDDNames(Var, DD)

}
