% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VarNamesToDD.R
\name{VarNamesToDD}
\alias{VarNamesToDD}
\title{Return a reduced \linkS4class{DD} involving the input variable names}
\usage{
VarNamesToDD(VarNames, DD)
}
\arguments{
\item{VarNames}{Character vector with the compound variable names.}

\item{DD}{Object of class \linkS4class{DD} with the definition and properties of the variables.}
}
\value{
\linkS4class{DD} involving only the compound variable names specified in the input
parameter \code{VarNames}.
}
\description{
\code{VarNamesToDD} returns a \linkS4class{DD} object including only the variable
names specified in the input parameter \code{VarNames}.
}
\examples{
data(ExampleDD)

VarNames <- c('Turnover')
VarNamesToDD(VarNames, ExampleDD)

VarNames <- c('NOrden', 'Turnover', 'EsRemuner', 'Employees')
VarNamesToDD(VarNames, ExampleDD)
}

