% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarNamesToFormula.R
\docType{methods}
\name{VarNamesToFormula}
\alias{VarNamesToFormula}
\alias{VarNamesToFormula,character,DD-method}
\title{Return a list of formulas for dcasting long data sets.}
\usage{
VarNamesToFormula(VarNames, DD)

\S4method{VarNamesToFormula}{character,DD}(VarNames, DD)
}
\arguments{
\item{VarNames}{Character vector with compound variable names.}

\item{DD}{Object of class \linkS4class{DD} with the definition and properties of the variables.}
}
\value{
\linkS4class{data.table} with columns \code{Variable} and \code{Form}. The column
\code{Variable} indicates the root name of each input variable name and the column \code{Form}
presents the corresponding formula.
}
\description{
\code{VarNamesToFormula} returns a list of formulas (in the sense of functions
\code{\link[data.table]{dcast}} and \code{\link[data.table]{melt}} from the package
\link{data.table}) for each compound variable name included in the input vector \code{VarNames}.

This function returns a \linkS4class{data.table} with two colums: the column \code{Variable} with
the root names of each compound variable name, and the column \code{Form} with the corresponding
formula having IDQual qualifiers and the sort of variable (\code{IDDD}, \code{IDQual} or
\code{NonIDQual}) in its lhs and the variable qualifiers whose values appear as suffixes in the
compound variable names in the rhs. Duplicate entries in the resulting \linkS4class{data.table}
drop out.

\code{VarNamesToFormula} is designed to be applied on compound variable names including suffixes
with qualifier values. It is fundamentally designed for the internal use in the construction of
editing strategies, but it can also be of utility in some scripts.
}
\examples{
data(ExampleDD)
VarNamesToFormula('Turnover', ExampleDD)

VarNamesToFormula(c('Employees_1.', 'Employees_2.2', 'Employees'), ExampleDD)

VarNamesToFormula(c('Turnover', 'Employees_1.', 'Employees_2.2', 'Employees'), ExampleDD)

}
