% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VNC.R
\name{VNC}
\alias{VNC}
\title{S3 class object constructor for the correspondence between variable names}
\usage{
VNC(ID = data.table(IDQual = character(0), NonIDQual = character(0), IDDD
  = character(0), UnitName = character(0), InFiles = character(0)),
  MicroData = data.table(IDQual = character(0), NonIDQual = character(0),
  IDDD = character(0), UnitName = character(0), InFiles = character(0)),
  ...)
}
\description{
Definition of an S3 class named \link{VarNameCorresp} with the correspondence 
between variable names used by the different production unit for a given statistical variable and
variable names according to the key-value pair data model with qualifiers given by statistical 
metadata.

The class \link{VarNameCorresp} comprises a \link{list} whose components are 
\linkS4class{data.table}s with a row per each variable and the following columns:

\itemize{
 \item \code{IDQual}: Names of unit qualifiers.
 \item \code{NonIDQual}: Names of variable names qualifiers.
 \item \code{IDDD}: Names of the variables.
 \item \emph{One column per each name of the \code{IDQual} qualifiers}: 
 Values of the corresponding \code{IDQual} Qualifier for each variable (row).
 \item \emph{One column per each name of the \code{NonIDQual} qualifiers}: 
 Values of the corresponding \code{NonIDQual} Qualifier for each variable 
 (row).
 \item \code{UnitName}: Variable name used in production.
 \item \code{InFiles}: code of the files where each variable is included.
}
}
\examples{
library(data.table)
VNC <- VNC(ID = data.table(IDQual = c('NumIdEst', rep('', 4)),
                           NonIDQual = c('','','','',''),
                           IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                           NumIdEst = c('', rep('.', 4)),
                           UnitName = c('numidest', 'nombre', 'apellidos', 'direccion', 'telefono'),
                           InFiles = rep('FI', 5)),
           MicroData = data.table(IDQual = c('NumIdEst', rep('', 2)),
                                  NonIDQual = c('', 'Market', ''),
                                  IDDD = c(rep('', 2), 'NewOrders'),
                                  NumIdEst = c(rep('', 2), '.'),
                                  Market = c(rep('', 2), '1'),
                                  UnitName = c('numidest', '', 'cp09'),
                                  InFiles = rep('FF, FD, FG', 3)),
           ParaData = data.table(IDQual = c('NumIdEst', rep('', 2)),
                                 NonIDQual = c('', 'Action', ''),
                                 IDDD = c(rep('', 2), 'Date'),
                                 NumIdEst = c(rep('', 2), '.'),
                                 Action = c(rep('', 2), 'Imputation'),
                                 UnitName = c('numidest', '', 'FechaImput'),
                                 InFiles = rep('FP', 3)))

}
