% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WriteRepoFile.R
\docType{methods}
\name{WriteRepoFile}
\alias{WriteRepoFile}
\alias{WriteRepoFile,StQ-method}
\alias{WriteRepoFile,data.table-method}
\title{Escritura de un fichero con estructura par clave-valor}
\usage{
WriteRepoFile(object, Name)

\S4method{WriteRepoFile}{data.table}(object, Name)

\S4method{WriteRepoFile}{StQ}(object, Name)
}
\arguments{
\item{object}{Objeto que se desea escribir con estructura par clave-valor en
un fichero ASCII.}

\item{Name}{\code{\link{vector}} de tipo \code{character} de longitud 1 con el
nombre del fichero ASCII a escribir. Debe incluir la ruta completa en la que
se quiere ubicar el mismo (directorio y nombre del fichero que se va a crear).}

\item{NIV}{\code{Vector} de tipo \code{integer} de longitud uno con el número
de niveles del fichero que se va a generar. Corresponde al número total de
calificadores que aparecen en el fichero.Por defecto toma el valor
\code{NIV} = \code{1}.}
}
\value{
Fichero ASCII con los datos del objeto de entrada con estructura par
clave-valor y de nombre el especificado como input.
}
\description{
\code{WriteRepoFile} reescribe el objeto de entrada con una estructura par
clave-valor y lo guarda en un fichero ASCII.
}
\details{
Esta función lee el objeto de entrada \code{object} y reescribe su contenido
con una estructura par clave-valor. Esta nueva estructura se graba en un
fichero ASCII con nombre \code{Name} especificado como argumento de entrada,
el cual debe definir la ruta completa de acceso al fichero (directorio y nombre
que queremos asignarle).

También es necesario incluir como input de la función el número de
calificadores de identificación (\code{NIV}) para que se incluya en la
primera línea del fichero ASCII generado.
}
\examples{
# Se escribirá un fichero ASCII en el escritorio del administrador:
Name <- 'C:/Users/Administrador/Desktop/E30103.FF_V1.MM032014.D_1' # Cambiar a discreción
WriteRepoFile(Example.Q, Name)
}
\seealso{
\link{ReadSASFile}, \link{ReadRepoFile}, \link{FirstLine}
}

