% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildrawStQList.R
\name{BuildrawStQList}
\alias{BuildrawStQList}
\title{Constructor of objects of class \linkS4class{rawStQList}}
\usage{
BuildrawStQList(Data)
}
\arguments{
\item{Data}{A named \code{list} of objects of class \linkS4class{rawStQ}.}
}
\value{
An object of class \linkS4class{rawStQList}.
}
\description{
This constructor returns an object of class \linkS4class{rawStQList}.
The input parameter is a named \code{list} of objects of class \linkS4class{rawStQ}. Notice that 
the names of the \code{list} must be valid time period names according to package 
\link[=RepoTime]{RepoTime}.
}
\examples{
mm <- c(paste0('0', 1:9), 10:12)
TimePer <- paste0('MM', mm, '2015')
EmptyrawQ <- rawStQ()
EmptyrawQList <- vector('list', 12)
EmptyrawQList <- lapply(EmptyrawQList, function(x) EmptyrawQ)
names(EmptyrawQList) <- TimePer
rawQList <- BuildrawStQList(EmptyrawQList)
rawQList
#Notice that it is indeed an object with complex structure:
str(rawQList)

}
