% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildStQList.R
\name{BuildStQList}
\alias{BuildStQList}
\title{Constructor of objects of class \linkS4class{StQList}}
\usage{
BuildStQList(Data)
}
\arguments{
\item{Data}{\code{List} of objects of class \linkS4class{StQ}.}
}
\value{
An object of class \linkS4class{StQList}.
}
\description{
This constructor returns an object of class \linkS4class{StQList}.
The input parameter is a named \code{list} of objects of class \linkS4class{StQ}. Notice that 
the names of the \code{list} must be valid time period names according to package 
\link[=RepoTime]{RepoTime}.
}
\examples{
EmptyQ <- StQ()
EmptyQList <- vector('list', 12)
EmptyQList <- lapply(EmptyQList, function(x) EmptyQ)
mm <- c(paste0('0', 1:9), 10:12)
names(EmptyQList) <- paste0('MM', mm, '2015')
QList <- BuildStQList(EmptyQList)
QList
#Notice that it is indeed an object with complex structure:
str(QList)

}
