% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildVCN.R
\name{BuildVNC}
\alias{BuildVNC}
\title{Constructor of objects of class \linkS4class{VarNameCorresp}}
\usage{
BuildVNC(Data)
}
\arguments{
\item{Data}{\code{List} of objects of class \linkS4class{VNCdt}.}
}
\value{
An object of class \linkS4class{VarNameCorresp}.
}
\description{
This constructor returns objects of class \linkS4class{VarNameCorresp}.
THe input parameter is a \code{list} of objects of class \linkS4class{VNCdt}.
}
\examples{
library(data.table)
VarList <- list(
  ID = new(Class = 'VNCdt', 
           .Data = data.table(
                 IDQual = c('NumIdEst', rep('', 4)),
                 NonIDQual = rep('', 5),
                 IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                 NumIdEst = c('', rep('.', 4)),
                 Unit1 = c('numidest', 'nombre', 'apellidos', 'direccion', 'telefono'))),
  MicroData = new(Class = 'VNCdt', 
                  .Data = data.table(
                      IDQual = c('NumIdEst', rep('', 4)),
                      NonIDQual = c('', 'IsNatMarket', 'IsEuroMarket', 'IsRWMarket', ''),
                      IDDD = c(rep('', 4), 'NewOrders'),
                      NumIdEst = c(rep('', 4), '.'),
                      IsNatMarket = c(rep('', 4), '0'),
                      IsEuroMarket = c(rep('', 4), '0'),
                      IsRWMarket = c(rep('', 4), '1'),
                      Unit1 = c('numidest', rep('', 3), 'cp09'))),
 Aggregates = new(Class = 'VNCdt', 
                  .Data = data.table(
                     IDQual = c('Province', 'NACE', 'IsNatMarket', ''),
                     NonIDQual = rep('', 4),
                     IDDD = c('', '', '', 'TotalTurnover'),
                     Province = c('', '', '', '.'),
                     NACE = c('', '', '', '.'),
                     IsNatMarket = c('', '', '', '1'),
                     Unit1 = c('provincia', 'actividad', '', 'cn01'))))
new(Class = 'VarNameCorresp', .Data = VarList)

VNC <- BuildVNC(VarList)
VNC
#Notice that it is indeed an object with complex structure:
str(VNC)

}

