% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatadtToDT.R
\name{DatadtToDT}
\alias{DatadtToDT}
\title{Returns a \linkS4class{data.table} with the data of the input Datadt object.}
\usage{
DatadtToDT(object)
}
\arguments{
\item{object}{Datadt which is desired to transform into a data.table.}
}
\value{
\linkS4class{data.table}
}
\description{
\code{DatadtToDT} returns a \linkS4class{data.table} with the data and values of the
 input \linkS4class{Datadt} object.
}
\examples{
library(data.table)
Ddt <- new(Class = 'Datadt', 
            data.table(ID = c('001', '001', '001', '001'), 
                       IsNatMarket = c('0', '1', '', ''),
                       EsRemuner = c('1','0','',''),
                       IDDD = c('Turnover', 'Turnover', 'Province', 'NACE09'),
                       Value = c('625000', '23154', '04', '0512')))
                        
                        
DatadtToDT(Ddt)                        

}

