% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcast_StQ.R, R/dcast_Datadt.R
\docType{methods}
\name{dcast_StQ}
\alias{dcast_Datadt,Datadt-method}
\alias{dcast_StQ}
\alias{dcast_StQ,StQ-method}
\title{Convert an \linkS4class{StQ} object into a dcasted data.table.}
\usage{
dcast_StQ(object, VarNames = NULL, DDslot = "MicroData")

\S4method{dcast_StQ}{StQ}(object, VarNames = NULL, DDslot = "MicroData")

\S4method{dcast_Datadt}{Datadt}(object, VarNames)
}
\arguments{
\item{object}{Object of class \linkS4class{StQ} whose slot \code{Data} will
be converted.}

\item{VarNames}{Character vector with names of the output variables.}

\item{DDslot}{Character vector of length 1 with the name of DD slot used
to make the transformation of the input object. Its default value is
\code{MicroData}.}
}
\value{
\linkS4class{data.table} with data from slot \code{Data} of the input
\linkS4class{StQ} object with statistical units by rows and variables by
 columns. Only variables in \code{VarNames} will be output. If no variable
 name is specified, all variables in the input object will be output.
}
\description{
\code{dcast_StQ} returns a \linkS4class{data.table} in dcasted
form (observations by row and variables by columns) with data from the input
\linkS4class{StQ} object.

This methods converts the slot \code{Data} from the input \code{StQ} object
 into a \linkS4class{data.table} with statistical units by row and variables 
 specified in the input parameter \code{VarNames} by columns.

To distinguish between variables and qualifiers this function makes use of
the slot \code{DD} of input \linkS4class{StQ} variable.

This method is indeed a wrapper for the function
\code{\link[data.table]{dcast.data.table}} of the package
\linkS4class{data.table}, adapted to the structure of objecto
\linkS4class{StQ}.
}
\examples{
data(ExampleQ)
Mat <- dcast_StQ(ExampleQ, c('Turnover'))
str(Mat)

}
\seealso{
\code{\link{melt_StQ}}, \code{\link[data.table]{dcast.data.table}},
\code{\link[data.table]{melt.data.table}}, \code{\link[reshape2]{melt}},
\code{\link[reshape2]{dcast}}
}

