% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD-class.R
\docType{class}
\name{DD-class}
\alias{DD-class}
\title{S4 class for the information contained in DD files}
\description{
Definition of an S4 class named \code{DD} with the information 
contained in a DD file with s slightly different structure to that of those
files. 

The class \code{DD} comprises a slot of class \linkS4class{data.table} with
at least four columns named \code{Variable}, \code{Sort}, \code{Class} and
\code{Qual1}. These columns have the same meanings: 

\itemize{
 \item \code{Variable}: Name of the variable. 
 \item \code{Sort}: Semantic sort of the variable, which can be a statistical
 unit qualifier (\code{IDQual}), a variable (non-unit) qualifier 
 (\code{NonIDQual}) and a variable name (\code{IDDD}).
 \item \code{Class}: Class of the variable (\code{integer}, \code{numeric},
 \code{character},...).
 \item \code{Qual1}: Name of the variable qualifier 1. 
}

This \linkS4class{data.table} is completed with as many columns named 
\code{Qualn} as necessary.
}
\section{Slots}{

\describe{
\item{\code{VarNameCorresp}}{Object of class \linkS4class{VarNameCorresp}.}

\item{\code{MicroData}}{\linkS4class{data.table} with at least the four columns 
\code{Variable}, \code{Sort}, \code{Class}, \code{Qual1} (in that order).}

\item{\code{Aggregates}}{\linkS4class{data.table} with at least the four columns 
\code{Variable}, \code{Sort}, \code{Class}, \code{Qual1} (in that order).}

\item{\code{AggWeights}}{\linkS4class{data.table} with at least the four columns 
\code{Variable}, \code{Sort}, \code{Class}, \code{Qual1} (in that order).}

\item{\code{Other}}{\linkS4class{data.table} with at least the four columns 
\code{Variable}, \code{Sort}, \code{Class}, \code{Qual1} (in that order).}
}}
\examples{
# An empty DD object is built through the code: 
new(Class = 'DD')

# An example with data extracted of a DD object:
library(data.table)
data(ExampleDD)
VNC <- ExampleDD@VarNameCorresp
DDData <- ExampleDD@MicroData
DDAggregates <- ExampleDD@Aggregates
new(Class = 'DD', VarNameCorresp = VNC, MicroData = DDData, Aggregates = DDAggregates)

}

