% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD-class.R
\docType{class}
\name{DD-class}
\alias{DD-class}
\title{S4 class with the dictionary of data (variable specifications)}
\description{
Definition of an S4 class named \code{DD} with the specification
of each variable.

The class \code{DD} comprises a slot of class \linkS4class{VarNameCorresp}
and slots of class \linkS4class{DDdt} of names \code{ID},
\code{MicroData}, \code{Aggregates}, \code{AggWeights}, \code{Other}
}
\examples{
# An empty DD object is built through the code:
new(Class = 'DD')

# An example:
library(data.table)
### We build the VNC object
VarList <- list(ID = new(Class = 'VNCdt',
                data.table(IDQual = c('NumIdEst', rep('', 4)),
                           NonIDQual = c('','','','',''),
                           IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                           NumIdEst = c('', rep('.', 4)),
                           UnitName = c('numidest', 'nombre', 'apellidos', 'direccion', 'telefono'))
 ),
MicroData =new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst', rep('', 2)),
                                           NonIDQual = c('', 'Market', ''),
                                           IDDD = c(rep('', 2), 'NewOrders'),
                                           NumIdEst = c(rep('', 2), '.'),
                                           Market = c(rep('', 2), '1'),
                                           UnitName = c('numidest', '', 'cp09'))),
ParaData = new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst', rep('', 2)),
                                           NonIDQual = c('', 'Action', ''),
                                           IDDD = c(rep('', 2), 'Date'),
                                           NumIdEst = c(rep('', 2), '.'),
                                           Action = c(rep('', 2), 'Imputation'),
                                           UnitName = c('numidest', '', 'FechaImput'))))

VNC <- new(Class = 'VarNameCorresp', VarList)

IDdt <- new( Class='DDdt',data.table(
    Variable = c('NumIdEst', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
    Sort = c('IDQual', rep('IDDD', 4)),
    Class = rep('character', 5),
    Length = c('11', '15', '15', '20','9'),
    Qual1 = c('', rep('NumIdEst', 4)),
    ValueRegExp = c('[0-9]{9}PP', '.+', '.+', '.+', '(6|9)[0-9]{8}')))
Microdt <- new( Class='DDdt',data.table(
    Variable = c('NumIdEst', 'Market', 'NewOrders'),
    Sort = c('IDQual', 'NonIDQual', 'IDDD'),
    Class = c(rep('character', 2), 'numeric'),
    Length = c('11', '2', '7'),
    Qual1 = c(rep('', 2), 'NumIdEst'),
    ValueRegExp = c('[0-9]{9}PP', '.+', '([0-9]{1, 10}| )')))
Paradt <-new( Class='DDdt', data.table(
    Variable = c('NumIdEst', 'Action', 'Date'),
    Sort = c('IDQual', 'NonIDQual', 'IDDD'),
    Class = rep('character', 3),
    Length = c('11', '10', '10'),
    Qual1 = c(rep('', 2), 'NumIdEst'),
    Qual2 = c(rep('', 2), 'Action'),
    ValueRegExp = c('[0-9]{9}PP', 'Collection|Editing|Imputation',
                    '(([0-9]{2}-(0[1-9]|1(0-2))-[0-9]{4})| )')))

DD <- new(Class = 'DD',
          VarNameCorresp = VNC,
          ID = IDdt,
          MicroData = Microdt,
          ParaData = Paradt)

}

