% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDdtToVNC.R
\name{DDdtToVNC}
\alias{DDdtToVNC}
\title{Returns the corresponding object of class \linkS4class{VarNameCorresp} according to the
 DDdt input object.}
\usage{
DDdtToVNC(DDdt, NameVNC, InFiles)
}
\arguments{
\item{DDdt}{Object of class \linkS4class{DDdt}.}

\item{NameVNC}{character vector with the name of the element in VNC which is being build.}

\item{InFiles}{character vector with as many components as the rows of \code{DDdt} specifying the 
files where to include each variable.}
}
\value{
the corresponding object of class \linkS4class{VarNameCorresp}.
}
\description{
\code{DDdtToVNC} returns the corresponding object of class
\linkS4class{VarNameCorresp} according to the DDdt input object.
}
\examples{
library(data.table)
AggWeights <- data.table(Variable = c('ID', 'Pond1'), Sort = c('IDQual', 'IDDD'),
                         Class = c('character', 'character'),
                         Length = c('11', '7'),
                         Qual1 = c('', 'ID'), ValueRegExp = c('', ''))
AggWeights <- new(Class = 'DDdt', AggWeights)

VNC <- DDdtToVNC(AggWeights, 'AggWeights', rep('FA', 2))

}

