% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDslotWith.R
\docType{methods}
\name{DDslotWith}
\alias{DDslotWith}
\alias{DDslotWith,DD,character-method}
\title{Return the slot of the input \linkS4class{DD} object containing the input variable names}
\usage{
DDslotWith(object, VarName, DDslot = "MicroData")

\S4method{DDslotWith}{DD,character}(object, VarName, DDslot = "MicroData")
}
\arguments{
\item{object}{Object of class \linkS4class{DD}.}

\item{VarName}{Character vector with the name of the variable.}

\item{DDslot}{Character vector of length 1 with the name of DD slot in which
variables in VarName are defined. Its default value is \code{MicroData}.}
}
\value{
Matrix with the queried values.
}
\description{
\code{DDslotWith} returns the slot of the input \linkS4class{DD} object which 
 contains the variable name specified as input parameter \code{VarName}. If the variable is 
 present in more than one slot, the input parameter \code{DDslot} is used to choose the correct
 slot; otherwise an error is triggered.
}
\examples{
data(ExampleDD)
DDslotWith(ExampleDD, 'Turnover')

}

