<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTToKey.R
\docType{methods}
\name{DTToKey}
\alias{DTToKey}
\alias{DTToKey,data.table-method}
\title{Return a character vector of keys out of a \linkS4class{data.table}.}
\usage{
DTToKey(DT)

\S4method{DTToKey}{data.table}(DT)
}
\arguments{
\item{DT}{\linkS4class{data.table} with the key components.}
}
\value{
Object of class \linkS4class{rawKey} with the keys of the input
}
\description{
\code{DTToKey} returns a character vector of keys from the input
\linkS4class{data.table} whose columns are the different qualifiers with as 
many rows as components (length) of this input key.
}
\examples{
key <- new(Class = 'rawKey', c('IDDD:IASSEmpleo_Norden:391092SS_EsRemuner:1_TipoRemuner:1',
                               'IDDD:IASSCifraNeg_Norden:asd2SS',
                               'IDDD:IASSLPCifraNeg_Norden:1231_CCAA:01'))
keyDT <- DTToKey(KeyToDT(key))

}

||||||| merged common ancestors
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTToKey.R
\docType{methods}
\name{DTToKey}
\alias{DTToKey}
\alias{DTToKey,data.table-method}
\title{Return a character vector of keys out of a \linkS4class{data.table}}
\usage{
DTToKey(DT)

\S4method{DTToKey}{data.table}(DT)
}
\arguments{
\item{DT}{\linkS4class{data.table} with the key components.}
}
\value{
\linkS4class{data.table} with one column per each different qualifier
found in the input key and as many rows as components (length) of this input
key.
}
\description{
\code{DTToKey} returns a character vector of keys from the input
\linkS4class{data.table} whose columns are the different qualifiers with as 
many rows as components (length) of this input key.
}
\examples{
key <- new(Class = 'rawKey', c('ID:391092SS_Turnover:9834.3_Province:09',
                               'ID:asd2SS',
                               'ID:1231_Employees:841_NACE:0502'))
DTToKey(KeyToDT(key))

}

>>>>>>> e3ac9308c1978eb314de9c1fda9971de40a9bce7
