% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDatadt.R
\docType{data}
\name{ExampleDatadt}
\alias{ExampleDatadt}
\title{Object of class \linkS4class{Datadt}}
\format{Object of class \linkS4class{Datadt} with 354458 rows and 4 columns.}
\usage{
data(ExampleDatadt)
}
\description{
\code{ExampleDatadt} is an object of class \linkS4class{Datadt} with key-value pair
structure. Data are severely perturbed anonymised data from a real statistical operation. The 
class \linkS4class{Datadt} is a subclass of \linkS4class{data.table}.

The columns of this object are:
\itemize{
  \item ID: First qualifier (statistical unit qualifier).
  \item IsRemun: Second qualifier (variable name qualifier).
  \item IsPartTime: Third qualifier (variable name qualifier).
  \item IDDD: Variable name.
  \item Value: Variable value.
}
}

