<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> e9efa7c0eccfd9de1f1e25ef8428b085f47e8788
% Please edit documentation in R/ExampleDD.R
\docType{data}
\name{ExampleDD}
\alias{ExampleDD}
<<<<<<< HEAD
\title{Object of class \linkS4class{DD}}
\format{Object of class \linkS4class{DD}. Its slot \code{Data} has 52 rows
and 6 columns.}
=======
\title{Objeto de clase DD.}
\format{Objeto de clase \code{DD}. Su slot \code{Data} tiene 52 filas y 6
columnas.}
>>>>>>> e9efa7c0eccfd9de1f1e25ef8428b085f47e8788
\usage{
data(ExampleDD)
}
\description{
<<<<<<< HEAD
\code{ExampleDD} is an object of class \linkS4class{DD}.

The columns of the slot \code{Data} are:
\itemize{
  \item Variable: statistical variable name.
  \item Sort: semantic sort of the variable (IDQual, NonIDQual, IDDD).
  \item Class: class of the variable (\code{numeric}, \code{character},...).
  \item Qual1: first variable qualifier.
  \item Qual2: second variable qualifier.
  \item Qual3: third variable qualifier.
=======
\code{ExampleDD} es un objeto de clase \linkS4class{DD}.
}
\details{
Las columnas del slot Data del objeto son las siguientes:
\itemize{
  \item Variable. Identificador de variable estadística.
  \item Sort. Rol de la variable (IDQual, NonIDQual, IDDD).
  \item Class. Clase en el lenguaje R de la variable.
  \item Qual1. Primer calificador de la variable.
  \item Qual2. Segundo calificador de la variable.
  \item Qual3. Tercer calificador de la variable.
>>>>>>> e9efa7c0eccfd9de1f1e25ef8428b085f47e8788
}
}

