% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDM.R
\docType{data}
\name{ExampleDM}
\alias{ExampleDM}
\title{\linkS4class{data.table} with a rectangular data set}
\format{\code{data.table} with 29499 rows and 59 columns.}
\usage{
data(ExampleDM)
}
\description{
\code{ExampleDM} is a \linkS4class{data.table} in data matrix
structure with anonymised data from an undisclosed period of the Sector
Service Indicators Survey.
}

