% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDM.R
\docType{data}
\name{ExampleDM}
\alias{ExampleDM}
\title{\linkS4class{data.table} with a rectangular data set (data matrix)}
\format{\code{data.table} with 29538 rows and 13 columns.}
\usage{
data(ExampleDM)
}
\description{
\code{ExampleDM} is a \linkS4class{data.table} with data matrix structure and
severely perturbed data by different methods from a real data set.
}

