<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> e9efa7c0eccfd9de1f1e25ef8428b085f47e8788
% Please edit documentation in R/ExampleQ.R
\docType{data}
\name{ExampleQ}
\alias{ExampleQ}
<<<<<<< HEAD
\title{Object of class \linkS4class{StQ}}
\format{Object of class \code{StQ}. Its slot \code{Data} has 1710942 rows and
 6 columns.}
=======
\title{Objeto de clase StQ.}
\format{Objeto de clase \code{StQ}. Su slot \code{Data} tiene 1710942 filas
 y 6 columnas.}
>>>>>>> e9efa7c0eccfd9de1f1e25ef8428b085f47e8788
\usage{
data(ExampleQ)
}
\description{
<<<<<<< HEAD
\code{ExampleQ} is an object of class \linkS4class{StQ} whose
slot \code{Data} is a \link{data.table} with key-value pair structure with
anonymised data of an undisclosed time period.

The columns of slot \code{Data} of this object is:
\itemize{
  \item NOrden: First qualifier (statistical unit qualifier).
  \item CCAA: Second qualifier (variable name qualifier).
  \item EsRemuner: Third qualifier (variable name qualifier).
  \item TipoRem: Fourth qualifier (variable name qualifier).
  \item IDDD: Variable names.
  \item Value: Variable values (default class: \code{character}).
=======
\code{ExampleQ} es un objeto de clase \linkS4class{StQ} cuyo slot
\code{Data} es una \code{\link{data.table}} con estructura par clave-valor
con los datos de la encuesta de IASS correspondientes a un periodo no
especificado (el mismo periodo utilizado para el objeto \code{ExampleDM}).
}
\details{
Las columnas del slot Data del objeto son las siguientes:
\itemize{
  \item NOrden. Primer calificador (identificador de unidad).
  \item CCAA. Segundo calificador (de variable).
  \item EsRemuner. Tercer calificador (de variable).
  \item TipoRem. Cuarto calificador (de variable).
  \item IDDD. Nombres de las variable.
  \item Valor. Valores de las variable (por defecto, de clase \code{character}).
>>>>>>> e9efa7c0eccfd9de1f1e25ef8428b085f47e8788
}
}

