% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAggWeights.R, R/getOtherDD.R
\docType{methods}
\name{getAggWeights}
\alias{getAggWeights}
\alias{getAggWeights,DD-method}
\alias{getOtherDD,DD-method}
\title{Return slot AggWeights from a DD object}
\usage{
getAggWeights(object)

\S4method{getAggWeights}{DD}(object)

\S4method{getOtherDD}{DD}(object)
}
\arguments{
\item{object}{a DD Object whose slot \code{AggWeights} is queried.}
}
\value{
\linkS4class{data.table} with data from slot \code{AggWeights} of the
input \linkS4class{DD} object.
}
\description{
\code{getAggWeights} returns slot \code{AggWeights} of the input
\linkS4class{DD} object.
}
\examples{
# An example:
library(data.table)
### We build the VNC object
VarList <- list(ID = new(Class = 'VNCdt',
                data.table(IDQual = c('NumIdEst', rep('', 4)),
                           NonIDQual = c('','','','',''),
                           IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                           NumIdEst = c('', rep('.', 4)),
                           UnitName = c('numidest', 'nombre', 'apellidos', 'direccion', 'telefono'),
                           InFiles = rep('FI', 5))),
MicroData =new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst', rep('', 2)),
                                           NonIDQual = c('', 'Market', ''),
                                           IDDD = c(rep('', 2), 'NewOrders'),
                                           NumIdEst = c(rep('', 2), '.'),
                                           Market = c(rep('', 2), '1'),
                                           UnitName = c('numidest', '', 'cp09'),
                                           InFiles = rep('FF', 3))),
ParaData = new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst', rep('', 2)),
                                           NonIDQual = c('', 'Action', ''),
                                           IDDD = c(rep('', 2), 'Date'),
                                           NumIdEst = c(rep('', 2), '.'),
                                           Action = c(rep('', 2), 'Imputation'),
                                           UnitName = c('numidest', '', 'FechaImput'),
                                           InFiles = rep('FP', 3))),
AggWeights = new(Class = 'VNCdt', data.table(IDQual = c('CCAA', 'NACE09', ''),
                                           NonIDQual = rep('', 3),
                                           IDDD = c('', '', 'Ponderacion'),
                                           CCAA = c('', '', '.'),
                                           NACE09 = c('', '', '.'),
                                           UnitName = c('Provincia', '', ''),
                                           InFiles = rep('FF', 3))))

VNC <- new(Class = 'VarNameCorresp', VarList)

### We build the specification data.tables
IDdt <- new( Class='DDdt',data.table(
    Variable = c('NumIdEst', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
    Sort = c('IDQual', rep('IDDD', 4)),
    Class = rep('character', 5),
    Length = c('11', '15', '15', '20','9'),
    Qual1 = c('', rep('NumIdEst', 4)),
    ValueRegExp = c('[0-9]{9}PP', '.+', '.+', '.+', '(6|9)[0-9]{8}')))
Microdt <- new( Class='DDdt',data.table(
    Variable = c('NumIdEst', 'Market', 'NewOrders'),
    Sort = c('IDQual', 'NonIDQual', 'IDDD'),
    Class = c(rep('character', 2), 'numeric'),
    Length = c('11', '2', '7'),
    Qual1 = c(rep('', 2), 'NumIdEst'),
    ValueRegExp = c('[0-9]{9}PP', '.+', '([0-9]{1, 10}| )')))
Paradt <-new( Class='DDdt', data.table(
    Variable = c('NumIdEst', 'Action', 'Date'),
    Sort = c('IDQual', 'NonIDQual', 'IDDD'),
    Class = rep('character', 3),
    Length = c('1', '10', '10'),
    Qual1 = c(rep('', 2), 'NumIdEst'),
    Qual2 = c(rep('', 2), 'Action'),
    ValueRegExp = c('[0-9]{9}PP', 'Collection|Editing|Imputation', 
                    '(([0-9]{2}-(0[1-9]|1(0-2))-[0-9]{4})| )')))
Aggdt <- new(Class='DDdt',
             data.table(Variable = c('CCAA', 'NACE09', 'Ponderacion'),
                        Sort = c(rep('IDQual', 2), 'IDDD'),
                        Class = c(rep('character', 2), 'numeric'),
                        Length = c('2', '4', '7'),
                        Qual1 = c(rep('', 2), 'CCAA'),
                        Qual2 = c(rep('', 2), 'NACE09'),
                        ValueRegExp = c('[0-9]{4}', '([0-4][0-9])|(5[0-2])', 
                                        '([0-9]{1, 15}| )')))

DD <- new(Class = 'DD', 
          VarNameCorresp = VNC, 
          ID = IDdt, 
          MicroData = Microdt, 
          ParaData = Paradt,
          AggWeights = Aggdt)

getAggWeights(DD)

}

