% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\docType{methods}
\name{getData}
\alias{getData}
\alias{getData,DD-method}
\alias{getData,StQ-method}
\alias{getData,StQList-method}
\title{Return slot Data from an object}
\usage{
getData(object, VarNames)

\S4method{getData}{DD}(object, VarNames)

\S4method{getData}{StQ}(object, VarNames)

\S4method{getData}{StQList}(object, VarNames)
}
\arguments{
\item{object}{Object whose (possibly subsetted) slot \code{Data} is queried.}

\item{VarNames}{Character vector with the variable names subsetting the data 
set.}
}
\value{
In the case of \linkS4class{StQ} objects, it returns a 
\linkS4class{data.table} with key-value pair structure corresponding to slot
\code{Data} from the input object with the values of the column \code{IDDD} 
restriCted to variable names specified in \code{VarNames}.
}
\description{
\code{getData} returns slot \code{Data} from the input object.  

In the case of objects of class \linkS4class{StQ}, it returns a data set 
restricted to those root variable names in the column \code{IDDD} of slot
\code{Data} specified in the input parameter \code{VarNames}.

Input objects can be of class: 
\itemize{
\item \code{StQ}: The input parameters are an object of class 
\linkS4class{StQ} and a character vector, \code{VarNames}, with variable 
names. It returns the \code{data.table} corresponding to slot \code{Data} of 
such an object, but only with variables included in \code{VarNames}.

If no variable name is specified in \code{VarNames}, it returns the complete  
slot \code{Data}. 

\item \code{DD}: The input parameter is an object of class \linkS4class{DD}. 
The parameter \code{VarNames} has no effect. 
}
}
\examples{
# From DD objects 
DDData <- data.table(Variable = c('NumIdEst', 'EsMercNac', 'EsMercEuro',
                                  'EsMercRM', 'Cod', 'IEPEntradaPed'),
                     Sort = c('IDQual', 'NonIDQual', 'NonIDQual', 
                              'NonIDQual', 'NonIDQual', 'IDDD'),
                     Class = c('character', 'character', 'character', 
                               'character', 'character', 'character'),
                     Qual1 = c('', '', '', '', '', 'NumIdEst'))
VarList <- list(data.table(IDQual = c('NumIdEst','','','',''),
                     NonIDQual = c('EsMercNac', 'EsMercEuro', 'EsMercRM', 
                                   'Cod',''),
                     IDDD = c('','','','','IEPEntradaPed'),
                     Unit1 = c('','','','','')))
VarNameCorresp <- new(Class = 'VarNameCorresp', VarNameCorresp = VarList)
DD <- new(Class = 'DD', MicroData = DDData, VarNameCorresp = VarNameCorresp)
getData(DD)


# From an StQ object 
VarNames <- c('IASSCifraNeg', 'IASSEmpleo')
getData(ExampleQ, VarNames)

VarNames <- c('IASSCifraNeg', 'IASSEmpleo_0')
getData(ExampleQ, VarNames)

# From an StQList object 
mm <- c(paste0('0', 1:9), 10:12)
TimePer <- paste0('MM', mm, '2015')
QList <- vector('list', 12)
QList <- lapply(QList, function(x) ExampleQ)
names(QList) <- TimePer
QList <- new(Class = 'StQList', QList)
VarNames <- c('IASSCifraNeg', 'IASSEmpleo')
getData(QList, VarNames)

}

