% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\alias{getData,StQ-method}
\alias{getData,rawStQ-method}
\alias{getData,StQList-method}
\alias{getData,rawStQList-method}
\title{Return slot \code{Data} from an object possibly subsetted to a set of variables}
\usage{
getData(object, VarNames)

\S4method{getData}{StQ}(object, VarNames)

\S4method{getData}{rawStQ}(object, VarNames)

\S4method{getData}{StQList}(object, VarNames)

\S4method{getData}{rawStQList}(object, VarNames)
}
\arguments{
\item{object}{Object whose (possibly subsetted) slot \code{Data} is queried.}

\item{VarNames}{\code{Character} vector with the variable names subsetting the data set.}
}
\value{
In the case of \link{StQ}/\linkS4class{rawStQ} objects, it returns a
\linkS4class{data.table} with key-value pair structure corresponding to slot \code{Data} from the
input object with the values of the variables restricted to variable names specified in
\code{VarNames}. In the case of \linkS4class{StQList}/\linkS4class{rawStQList} objects, it
returns a list of \link{StQ}/linkS4class{rawStQ} objects.
}
\description{
\code{getData} returns slot \code{Data} from the input object possibly subsetted to
those variables specified as an input parameter. 

Input objects can be of class:
\itemize{
\item \code{StQ}: The input parameters are an object of class \link{StQ} and a character
vector, \code{VarNames}, with variable names. It returns the \code{data.table} corresponding to
slot \code{Data} of such an object, but only with variables included in \code{VarNames}. If in 
\code{VarNames} only the roots of variable names in the column \code{IDDD} of slot \code{Data}
are specified, it returns a data set with all variables which have those roots.

If no variable name is specified in \code{VarNames}, it returns the complete slot \code{Data}.

\item \code{StQList}: The input parameters are an object of class \linkS4class{StQList} and a
character vector, \code{VarNames}, with variable names. It returns a list of \link{StQ}
objects, but only with variables included in \code{VarNames}, as it is explained above.

\item \code{rawStQ}: The input parameters are an object of class \linkS4class{rawStQ} and a
character vector, \code{VarNames}, with variable names. It returns the \code{data.table}
corresponding to slot \code{Data} of such an object, but only with variables included in
\code{VarNames}.

\item \code{rawStQList}: The input parameters are an object of class \linkS4class{rawStQList} and
 a character vector, \code{VarNames}, with variable names. It returns a list of
 \linkS4class{rawStQ} objects, but only with variables included in \code{VarNames}.
}
}
\examples{
# From an StQ object
VarNames <- c('Employees_1.')
getData(ExampleStQ, VarNames)

VarNames <- c('Turnover')
getData(ExampleStQ, VarNames)

# From a rawStQ object
VarNames <- c('Turnover')
getData(ExamplerawStQ, VarNames)

# From an StQList object
mm <- c(paste0('0', 1:9), 10:12)
TimePer <- paste0('MM', mm, '2015')
QList <- vector('list', 12)
QList <- lapply(QList, function(x) ExampleStQ)
names(QList) <- TimePer
QList <- BuildStQList(QList)
VarNames <- c('Turnover', 'Employees_1.')
getData(QList, VarNames)

}
