% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\docType{methods}
\name{getData}
\alias{getData}
\alias{getData,StQ-method}
\alias{getData,StQList-method}
\alias{getData,rawStQ-method}
\alias{getData,rawStQList-method}
\title{Return slot \code{Data} from an object possibly subsetted to a set of variables}
\usage{
getData(object, VarNames, DDslot = "MicroData")

<<<<<<< HEAD
\S4method{getData}{StQ}(object, VarNames)

\S4method{getData}{rawStQ}(object, VarNames)

\S4method{getData}{StQList}(object, VarNames)

\S4method{getData}{rawStQList}(object, VarNames)
=======
\S4method{getData}{DD}(object, VarNames, DDslot = "MicroData")

\S4method{getData}{StQ}(object, VarNames, DDslot = "MicroData")

\S4method{getData}{StQList}(object, VarNames, DDslot = "MicroData")

\S4method{getData}{rawStQ}(object, VarNames, DDslot = "MicroData")

\S4method{getData}{rawStQList}(object, VarNames, DDslot = "MicroData")
>>>>>>> 5034523f22c62817420f2f5687369d62b4523cd8
}
\arguments{
\item{object}{Object whose (possibly subsetted) slot \code{Data} is queried.}

\item{VarNames}{\code{Character} vector with the variable names subsetting the data set.}

\item{DDslot}{\code{Character} vector of length 1 with the name of DD slot whose variables are
queried in the input parameter VarNames. Its default value is \code{MicroData} and has no efect 
if the input object is a \linkS4class{DD} object.}
}
\value{
In the case of \linkS4class{StQ}/\linkS4class{rawStQ} objects, it returns a 
\linkS4class{data.table}/\linkS4class{rawDatadt} with key-value pair structure corresponding to 
slot \code{Data} from the input object with the values of the column \code{IDDD} restriCted to 
variable names specified in \code{VarNames}. In the case of \linkS4class{DD} objects, it returns 
the slot MicroData. In the case of \linkS4class{StQList}/\linkS4class{rawStQList} objects, it 
returns a list of \linkS4class{StQ}/linkS4class{rawStQ} objects.
}
\description{
\code{getData} returns slot \code{Data} from the input object possibly subsetted to
those variables specified as an input parameter. In the case of those variables pertaining to 
more than one slot of the input object (in particular for \linkS4class{DD} objects), a third 
parameter \code{DDslot} with default value MicroData must be specified. 
<<<<<<< HEAD

In the case of objects of class \linkS4class{DD}, it returns the slot \code{MicroData} of the
input object.

=======

In the case of objects of class \linkS4class{DD}, it returns the slot \code{MicroData} of the
input object.

>>>>>>> 5034523f22c62817420f2f5687369d62b4523cd8
In the case of objects of class \linkS4class{StQ}, it returns a data set restricted to those root
variable names in the column \code{IDDD} of slot \code{Data} specified in the input parameter
\code{VarNames}.

Input objects can be of class: 
\itemize{
\item \code{StQ}: The input parameters are an object of class \linkS4class{StQ} and a character 
vector, \code{VarNames}, with variable names. It returns the \code{data.table} corresponding to 
slot \code{Data} of such an object, but only with variables included in \code{VarNames}.

If no variable name is specified in \code{VarNames}, it returns the complete slot \code{Data}. 

\item \code{DD}: The input parameter is an object of class \linkS4class{DD}. The parameter
\code{VarNames} has no effect.

\item \code{StQList}: The input parameters are an object of class \linkS4class{StQList} and a 
character vector, \code{VarNames}, with variable names. It returns a list of \linkS4class{StQ} 
objects, but only with variables included in \code{VarNames}.

\item \code{rawStQ}: The input parameters are an object of class \linkS4class{rawStQ} and a 
character vector, \code{VarNames}, with variable names. It returns the \linkS4class{rawDatadt} 
corresponding to slot \code{Data} of such an object, but only with variables included in 
\code{VarNames}.

\item \code{rawStQList}: The input parameters are an object of class \linkS4class{rawStQList} and
 a character vector, \code{VarNames}, with variable names. It returns a list of 
 \linkS4class{rawStQ} objects, but only with variables included in \code{VarNames}.
}
}
\examples{
# From DD objects 
data(ExampleDD)
getData(ExampleDD)
getData(ExampleDD, DDslot = 'Aggregates')
getData(ExampleDD, VarNames = c('Turnover', 'Employees_1.'))

# From an StQ object 
VarNames <- c('Employees_1.')
getData(ExampleStQ, VarNames)

VarNames <- c('Turnover')
getData(ExampleStQ, VarNames)

# From a rawStQ object
VarNames <- c('Turnover')
getData(ExamplerawStQ, VarNames)

VarNames <- c('Turnover')
getData(ExampleStQ, VarNames)

# From an StQList object 
mm <- c(paste0('0', 1:9), 10:12)
TimePer <- paste0('MM', mm, '2015')
QList <- vector('list', 12)
QList <- lapply(QList, function(x) ExampleStQ)
names(QList) <- TimePer
QList <- BuildStQList(QList)
VarNames <- c('Turnover', 'Employees_2.1')
getData(QList, VarNames)

}

