% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDD.R
\docType{methods}
\name{getDD}
\alias{getDD}
\alias{getDD,StQ-method}
\alias{getDD,StQList-method}
\alias{getDD,rawStQ-method}
\title{Return slot DD from the input object}
\usage{
getDD(object)

\S4method{getDD}{StQ}(object)

\S4method{getDD}{rawStQ}(object)

\S4method{getDD}{StQList}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{StQ}, \linkS4class{rawStQ}, \linkS4class{StQList} or 
\linkS4class{rawStQList}.}
}
\value{
Object of class \linkS4class{DD} corresponding to the slot \code{DD} of the input 
parameter.
}
\description{
\code{getDD} extracts the slot \code{DD} of the input object.
}
\examples{
data(ExampleStQ)
DD <- getDD(ExampleStQ)
DD
str(DD)

}

