% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getDD.R
\docType{methods}
\name{getDD}
\alias{getDD}
\alias{getDD,StQ-method}
\alias{getDD,StQList-method}
\alias{getDD,rawStQ-method}
\title{Return slot \code{DD} from the input object}
\usage{
getDD(object)

\S4method{getDD}{StQ}(object)

\S4method{getDD}{StQList}(object)

\S4method{getDD}{rawStQ}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{StQ} or \linkS4class{StQList}.}
}
\value{
Object of class \linkS4class{DD} corresponding to the slot \code{DD} of the input
parameter.
}
\description{
\code{getDD} extracts the slot \code{DD} of the input object.


}
\examples{
data(ExampleStQ)
DD <- getDD(ExampleStQ)
DD
str(DD)
}

