% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDoubleDotQual.R
\docType{methods}
\name{getDoubleDotQual}
\alias{getDoubleDotQual}
\alias{getDoubleDotQual,DD-method}
\alias{getDoubleDotQual,StQ-method}
\alias{getDoubleDotQual,VNCdt-method}
\alias{getDoubleDotQual,VarNameCorresp-method}
\title{Return non-unit qualifiers used in compound variable names}
\usage{
getDoubleDotQual(object, Component = names(object))

\S4method{getDoubleDotQual}{VNCdt}(object, Component = names(object))

\S4method{getDoubleDotQual}{VarNameCorresp}(object, Component = names(object))

\S4method{getDoubleDotQual}{DD}(object, Component = names(getVNC(object)))

\S4method{getDoubleDotQual}{StQ}(object, Component = names(object))
}
\arguments{
\item{object}{Object containing the dot qualifiers.}

\item{Component}{\code{Character} vector with the components or slots from which dot qualifiers 
are required.}
}
\value{
\code{Character} vector with all the double dot qualifier names.
}
\description{
\code{getDoubleDotQual} returns a character vector with all non-unit qualifier names 
specified with a double wildcard dot in the VNC component of the input object.
}
\examples{
library(data.table)
VarList <- list(ID = new(Class = 'VNCdt',
                data.table(IDQual = c('NumIdEst', rep('', 4)),
                           NonIDQual = c('','','','',''),
                           IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                           NumIdEst = c('', rep('.', 4)),
                           UnitName = c('numidest', 'nombre', 'apellidos', 'direccion', 'telefono'),
                           InFiles = rep('FI', 5))),
MicroData = new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst', rep('', 2)),
                                            NonIDQual = c('', 'Market', ''),
                                            IDDD = c(rep('', 2), 'NewOrders'),
                                            NumIdEst = c(rep('', 2), '.'),
                                            Market = c(rep('', 2), '1.'),
                                            UnitName = c('numidest', '', 'cp09'),
                                            InFiles = rep('FF, FD, FG', 3))),
ParaData = new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst', rep('', 2)),
                                           NonIDQual = c('', 'Action', ''),
                                           IDDD = c(rep('', 2), 'Date'),
                                           NumIdEst = c(rep('', 2), '.'),
                                           Action = c(rep('', 2), '..'),
                                           UnitName = c('numidest', '', 'FechaImput'),
                                           InFiles = rep('FP', 3))),
AggWeights = new(Class = 'VNCdt', data.table(IDQual = c('CCAA', 'NACE09', ''),
                                           NonIDQual = rep('', 3),
                                           IDDD = c('', '', 'Ponderacion'),
                                           CCAA = c('', '', '.'),
                                           NACE09 = c('', '', '.'),
                                           UnitName = c('Provincia', '', ''),
                                           InFiles = rep('FA', 3))))
VNC <- new(Class = 'VarNameCorresp', .Data = VarList)
getDotQual(VNC)


}

