% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDDD.R
\docType{methods}
\name{getIDDD}
\alias{getIDDD}
\alias{getIDDD,DD-method}
\alias{getIDDD,StQ-method}
\title{Return the variable names (IDDD) included in the input object}
\usage{
getIDDD(object)

\S4method{getIDDD}{DD}(object)

\S4method{getIDDD}{StQ}(object)
}
\arguments{
\item{object}{Object with the IDDD variable identifiers.}
}
\value{
Character vector with all the variable names.
}
\description{
\code{getIDDD} returns a character vector with all variable 
names (IDDD) included in the input object.
}
\examples{
getIDDD(ExampleQ)

}

