% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDDD.R
\docType{methods}
\name{getIDDD}
\alias{getIDDD}
\alias{getIDDD,DD-method}
\alias{getIDDD,DDdt-method}
\alias{getIDDD,Datadt-method}
\alias{getIDDD,StQ-method}
\alias{getIDDD,VNCdt-method}
\alias{getIDDD,VarNameCorresp-method}
\title{Return IDDD identifiers from an object}
\usage{
getIDDD(object, CompNames)

\S4method{getIDDD}{VNCdt}(object, CompNames)

\S4method{getIDDD}{VarNameCorresp}(object, CompNames)

\S4method{getIDDD}{DDdt}(object, CompNames)

\S4method{getIDDD}{DD}(object, CompNames)

\S4method{getIDDD}{Datadt}(object, CompNames)

\S4method{getIDDD}{StQ}(object, CompNames)
}
\arguments{
\item{object}{Object whose IDDD identifiers are required.}

\item{CompNames}{Character vector with the names of components or slots of the 
object from which the IDDD identifiers are requested.}
}
\value{
In the case of \linkS4class{VarNameCorresp} it returns a character
vector with the IDDD identifiers from the components specified in Namesdt. 
If no Namesdt is specified, it returns the IDDD identifier from all the 
components of the object.
}
\description{
\code{getIDDD} returns the IDDD identifiers from the input 
object.
}
\examples{
library(data.table)
VarList <- list(ID = new(Class = 'VNCdt',
                data.table(IDQual = c('NumIdEst', rep('', 4)),
                           NonIDQual = c('','','','',''),
                           IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                           NumIdEst = c('', rep('.', 4)),
                           Unit1 = c('numidest', 'nombre', 'apellidos', 'direccion', 'telefono'))
 ),
MicroData =new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst', rep('', 2)),
                                           NonIDQual = c('', 'Market', ''),
                                           IDDD = c(rep('', 2), 'NewOrders'),
                                           NumIdEst = c(rep('', 2), '.'),
                                           Market = c(rep('', 2), '1.'),
                                           Unit1 = c('numidest', '', 'cp09'))),
ParaData = new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst', rep('', 2)),
                                           NonIDQual = c('', 'Action', ''),
                                           IDDD = c(rep('', 2), 'Date'),
                                           NumIdEst = c(rep('', 2), '.'),
                                           Action = c(rep('', 2), 'Imputation'),
                                           Unit1 = c('numidest', '', 'FechaImput'))),
AggWeights = new(Class = 'VNCdt', data.table(IDQual = c('CCAA', 'NACE09', ''),
                                           NonIDQual = rep('', 3),
                                           IDDD = c('', '', 'Ponderacion'),
                                           CCAA = c('', '', '.'),
                                           NACE09 = c('', '', '.'),
                                           Unit1 = c('Provincia', '', ''))))

VNC <- new(Class = 'VarNameCorresp', .Data = VarList)
getIDDD(VNC)


}

