% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDDD.R
\docType{methods}
\name{getIDDD}
\alias{getIDDD}
\alias{getIDDD,DD-method}
\alias{getIDDD,StQ-method}
\alias{getIDDD,VarNameCorresp-method}
\title{Return IDDD identifiers from an object}
\usage{
getIDDD(object)

\S4method{getIDDD}{VarNameCorresp}(object)

\S4method{getIDDD}{DD}(object)

\S4method{getIDDD}{StQ}(object)
}
\arguments{
\item{object}{Object whose IDDD identifiers are queried.}
}
\value{
In the case of \linkS4class{VarNameCorresp} it returns the IDDD
identifiers from all components of of its slot \code{VarNameCorresp}.
}
\description{
\code{getIDDD} returns all IDDD identifiers from the input 
object.
}
\examples{
# A more elaborate example
VarList <- list(MicroData = data.table(IDQual = c('NumIdEst','','', '', '', '', ''),
                NonIDQual = c('', 'IsNatMarket', 'IsEuroMarket', 'IsRWMarket', '', '', ''),
                IDDD = c('','','','','Turnover', 'Turnover', 'Turnover'),
                NumIdEst = c('', '', '', '', '.', '.', '.'),
                IsNatMarket = c('', '', '', '', '1', '0', '0'),
                IsEuroMarket = c('', '', '', '', '', '1', '0'),
                IsRWMarket = c('', '', '', '', '', '', '1'),
                Unit1 = c('','','','','cn01', 'cn02', 'cn03')))
Example <- new(Class = 'VarNameCorresp', VarList)
getIDDD(Example)

}

