% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDDD.R
\docType{methods}
\name{getIDDD}
\alias{getIDDD}
\alias{getIDDD,DD-method}
\alias{getIDDD,DDdt-method}
\alias{getIDDD,Datadt-method}
\alias{getIDDD,StQ-method}
\alias{getIDDD,VNCdt-method}
\alias{getIDDD,VarNameCorresp-method}
\title{Return IDDD identifiers from an object}
\usage{
getIDDD(object, Namesdt)

\S4method{getIDDD}{VNCdt}(object)

\S4method{getIDDD}{VarNameCorresp}(object, Namesdt)

\S4method{getIDDD}{DDdt}(object)

\S4method{getIDDD}{DD}(object, Namesdt)

\S4method{getIDDD}{Datadt}(object)

\S4method{getIDDD}{StQ}(object)
}
\arguments{
\item{object}{Object whose IDDD identifiers are queried.}
}
\value{
In the case of \linkS4class{VarNameCorresp} it returns the IDDD
identifiers from all components of of its slot \code{VarNameCorresp}.
}
\description{
\code{getIDDD} returns all IDDD identifiers from the input 
object.
}
\examples{
# A more elaborate example
VarList <- list(
  ID = new(Class = 'VNCdt', 
           .Data = data.table(
                 IDQual = c('NumIdEst', rep('', 4)),
                 NonIDQual = rep('', 5),
                 IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                 NumIdEst = c('', rep('.', 4)),
                 Unit1 = c('numidest', 'nombre', 'apellidos', 'direccion', 'telefono'))),
  MicroData = new(Class = 'VNCdt', 
                  .Data = data.table(
                      IDQual = c('NumIdEst', rep('', 4)),
                      NonIDQual = c('', 'IsNatMarket', 'IsEuroMarket', 'IsRWMarket', ''),
                      IDDD = c(rep('', 4), 'NewOrders'),
                      NumIdEst = c(rep('', 4), '.'),
                      IsNatMarket = c(rep('', 4), '0'),
                      IsEuroMarket = c(rep('', 4), '0'),
                      IsRWMarket = c(rep('', 4), '1'),
                      Unit1 = c('numidest', rep('', 3), 'cp09'))),
 Aggregates = new(Class = 'VNCdt', 
                  .Data = data.table(
                     IDQual = c('Province', 'NACE', 'IsNatMarket', ''),
                     NonIDQual = rep('', 4),
                     IDDD = c('', '', '', 'TotalTurnover'),
                     Province = c('', '', '', '.'),
                     NACE = c('', '', '', '.'),
                     IsNatMarket = c('', '', '', '1'),
                     Unit1 = c('provincia', 'actividad', '', 'cn01'))))
Example <- new(Class = 'VarNameCorresp', .Data = VarList)
getIDDD(Example)

}

