% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDDD.R
\docType{methods}
\name{getIDDD}
\alias{getIDDD}
\alias{getIDDD,VNC-method}
\alias{getIDDD,DD-method}
\alias{getIDDD,StQ-method}
\title{Return IDDD identifiers (statistical variable names) from the input object}
\usage{
getIDDD(object, CompNames = names(object))

\S4method{getIDDD}{VNC}(object, CompNames = names(object))

\S4method{getIDDD}{DD}(object, CompNames = setdiff(names(object), "VNC"))

\S4method{getIDDD}{StQ}(object, CompNames = setdiff(names(getDD(object)),
  "VNC"))
}
\arguments{
\item{object}{Object whose IDDD identifiers in the slot \link{DD} are required.}

\item{CompNames}{\code{Character} vector with the names of components or slots of the object from
which the IDDD identifiers are requested.}
}
\value{
Returns a character vector with the IDDD identifiers from the components specified in
CompNames. If no CompNames parameter is specified, it returns the IDDD identifiers from all the
components of the input object.
}
\description{
\code{getIDDD} returns the IDDD identifiers from input object.
}
\examples{
# An example:
library(data.table)
### We build the VNC object
VarList <- list(ID = data.table(IDQual = c('NumIdEst', rep('', 4)),
                                NonIDQual = c('','','','',''),
                                IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                                NumIdEst = c('', rep('.', 4)),
                                UnitName = c('numidest', 'nombre', 'apellidos', 'direccion', 'telefono'),
                                InFiles = rep('FI', 5)),
MicroData = data.table(IDQual = c('NumIdEst', rep('', 2)),
                       NonIDQual = c('', 'Market', ''),
                       IDDD = c(rep('', 2), 'NewOrders'),
                       NumIdEst = c(rep('', 2), '.'),
                       Market = c(rep('', 2), '1.'),
                       UnitName = c('numidest', '', 'cp09'),
                       InFiles = rep('FF', 3)),
ParaData = data.table(IDQual = c('NumIdEst', rep('', 2)),
                      NonIDQual = c('', 'Action', ''),
                      IDDD = c(rep('', 2), 'Date'),
                      NumIdEst = c(rep('', 2), '.'),
                      Action = c(rep('', 2), 'Imputation'),
                      UnitName = c('numidest', '', 'FechaImput'),
                      InFiles = rep('FP', 3)),
AggWeights = data.table(IDQual = c('CCAA', 'NACE09', ''),
                        NonIDQual = rep('', 3),
                        IDDD = c('', '', 'Ponderacion'),
                        CCAA = c('', '', '.'),
                        NACE09 = c('', '', '.'),
                        UnitName = c('Provincia', '', ''),
                        InFiles = rep('FF', 3)))

VNC <- BuildVNC(VarList)

getIDDD(VNC)


}
