% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDQual.R
\docType{methods}
\name{getIDQual}
\alias{getIDQual}
\alias{getIDQual,DD-method}
\alias{getIDQual,DDdt-method}
\alias{getIDQual,Datadt-method}
\alias{getIDQual,StQ-method}
\alias{getIDQual,VNCdt-method}
\alias{getIDQual,VarNameCorresp-method}
\title{Return IDDD identifiers from an object}
\usage{
getIDQual(object, Namesdt)

\S4method{getIDQual}{VNCdt}(object)

\S4method{getIDQual}{VarNameCorresp}(object, Namesdt)

\S4method{getIDQual}{DDdt}(object)

\S4method{getIDQual}{DD}(object, Namesdt)

\S4method{getIDQual}{Datadt}(object)

\S4method{getIDQual}{StQ}(object)
}
\arguments{
\item{object}{Object with the IDQual unit qualifier.}

\item{Namesdt}{Character vector with the components or slots from which IDQuals
are requiered.}
}
\value{
Character vector with all the unit qualifier names.
}
\description{
\code{getIDQual} returns a character vector with all unit qualifier 
names (IDQual) included in the input object.
}
\examples{
# A more elaborate example
VarList <- list(
  ID = new(Class = 'VNCdt', 
           .Data = data.table(
                 IDQual = c('NumIdEst', rep('', 4)),
                 NonIDQual = rep('', 5),
                 IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                 NumIdEst = c('', rep('.', 4)),
                 Unit1 = c('numidest', 'nombre', 'apellidos', 'direccion', 'telefono'))),
  MicroData = new(Class = 'VNCdt', 
                  .Data = data.table(
                      IDQual = c('NumIdEst', rep('', 4)),
                      NonIDQual = c('', 'IsNatMarket', 'IsEuroMarket', 'IsRWMarket', ''),
                      IDDD = c(rep('', 4), 'NewOrders'),
                      NumIdEst = c(rep('', 4), '.'),
                      IsNatMarket = c(rep('', 4), '0'),
                      IsEuroMarket = c(rep('', 4), '0'),
                      IsRWMarket = c(rep('', 4), '1'),
                      Unit1 = c('numidest', rep('', 3), 'cp09'))),
 Aggregates = new(Class = 'VNCdt', 
                  .Data = data.table(
                     IDQual = c('Province', 'NACE', 'IsNatMarket', ''),
                     NonIDQual = rep('', 4),
                     IDDD = c('', '', '', 'TotalTurnover'),
                     Province = c('', '', '', '.'),
                     NACE = c('', '', '', '.'),
                     IsNatMarket = c('', '', '', '1'),
                     Unit1 = c('provincia', 'actividad', '', 'cn01'))))
Example <- new(Class = 'VarNameCorresp', .Data = VarList)
getIDQual(Example)


}

