% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDQual.R
\docType{methods}
\name{getIDQual}
\alias{getIDQual}
\alias{getIDQual,DD-method}
\alias{getIDQual,DDdt-method}
\alias{getIDQual,Datadt-method}
\alias{getIDQual,StQ-method}
\alias{getIDQual,VNCdt-method}
\alias{getIDQual,VarNameCorresp-method}
\title{Return IDDD identifiers from an object}
\usage{
getIDQual(object, Namesdt)

\S4method{getIDQual}{VNCdt}(object)

\S4method{getIDQual}{VarNameCorresp}(object, Namesdt)

\S4method{getIDQual}{DDdt}(object)

\S4method{getIDQual}{DD}(object, Namesdt)

\S4method{getIDQual}{Datadt}(object)

\S4method{getIDQual}{StQ}(object)
}
\arguments{
\item{object}{Object with the IDQual unit qualifier.}

\item{Namesdt}{Character vector with the components or slots from which IDQuals
are requiered.}
}
\value{
Character vector with all the unit qualifier names.
}
\description{
\code{getIDQual} returns a character vector with all unit qualifier 
names (IDQual) included in the input object.
}
\examples{
library(data.table)
VarList <- list(ID = new(Class = 'VNCdt',
                data.table(IDQual = c('NumIdEst', rep('', 4)),
                           NonIDQual = c('','','','',''),
                           IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                           NumIdEst = c('', rep('.', 4)),
                           Unit1 = c('numidest', 'nombre', 'apellidos', 'direccion', 'telefono'))
 ),
MicroData =new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst', rep('', 2)),
                                           NonIDQual = c('', 'Market', ''),
                                           IDDD = c(rep('', 2), 'NewOrders'),
                                           NumIdEst = c(rep('', 2), '.'),
                                           Market = c(rep('', 2), '1.'),
                                           Unit1 = c('numidest', '', 'cp09'))),
ParaData = new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst', rep('', 2)),
                                           NonIDQual = c('', 'Action', ''),
                                           IDDD = c(rep('', 2), 'Date'),
                                           NumIdEst = c(rep('', 2), '.'),
                                           Action = c(rep('', 2), 'Imputation'),
                                           Unit1 = c('numidest', '', 'FechaImput'))),
AggWeights = new(Class = 'VNCdt', data.table(IDQual = c('CCAA', 'NACE09', ''),
                                           NonIDQual = rep('', 3),
                                           IDDD = c('', '', 'Ponderacion'),
                                           CCAA = c('', '', '.'),
                                           NACE09 = c('', '', '.'),
                                           Unit1 = c('Provincia', '', ''))))
VNC <- new(Class = 'VarNameCorresp', .Data = VarList)
getIDQual(VNC)


}

