% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDQual.R
\docType{methods}
\name{getIDQual}
\alias{getIDQual}
\alias{getIDQual,VNC-method}
\alias{getIDQual,DD-method}
\alias{getIDQual,StQ-method}
\alias{getIDQual,StQList-method}
\title{Return unit qualifiers from an object}
\usage{
getIDQual(object, CompNames = names(object))

\S4method{getIDQual}{VNC}(object, CompNames = names(object))

\S4method{getIDQual}{DD}(object, CompNames = setdiff(names(object),
  "VNC"))

\S4method{getIDQual}{StQ}(object,
  CompNames = setdiff(names(getDD(object)), "VNC"))

\S4method{getIDQual}{StQList}(object, CompNames = names(object))
}
\arguments{
\item{object}{Object with the unit qualifiers (IDQual) to be queried.}

\item{CompNames}{Character vector with the components or slots from which IDQuals are queried.}
}
\value{
Character vector with all the unit qualifier names.
}
\description{
\code{getIDQual} returns a character vector with all unit qualifier names (IDQual)
contained in the input object.
}
\details{
Unit qualifiers are those qualifiers identifying statistical units.
}
\examples{
# An example:
library(data.table)
### We build the VNC object
VarList <- list(ID = data.table(IDQual = c('NumIdEst', rep('', 4)),
                                NonIDQual = c('','','','',''),
                                IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                                NumIdEst = c('', rep('.', 4)),
                                UnitName = c('numidest', 'nombre', 'apellidos', 'direccion', 'telefono'),
                                InFiles = rep('FI', 5)),
MicroData = data.table(IDQual = c('NumIdEst', rep('', 2)),
                       NonIDQual = c('', 'Market', ''),
                       IDDD = c(rep('', 2), 'NewOrders'),
                       NumIdEst = c(rep('', 2), '.'),
                       Market = c(rep('', 2), '1.'),
                       UnitName = c('numidest', '', 'cp09'),
                       InFiles = rep('FF', 3)),
ParaData = data.table(IDQual = c('NumIdEst', rep('', 2)),
                      NonIDQual = c('', 'Action', ''),
                      IDDD = c(rep('', 2), 'Date'),
                      NumIdEst = c(rep('', 2), '.'),
                      Action = c(rep('', 2), 'Imputation'),
                      UnitName = c('numidest', '', 'FechaImput'),
                      InFiles = rep('FP', 3)),
AggWeights = data.table(IDQual = c('CCAA', 'NACE09', ''),
                        NonIDQual = rep('', 3),
                        IDDD = c('', '', 'Ponderacion'),
                        CCAA = c('', '', '.'),
                        NACE09 = c('', '', '.'),
                        UnitName = c('Provincia', '', ''),
                        InFiles = rep('FF', 3)))

VNC <- BuildVNC(VarList)

getIDQual(VNC)
getIDQual(VNC, 'MicroData')
getIDQual(VNC, 'ParaData')

}
