% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNonIDQual.R
\docType{methods}
\name{getNonIDQual}
\alias{getNonIDQual}
\alias{getNonIDQual,DD-method}
\alias{getNonIDQual,DDdt-method}
\alias{getNonIDQual,StQ-method}
\alias{getNonIDQual,VNCdt-method}
\alias{getNonIDQual,VarNameCorresp-method}
\title{Return the variable names included in columns 'NonIDQual" of the input object}
\usage{
getNonIDQual(object, Namesdt)

\S4method{getNonIDQual}{VNCdt}(object)

\S4method{getNonIDQual}{VarNameCorresp}(object, Namesdt)

\S4method{getNonIDQual}{DDdt}(object)

\S4method{getNonIDQual}{DD}(object, Namesdt)

\S4method{getNonIDQual}{StQ}(object)
}
\arguments{
\item{object}{Object with the NonIDQual variable qualifiers.}

\item{Namesdt}{Character vector with the components or slots of the object
from which NonIDQuals are required.}
}
\value{
Character vector with the NonIDQual variable qualifiers names.
}
\description{
\code{getNonIDQual} returns a character vector with all variable 
qualifiers names (NonIDQual) included in the input object.
}
\examples{
library(data.table)
VarList <- list(ID = new(Class = 'VNCdt',
                data.table(IDQual = c('NumIdEst', rep('', 4)),
                           NonIDQual = c('','','','',''),
                           IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                           NumIdEst = c('', rep('.', 4)),
                           Unit1 = c('numidest', 'nombre', 'apellidos', 'direccion', 'telefono'))
 ),
MicroData =new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst', rep('', 2)),
                                           NonIDQual = c('', 'Market', ''),
                                           IDDD = c(rep('', 2), 'NewOrders'),
                                           NumIdEst = c(rep('', 2), '.'),
                                           Market = c(rep('', 2), '1.'),
                                           Unit1 = c('numidest', '', 'cp09'))),
ParaData = new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst', rep('', 2)),
                                           NonIDQual = c('', 'Action', ''),
                                           IDDD = c(rep('', 2), 'Date'),
                                           NumIdEst = c(rep('', 2), '.'),
                                           Action = c(rep('', 2), 'Imputation'),
                                           Unit1 = c('numidest', '', 'FechaImput'))),
AggWeights = new(Class = 'VNCdt', data.table(IDQual = c('CCAA', 'NACE09', ''),
                                           NonIDQual = rep('', 3),
                                           IDDD = c('', '', 'Ponderacion'),
                                           CCAA = c('', '', '.'),
                                           NACE09 = c('', '', '.'),
                                           Unit1 = c('Provincia', '', ''))))
VNC <- new(Class = 'VarNameCorresp', .Data = VarList)
getNonIDQual(VNC)


}

