% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOtherDD.R
\name{getOtherDD}
\alias{getOtherDD}
\title{Return slot Other from a DD object}
\usage{
getOtherDD(object)
}
\arguments{
\item{object}{a DD Object whose slot \code{Other} is queried.}
}
\value{
\linkS4class{data.table} with data from slot \code{Other} of the
input \linkS4class{DD} object.
}
\description{
\code{getOtherDD} returns slot \code{Other} of the input
\linkS4class{DD} object.
}
\examples{
# An example with data created previosly:
data(ExampleDD)
getOtherDD(ExampleDD)

}

