% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParaData.R
\docType{methods}
\name{getParaData}
\alias{getParaData}
\alias{getParaData,DD-method}
\title{Return slot ParaData from a DD object}
\usage{
getParaData(object)

\S4method{getParaData}{DD}(object)
}
\arguments{
\item{object}{a DD Object whose slot \code{ParaData} is queried.}
}
\value{
\linkS4class{data.table} with data from slot \code{ParaData} of the
input \linkS4class{DD} object.
}
\description{
\code{getParaData} returns slot \code{ParaData} of the input
\linkS4class{DD} object.
}
\examples{
# An example with data created previosly:
data(ExampleDD)
getParaData(ExampleDD)

}

