% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPeriods.R
\docType{methods}
\name{getPeriods}
\alias{getPeriods}
\alias{getPeriods,rawStQList-method}
\alias{getPeriods,StQList-method}
\title{Return a vector with the names of the time period intervals}
\usage{
getPeriods(object)

\S4method{getPeriods}{rawStQList}(object)

\S4method{getPeriods}{StQList}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{rawStQList} or \linkS4class{StQList}.}
}
\value{
Character vector with the names of the time periods according to the notation of package
\linkS4class{RepoTimeInt}.
}
\description{
\code{getPeriods} returns a vector with the names of the time periods from the input
 object.
}
\examples{
mm <- c(paste0('0', 1:9), 10:12)
TimePer <- paste0('MM', mm, '2015')
PeriodList <- RepoTime::newRepoTime(TimePer)

EmptyrawQ <- rawStQ())
EmptyrawQList <- vector('list', 12)
EmptyrawQList <- lapply(EmptyrawQList, function(x) EmptyrawQ)
rawQList <- rawStQList(Data = EmptyrawQList, Periods = PeriodList)
getPeriods(rawQList)

EmptyQ <- StQ()
EmptyQList <- vector('list', 12)
EmptyQList <- lapply(EmptyQList, function(x) EmptyQ)
QList <- StQList(Data = EmptyQList, Periods = PeriodList)
getPeriods(QList)

}
