% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getUnits.R
\docType{methods}
\name{getUnits}
\alias{getUnits}
\alias{getUnits,StQ-method}
\title{Extract statistical units from an object of class \linkS4class{StQ}}
\usage{
getUnits(object)

\S4method{getUnits}{StQ}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{StQ}.}
}
\value{
It returns a \code{data.table} with the statistical units in the
input object.
}
\description{
This method identifies the IDQual qualifiers in the input object
and returns a \linkS4class{data.table} with the values of these qualifiers
for each statistical unit.
}
\examples{
data(ExampleQ)
getUnits(ExampleQ)
}

