% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUnits.R
\name{getUnits}
\alias{getUnits}
\alias{getUnits,StQ-method}
\title{Extract statistical units from an object}
\usage{
getUnits(object, DDslot = "MicroData")

\S4method{getUnits}{StQ}(object, DDslot = "MicroData")
}
\arguments{
\item{object}{Object of class \link{StQ}.}

\item{DDslot}{\code{Character} vector of length 1 with the name of slot of the attribute 
\link{DD}. The default value is \code{MicroData}.}
}
\value{
It returns a \code{data.table} with the statistical units from the input object.
}
\description{
This method identifies the IDQual qualifiers in the specified slot of the input 
object and returns a \linkS4class{data.table} with the values of these qualifiers for each 
statistical unit of this slot.
}
\examples{
data(ExampleStQ)
getUnits(ExampleStQ)

}
