% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVar.R
\docType{methods}
\name{getVar}
\alias{getVar}
\alias{getVar,StQ,character-method}
\alias{getVar,StQList,character-method}
\title{Return the values of a variable of specified statistical units}
\usage{
getVar(object, VarName, DDslot = "MicroData", Units = getUnits(object,
  "MicroData"))

\S4method{getVar}{StQ,character}(object, VarName, DDslot = "MicroData",
  Units = getUnits(object, "MicroData"))

\S4method{getVar}{StQList,character}(object, VarName,
  Units = getUnits(object[[length(object)]]))
}
\arguments{
\item{object}{Object of class \linkS4class{StQ}.}

\item{VarName}{Character vector with the name of the variable.}

\item{DDslot}{Character vector of length 1 with the name of DD slot in which
variable in VarName is defined. Its default value is \code{MicroData}.}

\item{Units}{\linkS4class{data.table} with the qualifier values identifying
each statistical unit in the input object.}
}
\value{
Matrix with the queried values.
}
\description{
\code{getVar} returns the values of the variable specified as
input parameter of those statistical units also specified as an input
parameter of an input object.
}
\examples{
library(data.table)
getVar(ExampleQ, 'Orders_0')
getVar(ExampleQ, 'Turnover')

}

