% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariables.R
\docType{methods}
\name{getVariables}
\alias{getVariables}
\alias{getVariables,DD-method}
\alias{getVariables,DDdt-method}
\alias{getVariables,StQ-method}
\title{Return the set of Variable Names from an object}
\usage{
getVariables(object, Sort, slots)

\S4method{getVariables}{DDdt}(object, Sort = c("IDDD", "IDQual", "NonIDQual"))

\S4method{getVariables}{DD}(object, Sort, slots)

\S4method{getVariables}{StQ}(object, Sort, slots)
}
\arguments{
\item{object}{Object whose Variable names are queried.}

\item{Sort}{Character vector with the Sort of Variables ('IDDD', 'IDQual', 
'NonIDQual').}

\item{slots}{Character vector with the slots names if object class is 
\linkS4class{DD}.}
}
\value{
In the case of \linkS4class{DD} it returns the Variable columns
from each of its slots different from \code{VarNameCorresp} slot.
}
\description{
\code{getVariables} returns all Variable names from the input 
object.
}
\examples{
data(ExampleDD) 
getVariables(ExampleDD)
getVariables(ExampleDD, Sort = 'IDDD', slots = 'MicroData')


}

