% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDDDToUnitNames.R
\docType{methods}
\name{IDDDToUnitNames}
\alias{IDDDToUnitNames}
\alias{IDDDToUnitNames,DD-method}
\alias{IDDDToUnitNames,StQ-method}
\alias{IDDDToUnitNames,VNCdt-method}
\alias{IDDDToUnitNames,VarNameCorresp-method}
\title{Convert IDDD variable names into their corresponding unit\emph{j} names}
\usage{
IDDDToUnitNames(object, IDDDNames, Unit)

\S4method{IDDDToUnitNames}{VNCdt}(object, IDDDNames, Unit)

\S4method{IDDDToUnitNames}{VarNameCorresp}(object, IDDDNames, Unit)

\S4method{IDDDToUnitNames}{DD}(object, IDDDNames, Unit)

\S4method{IDDDToUnitNames}{StQ}(object, IDDDNames, Unit)
}
\arguments{
\item{object}{Object with the IDDD variable identifiers.}

\item{IDDDNames}{character vector with the IDDD variables.}

\item{Unit}{character vector with the unit names.}
}
\value{
Data table with the IDDD variable names and their corresponding 
Unit\emph{j} names.
}
\description{
\code{IDDDToUnitNames} returns a data table with the unit
variable name for each IDDD variable name.
}
\examples{
library(data.table)
### We build the VNC object
VarList <- list(ID = new(Class = 'VNCdt',
                         .Data = data.table(IDQual = c('NumIdEst', rep('', 4)),
                                            NonIDQual = c(rep('',5)),
                                            IDDD = c('', 'Name', 'Surname', 'PostalAddr',
                                                 'PhoneNo'),
                                            NumIdEst = c('', rep('.', 4)),
                                            Unit1 = c('numidest', 'nombre', 'apellidos', 
                                               'direccion', 'telefono'))),
                MicroData = new(Class = 'VNCdt',
                                .Data =data.table(
                                       IDQual = c('NumIdEst', rep('', 4)),
                                       NonIDQual = c('', 'IsNatMarket', 
                                                     'IsEuroMarket', 
                                                     'IsRWMarket',
                                                     ''),
                                       IDDD = c(rep('', 4), 'NewOrders'),
                                       NumIdEst = c(rep('', 4), '.'),
                                       IsNatMarket = c(rep('', 4), '0'),
                                       IsEuroMarket = c(rep('', 4), '0'),
                                       IsRWMarket = c(rep('', 4), '1'),
                                       Unit1 = c('numidest', rep('', 3), 'cp09'))),
                ParaData = new(Class = 'VNCdt',
                                .Data = data.table(
                                       IDQual = c('NumIdEst', rep('', 2)),
                                       NonIDQual = c('', 'Action', ''),
                                       IDDD = c(rep('', 2), 'Date'),
                                       NumIdEst = c(rep('', 2), '.'),
                                       Action = c(rep('', 2), 'Imputation'),
                                       Unit1 = c('numidest', '', 'FechaImput'))),
                Aggregates = new(Class = 'VNCdt',
                                 .Data = data.table(
                                        IDQual = c('Province', 'NACE09', '', ''),
                                        NonIDQual = c(rep('', 2), 'IsNatMarket', ''),
                                        IDDD = c('', '', '', 'Turnover'),
                                        Province = c('', '', '', '.'),
                                        NACE09 = c('', '', '', '.'),
                                        IsNatMarket = c('', '', '', '1'),
                                        Unit1 = c('provincia', 'actividad', '', 'cn01'))))

VNC <- new(Class = 'VarNameCorresp', .Data = VarList)

### We build the specification data.tables
IDdt <- new(Class = "DDdt", 
            .Data = data.table(
                     Variable = c('NumIdEst', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                     Sort = c('IDQual', rep('IDDD', 4)),
                     Class = rep('character', 5),
                     Qual1 = c('', rep('NumIdEst', 4)),
                     ValueRegExp = c('[0-9]{9}PP', '.+', '.+', '.+', '(6|9)[0-9]{8}')))
Microdt <- new(Class = "DDdt", 
            .Data = data.table(
                     Variable = c('NumIdEst', 'IsNatMarket', 'IsEuroMarket', 
                      'IsRWMarket', 'NewOrders'),
                     Sort = c('IDQual', rep('NonIDQual', 3), 'IDDD'),
                     Class = c(rep('character', 4), 'numeric'),
                     Qual1 = c(rep('', 4), 'NumIdEst'),
                     ValueRegExp = c('[0-9]{9}PP', rep('(0|1| )', 3), '([0-9]{1, 10}| )')
))
Paradt <- new(Class = "DDdt",
            .Data = data.table(
                     Variable = c('NumIdEst', 'Action', 'Date'),
                     Sort = c('IDQual', 'NonIDQual', 'IDDD'),
                     Class = rep('character', 3),
                     Qual1 = c(rep('', 2), 'NumIdEst'),
                     Qual2 = c(rep('', 2), 'Action'),
                     ValueRegExp = c('[0-9]{9}PP', 'Collection|Editing|Imputation', 
                     '(([0-9]{2}-(0[1-9]|1(0-2))-[0-9]{4})| )')
))
Aggdt <- new(Class = "DDdt", 
            .Data = data.table(
                     Variable = c('Province', 'NACE09', 'Turnover'),
                     Sort = c(rep('IDQual', 2), 'IDDD'),
                     Class = c(rep('character', 2), 'numeric'),
                     Qual1 = c(rep('', 2), 'Province'),
                     Qual2 = c(rep('', 2), 'NACE09'),
                     ValueRegExp = c('[0-9]{4}', '([0-4][0-9])|(5[0-2])', '([0-9]{1, 15}| )')
))

DD <- new(Class = 'DD', 
          VarNameCorresp = VNC, 
          ID = IDdt, 
          MicroData = Microdt, 
          ParaData = Paradt,
          Aggregates = Aggdt)


IDDDToUnitNames(VNC, IDDDNames = 'Employees_1_1')

IDDDToUnitNames(DD, IDDDNames = 'Employees_1_1')


}

