% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyToDT.R
\docType{methods}
\name{KeyToDT}
\alias{KeyToDT}
\alias{KeyToDT,rawKey-method}
\title{Return \linkS4class{data.table} out of a set of parsed keys}
\usage{
KeyToDT(key)

\S4method{KeyToDT}{rawKey}(key)
}
\arguments{
\item{key}{key implemented either through the class \linkS4class{rawKey} or 
other.}
}
\value{
\linkS4class{data.table} with one column per each different qualifier
found in the input key and as many rows as components (length) of this input
key.
}
\description{
\code{KeyToDT} returns a \linkS4class{data.table} whose columns
are the different qualifiers parsed in the input key with as many raws as
components (length) of this input key.
}
\examples{
key <- new(Class = 'rawKey', c('IDDD:IASSEmpleo_Norden:391092SS_EsRemuner:1_TipoRemuner:1',
                               'IDDD:IASSCifraNeg_Norden:asd2SS',
                               'IDDD:IASSLPCifraNeg_Norden:1231_CCAA:01'))
KeyToDT(key)

data(ExamplerawKey)
KeyToDT(ExamplerawKey)

}

