% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyToDT.R
\docType{methods}
\name{KeyToDT}
\alias{KeyToDT}
\alias{KeyToDT,rawKey,DD-method}
\title{Return a \linkS4class{data.table} out of a set of parsed keys}
\usage{
KeyToDT(key, DD)

\S4method{KeyToDT}{rawKey,DD}(key, DD)
}
\arguments{
\item{key}{key implemented either through the class \linkS4class{rawKey} or 
other.}

\item{DD}{object DD containing the definition and order of the qualifiers.}
}
\value{
\linkS4class{data.table} with one column per each different qualifier
found in the input key and as many rows as components (length) of this input
key.
}
\description{
\code{KeyToDT} returns a \linkS4class{data.table} whose columns
are the different qualifiers parsed in the input key with as many raws as
components (length) of this input key.
}
\examples{
data(ExamplerawKey)
data(ExampleDD)
KeyToDT(ExamplerawKey, ExampleDD)

}

