% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseUnitName.R
\docType{methods}
\name{ParseUnitName}
\alias{ParseUnitName}
\alias{ParseUnitName,character,RepoTimeInt-method}
\alias{ParseUnitName,character,integer-method}
\title{Produce a vector of file names parsed out of a pattern and a sequence of time period names}
\usage{
ParseUnitName(UnitNames, MetaValues)

\S4method{ParseUnitName}{character,RepoTimeInt}(UnitNames, MetaValues)

\S4method{ParseUnitName}{character,integer}(UnitNames, MetaValues)
}
\arguments{
\item{UnitNames}{Character vector with the names of the files to parse.}

\item{MetaValues}{Object of the appropriate class (\linkS4class{RepoTimeInt}, 
\linkS4class{integer}, ...) with the values of the metacharacters to be parsed.}
}
\value{
Return a vector with each component of the input vector \code{UnitNames} parsed according
 to the input parameter \code{TimePeriods}.
}
\description{
This function parses each component of a vector of file names with a pattern filling
 out the corresponding codes with the input time periods.
}
\examples{
library(RepoTime)
ParseUnitName(c('dat[mm][aa].sas7bdat'), newRepoTime(c('MM122015', 'MM012016', 'MM022016')))
ParseUnitName(c('dat[mm][aa].sas7bdat', 'in[aaaa][mm].txt'), 
              newRepoTime(c('MM122015', 'MM012016', 'MM022016')))
      
}

