% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.DD.R
\docType{methods}
\name{+,DD,DD-method}
\alias{+,DD,DD-method}
\title{Method \code{+} for the class \linkS4class{DD}}
\usage{
\S4method{+}{DD,DD}(e1, e2)
}
\arguments{
\item{e1}{Object of class \linkS4class{DD}.}

\item{e2}{Object of class \linkS4class{DD}.}
}
\value{
Object of class \linkS4class{DD} resulting from integrating both
\linkS4class{DD} objects in a single \linkS4class{DD} object.
}
\description{
\code{+} sums two objects of class \linkS4class{DD}. This method
overloads the operator \link{+} and returns a new object of class
\linkS4class{DD}.
}
\examples{
library(data.table)
VarList1 <- list(ID = new(Class = "VNCdt",
                          .Data = data.table(IDQual = c('ID', rep('', 4)),
                                             NonIDQual = c(rep('', 5)),
                                             IDDD = c('', 'Name', 'Surname', 'PostalAddr', 
                                                      'PhoneNo'),
                                             ID = c('', rep('.', 4)),
                                             Unit1 = c('numidest', 'nombre', 'apellidos', 
                                                'direccion', 'telefono'))),
                 MicroData = new(Class = "VNCdt",
                                 .Data = data.table(IDQual = c('ID', rep('', 4)),
                                                    NonIDQual = c('', 'IsNatMarket', 
                                                                  'IsEuroMarket', 
                                                                  'IsRWMarket', ''),
                                                   IDDD = c(rep('', 4), 'Turnover'),
                                                   ID = c(rep('', 4), '.'),
                                                   IsNatMarket = c(rep('', 4), '0'),
                                                   IsEuroMarket = c(rep('', 4), '0'),
                                                   IsRWMarket = c(rep('', 4), '1'),
                                                   Unit1 = c('numidest', rep('', 3), 'cn05'))),
                Aggregates = new(Class = "VNCdt",
                                 .Data = data.table(IDQual = c('Province', 'NACE', 'IsNatMarket',
                                                               ''),
                                                    NonIDQual = c(rep('', 4)),
                                                    IDDD = c('', '', '', 'Turnover'),
                                                    Province = c('', '', '', '.'),
                                                    NACE = c('', '', '', '.'),
                                                    IsNatMarket = c('', '', '', '1'),
                                                    Unit1 = c('provincia', 'actividad', '', 
                                                              'cn01'))))
VNC1 <- BuildVNC(VarList1)

ID1dt <- new(Class = 'DDdt', 
             .Data = data.table(Variable = c('ID', 'Name', 'Surname', 'PostalAddr', 
                                             'PhoneNo'),
                                Sort = c('IDQual', rep('IDDD', 4)),
                                Class = rep('character', 5),
                                QualOrder = c('1', '', '', '', ''),
                                Qual1 = c('', rep('ID', 4)),
                                ValueRegExp = c('[0-9]{9}PP', '.+', '.+', '.+', 
                                                '(6|9)[0-9]{8}')))
Micro1dt <- new(Class = 'DDdt',
                .Data = data.table(Variable = c('ID', 'IsNatMarket', 'IsEuroMarket', 
                                                'IsRWMarket', 'Turnover'),
                                   Sort = c('IDQual', rep('NonIDQual', 3), 'IDDD'),
                                   Class = c(rep('character', 4), 'numeric'),
                                   QualOrder = c('1', '2', '3', '4', ''),
                                   Qual1 = c(rep('', 4), 'ID'),
                                   ValueRegExp = c('[0-9]{9}PP', rep('(0|1| )', 3), 
                                                   '([0-9]{1, 10}| )')))
Agg1dt <- new(Class = 'DDdt',
              .Data = data.table(Variable = c('Province', 'NACE09', 'Turnover'),
                                 Sort = c(rep('IDQual', 2), 'IDDD'),
                                 Class = c(rep('character', 2), 'numeric'),
                                 QualOrder = c('5', '6', ''),
                                 Qual1 = c(rep('', 2), 'Province'),
                                 Qual2 = c(rep('', 2), 'NACE09'),
                                 ValueRegExp = c('[0-9]{4}', '([0-4][0-9])|(5[0-2])',
                                          '([0-9]{1, 15}| )'))) 

DD1 <- new(Class = 'DD', VarNameCorresp = VNC1, ID = ID1dt, MicroData = Micro1dt, 
           Aggregates = Agg1dt)

VarList2 <- list(ID = new(Class = "VNCdt",
                          .Data = data.table(IDQual = c('ID', rep('', 4)),
                                             NonIDQual = c(rep('', 5)),
                                             IDDD = c('', 'Name', 'Surname', 'PostalAddr', 
                                                      'PhoneNo'),
                                             ID = c('', rep('.', 4)),
                                             Unit1 = c('numidest', 'nombre', 'apellidos', 
                                                   'direccion', 'telefono'))),     
                 MicroData = new(Class = "VNCdt",
                                 .Data =data.table(IDQual = c('ID', rep('', 4)),
                                                   NonIDQual = c('', 'IsNatMarket', 
                                                                 'IsEuroMarket', 
                                                         'IsRWMarket', ''),
                                                   IDDD = c(rep('', 4), 'NewOrders'),
                                                   ID = c(rep('', 4), '.'),
                                                   IsNatMarket = c(rep('', 4), '0'),
                                                   IsEuroMarket = c(rep('', 4), '0'),
                                                   IsRWMarket = c(rep('', 4), '1'),
                                                   Unit1 = c('numidest', rep('', 3), 'cp09'))),
                Aggregates = new(Class = "VNCdt",
                                 .Data =data.table(IDQual = c('Province', 'NACE', 'IsNatMarket', 
                                                              ''),
                                                   NonIDQual = c(rep('', 4)),
                                                   IDDD = c('', '', '', 'NewOrders'),
                                                   Province = c('', '', '', '.'),
                                                   NACE = c('', '', '', '.'),
                                                   IsNatMarket = c('', '', '', '1'),
                                                   Unit1 = c('provincia', 'actividad','', 
                                                             'cp02'))))
VNC2 <- BuildVNC(VarList2)

ID2dt <- new(Class = 'DDdt',
             .Data = data.table(Variable = c('ID', 'Name', 'Surname', 'PostalAddr', 
                                             'PhoneNo'),
                                Sort = c('IDQual', rep('IDDD', 4)),
                                Class = rep('character', 5),
                                QualOrder = c('1', rep('', 4)),
                                Qual1 = c('', rep('ID', 4)),
                                ValueRegExp = c('[0-9]{9}PP', '.+', '.+', '.+', 
                                                '(6|9)[0-9]{8}')))
Micro2dt <- new(Class = 'DDdt',
                .Data = data.table(Variable = c('ID', 'IsNatMarket', 'IsEuroMarket', 
                                                'IsRWMarket', 'NewOrders'),
                                   Sort = c('IDQual', rep('NonIDQual', 3), 'IDDD'),
                                   Class = c(rep('character', 4), 'numeric'),
                                   QualOrder = c('1', '2', '3', '4', ''),
                                   Qual1 = c(rep('', 4), 'NumIdEst'),
                                   ValueRegExp = c('[0-9]{9}PP', rep('(0|1| )', 3), 
                                               '([0-9]{1, 10}| )')))
Agg2dt <- new(Class = 'DDdt',
              .Data = data.table(Variable = c('Province', 'NACE09', 'NewOrders'),
                                 Sort = c(rep('IDQual', 2), 'IDDD'),
                                 Class = c(rep('character', 2), 'numeric'),
                                 QualOrder = c('5', '6', ''),
                                 Qual1 = c(rep('', 2), 'Province'),
                                 Qual2 = c(rep('', 2), 'NACE09'),
                                 ValueRegExp = c('[0-9]{4}', '([0-4][0-9])|(5[0-2])',
                                             '([0-9]{1, 15}| )'))) 

DD2 <- new(Class = 'DD', VarNameCorresp = VNC2, ID = ID2dt, MicroData = Micro2dt,
           Aggregates = Agg2dt)

DD1 + DD2

}

