% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plus.DD.R
\docType{methods}
\name{+,DD,DD-method}
\alias{+,DD,DD-method}
\title{Method \code{+} for the class \linkS4class{DD}}
\usage{
\S4method{+}{DD,DD}(e1, e2)
}
\arguments{
\item{e1}{Object of class \linkS4class{DD}.}

\item{e2}{Object of class \linkS4class{DD}.}
}
\value{
Object of class \linkS4class{DD} resulting from integrating both
\linkS4class{DD} objects in a single \linkS4class{DD} object.
}
\description{
\code{+} sums two objects of class \linkS4class{DD}. This method
overloads the operator \link{+} and returns a new object of class
\linkS4class{DD}.
}
\examples{
library(data.table)
VNC1List <- list(data.table(IDQual = c('Norden', '', ''),
                            NonIDQual = c('', 'CCAA', ''),
                            IDDD = c('', '', 'CifraNeg'),
                            Norden = c('', '', ''),
                            CCAA = c('', '', ''),
                            Unit1 = c('', '', '')))
names(VNC1List) <- seq(along = VNC1List)
VNC1 <- new(Class = 'VarNameCorresp', VarNameCorresp = VNC1List)
DD1 <- new(Class = 'DD',
           MicroData = data.table(Variable = c('NOrden', 'CCAA', 'CifraNeg'),
                             Sort = c('IDQual', 'NonIDQual', 'IDDD'),
                             Class = c('character', 'character', 'numeric'),
                             Qual1 = c('', '', 'NOrden')),
           VarNameCorresp = VNC1)
VNC2List <- list(data.table(IDQual = c('Norden',''), NonIDQual = c('', ''),
                            IDDD = c('', 'logCifraNeg'),
                            Norden = c('', ''),
                            Unit1 = c('', '')))
names(VNC2List) <- seq(along = VNC2List)
VNC2 <- new(Class = 'VarNameCorresp', VarNameCorresp = VNC2List)
DD2 <- new(Class = 'DD',
           MicroData = data.table(Variable = c('Norden', 'logCifraNeg'),
                             Sort = c('IDQual', 'IDDD'),
                             Class = c('character', 'numeric'),
                             Qual1 = c('', 'NOrden')),
           VarNameCorresp = VNC2)
DD1 + DD2
}

