% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.StQ.R
\docType{methods}
\name{+,StQ,StQ-method}
\alias{+,StQ,StQ-method}
\title{Method \code{+} for the class \linkS4class{StQ}}
\usage{
\S4method{+}{StQ,StQ}(e1, e2)
}
\arguments{
\item{e1}{Object of class \code{StQ}.}

\item{e2}{Object of class \code{StQ}.}
}
\value{
Object of class \code{StQ} with the join of both slots \code{DD} and
\code{Data}.
}
\description{
\code{+} joins two objects of class \linkS4class{StQ} in a
single object of the same class.

This method overloads the operator \code{\link{+}} and builds a new
\linkS4class{StQ} object joining both input objects.
}
\examples{
# We build two trivial data sets:
library(data.table)
Data1 <- data.table(NOrden = c('ID1', 'ID2'),
                    IDDD = c('CifraNeg', 'CifraNeg'),
                    Value = c(0, 187364))

Data2 <- data.table(NOrden = c('ID1', 'ID2', 'ID1', 'ID2'),
                    EsRemuner = c(0, 0, 1, 1),
                    IDDD = c('Empleo', 'Empleo', 'Empleo', 'Empleo'),
                    Value = c(0, 1, 23, 41))

# We build two corresponding DD objects:
DDData1 <- data.table(Variable = c('NOrden',
                                   'CifraNeg'),
                     Sort = c('IDQual',
                              'IDDD'),
                     Class = c('character',
                               'numeric'),
                     Qual1 = c('', 'NOrden'))
DD1 <- new(Class = 'DD', Data = DDData1)

DDData2 <- data.table(Variable = c('NOrden',
                                   'EsRemuner',
                                   'Empleo'),
                     Sort = c('IDQual',
                              'NonIDQual',
                              'IDDD'),
                     Class = c('character',
                               'integer',
                               'integer'),
                     Qual1 = c(rep('', 2), 'NOrden'),
                     Qual2 = c(rep('', 2), 'EsRemuner'))
DD2 <- new(Class = 'DD', Data = DDData2)

# We build both StQ objects and join them in a single object:
Q1 <- new(Class = 'StQ', Data = Data1, DD = DD1)
Q2 <- new(Class = 'StQ', Data = Data2, DD = DD2)
Q <- Q1 + Q2
str(Q)

}

