% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.StQ.R
\docType{methods}
\name{+,StQ,StQ-method}
\alias{+,StQ,StQ-method}
\title{Method \code{+} for the class \linkS4class{StQ}}
\usage{
\S4method{+}{StQ,StQ}(e1, e2)
}
\arguments{
\item{e1}{Object of class \code{StQ}.}

\item{e2}{Object of class \code{StQ}.}
}
\value{
Object of class \code{StQ} with the join of both slots \code{DD} and
\code{Data}.
}
\description{
\code{+} joins two objects of class \linkS4class{StQ} in a
single object of the same class.

This method overloads the operator \code{\link{+}} and builds a new
\linkS4class{StQ} object joining both input objects.
}
\examples{
# We build a trivial data set:
library(data.table)

Data1 <- data.table(NumIdEst = c('002', '003', '004'),
                    IDDD = c('NewOrders', 'NewOrders', 'NewOrders'),
                    Value = c(32150, 12574, 23896))
Data1 <- new(Class = 'Datadt', Data1)
VNC1 <- data.table(IDQual = c('NumIdEst', ''),
                   NonIDQual = c('', ''),
                   IDDD = c('', 'NewOrders'),
                   NumIdEst = c('', '.'),
                   Unit1 = c('numidest', 'cp01'))
VNC1 <- new(Class = 'VNCdt', VNC1)
VNC1 <- BuildVNC(list(MicroData = VNC1))

MicroDD1 <- data.table(Variable = c('NumIdEst', 'NewOrders'),
                       Sort = c('IDQual', 'IDDD'),
                       Class = c('character', 'numeric'),
                       Qual1 = c('', 'NumIdEst'),
                       ValueRegExp = c('[0-9]{9}PP', '([0-9]{1, 10}| )'))
MicroDD1 <- new(Class = 'DDdt', MicroDD1)
DD1 <- new(Class = 'DD', VarNameCorresp = VNC1, MicroData = MicroDD1)

# We build the StQ object, and join it with another previously created in a single object:
data(ExampleQ)  
Q1 <- new(Class = 'StQ', Data = Data1, DD = DD1)
Q2 <- ExampleQ
Q <- Q1 + Q2
str(Q)

}

