% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.VarNameCorresp.R
\docType{methods}
\name{+,VarNameCorresp,VarNameCorresp-method}
\alias{+,VarNameCorresp,VarNameCorresp-method}
\title{Method \code{+} for the class \linkS4class{VarNameCorresp}}
\usage{
\S4method{+}{VarNameCorresp,VarNameCorresp}(e1, e2)
}
\arguments{
\item{e1}{Object of class \linkS4class{VarNameCorresp}.}

\item{e2}{Object of class \linkS4class{VarNameCorresp}.}
}
\value{
Object of class \linkS4class{VarNameCorresp} resulting from 
integrating both \linkS4class{VarNameCorresp} objects in a single 
\linkS4class{VarNameCorresp} object.
}
\description{
\code{+} sums two objects of class \linkS4class{VarNameCorresp}.
 This method overloads the operator \link{+} and returns a new object of 
 class \linkS4class{VarNameCorresp}.

The integration is carried out according to the names of the components.
}
\examples{
library(data.table)
VarList1 <- list(MicroData = data.table(IDQual = c('NumIdEst', '', '', '', 
                                                   '', ''),
                                        NonIDQual = c('', 'EsMercNac', 
                                                      'EsMercEuro', 
                                                      'EsMercRM',
                                                      'Cod',''),
                                        IDDD = c('', '', '', '', '', 
                                                 'IEPEntradaPed'),
                                        NumIdEst = c('', '', '', '', '', ''),
                                        EsMercNac = c('', '', '', '', '', '0'),
                                        EsMercEuro = c('', '', '', '', '', '0'),
                                        EsMercRM = c('', '', '', '', '', '1'),
                                        Cod = c('', '', '', '', '', ''),
                                        Unit1 = c('', '', '', '', '', 'cp09')))
VNC1 <- new(Class = 'VarNameCorresp', VarNameCorresp = VarList1)
VarList2 <- list(MicroData = data.table(IDQual = c('NumIdEst', '', '', '', 
                                                   '', ''),
                                        NonIDQual = c('', 'EsMercNac', 
                                                      'EsMercEuro', 
                                                      'EsMercRM',
                                                      'Cod',''),
                                        IDDD = c('', '', '', '', '', 
                                                 'IEPEntradaPed'),
                                        NumIdEst = c('', '', '', '', '', ''),
                                        EsMercNac = c('', '', '', '', '', '0'),
                                        EsMercEuro = c('', '', '', '', '', '1'),
                                        EsMercRM = c('', '', '', '', '', ''),
                                        Cod = c('', '', '', '', '', ''),
                                        Unit1 = c('', '', '', '', '', 'cp08')))
VNC2 <- new(Class = 'VarNameCorresp', VarNameCorresp = VarList2)
VNC1 + VNC2

}

