% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plus.VNCdt.R
\docType{methods}
\name{+,VNCdt,VNCdt-method}
\alias{+,VNCdt,VNCdt-method}
\title{Method \code{+} for the class \linkS4class{VNCdt}}
\usage{
\S4method{+}{VNCdt,VNCdt}(e1, e2)
}
\arguments{
\item{e1}{Object of class \linkS4class{VNCdt}.}

\item{e2}{Object of class \linkS4class{VNCdt}.}
}
\value{
Object of class \linkS4class{VNCdt} resulting from
integrating both \linkS4class{VNCdt} objects in a single
\linkS4class{VNCdt} object.
}
\description{
\code{+} sums two objects of class \linkS4class{VNCdt}.
 This method overloads the operator \link{+} and returns a new object of
 class \linkS4class{VNCdt}.

The integration is carried out according to the names of the components.
}
\examples{
library(data.table)

ID1 <- new(Class = "VNCdt",
          .Data = data.table(IDQual = c('NumIdEst', rep('', 4)),
                             NonIDQual = rep('', 5),
                             IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                             NumIdEst = c('', rep('.', 4)),
                             Unit1 = c('numidest', 'nombre', 'apellidos', 'direccion',
                                       'telefono')))

ID2 <- new(Class = 'VNCdt',
          .Data = data.table(IDQual = c('NumIdEst', rep('', 4)),
                             NonIDQual = c(rep('', 5)),
                             IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                             NumIdEst = c('', rep('.', 4)),
                             Unit1 = c('numidest', 'nombre', 'apellidos', 'direccion',
                                       'telefono')))

ID1 + ID2

MicroData1 <- new(Class = "VNCdt",
                  .Data = data.table(IDQual = c('NumIdEst', rep('', 4)),
                                     NonIDQual = c('', 'IsNatMarket', 'IsEuroMarket',
                                                   'IsRWMarket', ''),
                                     IDDD = c(rep('', 4), 'Turnover'),
                                     NumIdEst = c(rep('', 4), '.'),
                                     IsNatMarket = c(rep('', 4), '0'),
                                     IsEuroMarket = c(rep('', 4), '0'),
                                     IsRWMarket = c(rep('', 4), '1'),
                                     Unit1 = c('numidest', rep('', 3), 'cn05')))

MicroData2 <- new(Class ='VNCdt',
                  .Data = data.table(IDQual = c('NumIdEst', rep('', 4)),
                                     NonIDQual = c('', 'IsNatMarket', 'IsEuroMarket', 'IsNew',
                                                   ''),
                                     IDDD = c(rep('', 4), 'NewOrders'),
                                     NumIdEst = c(rep('', 4), '.'),
                                     IsNatMarket = c(rep('', 4), '0'),
                                     IsEuroMarket = c(rep('', 4), '0'),
                                     IsNew = c(rep('', 4), '1'),
                                     Unit1 = c('numidest', rep('', 3), 'cp09'),
                                     Unit2 = c('norden', rep('', 4))))

MicroData1 + MicroData2

Aggregates1 <- new(Class = "VNCdt",
                   .Data = data.table(IDQual = c('Province', 'NACE', 'IsNatMarket', ''),
                                      NonIDQual = c(rep('', 4)),
                                      IDDD = c('', '', '', 'Turnover'),
                                      Province = c('', '', '', '.'),
                                      NACE = c('', '', '', '.'),
                                      IsNatMarket = c('', '', '', '1'),
                                      Unit1 = c('provincia', 'actividad', '', 'cn01')))


Aggregates2 <- new(Class = 'VNCdt',
                  .Data = data.table(IDQual = c('Province', 'NACE', 'IsNatMarket', ''),
                                     NonIDQual = c('', '', '', ''),
                                     IDDD = c('', '', '', 'NewOrders'),
                                     Province = c('', '', '', '.'),
                                     NACE = c('', '', '', '.'),
                                     IsNatMarket = c('', '', '', '1'),
                                     Unit1 = c('provincia', 'actividad', '', 'cp02')))

Aggregates1 + Aggregates2
}

