% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDatadt-class.R
\docType{class}
\name{rawDatadt-class}
\alias{rawDatadt-class}
\title{S4 class for \linkS4class{data.table}s as components of \linkS4class{rawStQ} objects}
\description{
Definition of an S4 class named \code{rawDatadt} as a subclass of class
\linkS4class{data.table}.

The class \code{rawDatadt} is a \linkS4class{data.table} with the following columns:

\itemize{
 \item \code{IDDDKey}: compund key of class \linkS4class{rawKey} per each value.
 \item \code{Value}: value.
}
}
\examples{

# An empty rawDatadt object:
library(data.table)
new(Class = 'rawDatadt')
               
}

