% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDatadt-class.R
\docType{class}
\name{rawDatadt-class}
\alias{rawDatadt-class}
\title{S4 class for \linkS4class{data.table}s as components of 
\linkS4class{rawStQ} objects}
\description{
Definition of an S4 class named \code{rawDatadt} as a subclass 
of class \linkS4class{data.table}.

The class \code{rawDatadt} is a \linkS4class{data.table} with the following 
columns:

\itemize{
 \item \code{Key}: compund key of class \linkS4class{rawKey} per each value.
 \item \code{Value}: value.
}
}
\examples{
key <- new(Class = 'rawKey', 
          c('IDDD:Turnover_NumIdEst:001_IsNatMarket:0', 
            'IDDD:Turnover_NumIdEst:001_IsNatMarket:1', 
            'IDDD:Turnover_NumIdEst:001_Province:04', 
            'IDDD:Turnover_NumIdEst:001_NACE09:0512'))
Data <- new(Class = 'rawDatadt', 
            data.table(Key = key, Value = c('625000', '23154', '25004', '10512')))
               
}

