% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawKey-class.R
\docType{class}
\name{rawKey-class}
\alias{rawKey-class}
\title{S4 class for the raw key of \linkS4class{rawDatadt} objects.}
\description{
Definition of an S4 class named \code{rawKey} implementing the
key of each key-value pair in a \linkS4class{rawDatadt} object.

The structure of the class \code{rawDatadt} comprises a character vector whose
components follow precisely defined syntax rules given by the regexp 
\code{[A-Za-z]+:[A-Za-z0-9]+(_[A-Za-z]+:[A-Za-z0-9]+)*}.
}
\examples{
# An empty key
new(Class = 'rawKey')

new(Class = 'rawKey', c('IASSEmpleo@391092SS@@1@1',
                        'IASSCifraNeg@asd2SS@@@',
                        'IASSLPCifraNeg@1231@01@@'))
                        
new(Class = 'rawKey', c('IASSEmpleo@391092SS@9834.3@09@@',
                        'IASSCifraNeg@asd2SS@@@@',
                        'IASSLPCifraNeg@1231@@@841@0502'))

}

