% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rawKey-class.R
\docType{class}
\name{rawKey-class}
\alias{rawKey-class}
\title{S4 class for the raw key of \linkS4class{rawDatadt} objects.}
\description{
Definition of an S4 class named \code{rawKey} implementing the
key of each key-value pair in a \linkS4class{rawDatadt} object.

The structure of the class \code{rawDatadt} comprises a character vector whose
components follow precisely defined syntax rules given by the regexp
\code{[A-Za-z]+:[A-Za-z0-9]+(_[A-Za-z]+:[A-Za-z0-9]+)*}.
}
\examples{
# An empty key
new(Class = 'rawKey')

new(Class = 'rawKey', c('IDDD:IASSEmpleo_Norden:391092SS_EsRemuner:1_TipoRemuner:1',
                        'IDDD:IASSCifraNeg_Norden:asd2SS',
                        'IDDD:IASSLPCifraNeg_Norden:1231_CCAA:01'))

new(Class = 'rawKey', c('IDDD:IASSEmpleo_Norden:391092SS_Turnover:9834.3_Province:09',
                        'IDDD:IASSCifraNeg_Norden:asd2SS',
                        'IDDD:IASSLPCifraNeg_Norden:1231_Employees:841_NACE:0502'))
}

