% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawKey-class.R
\docType{class}
\name{rawKey-class}
\alias{rawKey-class}
\title{S4 class for the raw key of \linkS4class{rawStQ} objects}
\description{
Definition of an S4 class named \code{rawKey} implementing the
key of each key-value pair in a \linkS4class{rawStQ} object.

The structure of the class \code{rawStQ} comprises a character vector whose
components follow precisely defined syntax rules given by the regexp 
\code{[A-Za-z]+:[A-Za-z0-9]+(_[A-Za-z]+:[A-Za-z0-9]+)*}.
}
\section{Slots}{

\describe{
\item{\code{Data}}{Object of class \linkS4class{data.table} with key-value pair
structure. It must have exactly two columns: \code{Key} and \code{Value}.
It contains all statistical variables (including some metadata) together with
their corresponding values. If \code{Data} is not specified as an input
parameter, an empty \linkS4class{data.table} with columns \code{Key} and
\code{Value} will be initiated.}

\item{\code{DD}}{Object of class \linkS4class{DD} with the definition and properties
of all variables. If \code{DD} is not specified as an input parameter, an
empty \linkS4class{DD} object with columns \code{Variable}, \code{Sort},
\code{Class} and \code{Qual1} will be initiated.}
}}
\examples{
# An empty key
new(Class = 'rawKey')

new(Class = 'rawKey', c('ID:391092SS_Turnover:9834.3_Province:09',
                        'ID:asd2SS',
                        'ID:1231_Employees:841_NACE:0502'))
                        
new(Class = 'rawKey', c('ID:391092SS.Turnover:9834.3_Province:09',
                        'ID:asd2-SS',
                        'ID:1231 Employees:841 NACE:0502'))

}

