% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawKey-class.R
\docType{class}
\name{rawKey-class}
\alias{rawKey-class}
\title{S4 class for the raw key of \linkS4class{rawDatadt} objects.}
\description{
Definition of an S4 class named \code{rawKey} implementing the
key of each key-value pair in a \linkS4class{rawDatadt} object.

The structure of the class \code{rawDatadt} comprises a character vector whose
components follow precisely defined syntax rules given by the regexp 
\code{[A-Za-z]+:[A-Za-z0-9]+(_[A-Za-z]+:[A-Za-z0-9]+)*}.
}
\examples{
# An empty key
library(data.table)
new(Class = 'rawKey')

# An example:
library(data.table)
key <- new(Class = 'rawKey', 
           data.table(IDDDKey = c('Employees', 'Employees', 'RemEmployees', 'Turnover'),
                      QualKey = c('25641378SS2.1.1.', '25641378SS1.    ', '25641378SS    1.', 
                                  '25641378SS')))
                                  
}

