% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawStQ.R
\name{rawStQ}
\alias{rawStQ}
\title{S4 class for sets of \emph{raw} \emph{St}andarized \emph{Q}uestionnaires}
\usage{
rawStQ(rawData = data.table(IDDDKey = character(0), QualKey = character(0),
  Value = character(0)), DD = BuildDD())
}
\description{
Definition of an S4 class named \code{rawStQ} for sets of raw standardized 
questionnaires.

The structure of the class \code{rawStQ} comprises 2 attributes:
\itemize{
\item The attribute \code{rawData}, which is an object of class \linkS4class{data.table} with 
key-value pair structure containing all statistical variables, both from the questionnaire and 
any resulting metadata from the data processing.

\item The attribute \code{DD}, which is an object of class \linkS4class{DD}. It basically 
contains the definition and properties of each variable.
}
}
\section{Slots}{

\describe{
\item{\code{rawData}}{Object of class \linkS4class{data.table} with key-value pair structure. It must 
have exactly three columns: \code{IDDDKey},  \code{QualKey} and \code{Value}. It contains all 
statistical variables (including some metadata) together with their corresponding values. If 
\code{rawData} is not specified as an input parameter, an empty \linkS4class{data.table} object 
with columns \code{IDDDKey}, \code{QualKey} and \code{Value} will be initiated.}

\item{\code{DD}}{Object of class \linkS4class{DD} with the definition and properties of all variables. If
\code{DD} is not specified as an input parameter, an empty \linkS4class{DD} object with all its
components  will be initiated.}
}}
\examples{
# An empty standardized questionnaire set:
rawStQ()

library(data.table)
data(ExampleDD)
rawData <- data.table(IDDDKey = c('Employees', 'Employees', 'RemEmployees', 'Turnover'),
                      QualKey = c('25641378SS2.1.1.', '25641378SS1.    ', '25641378SS    1.', '25641378SS'),
                      Value = c('625', '954', '122', '105124'))
rawStQ(rawData, ExampleDD)

}

