% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawStQToStQ.R
\docType{methods}
\name{rawStQToStQ}
\alias{rawStQToStQ}
\alias{rawStQToStQ,rawStQ-method}
\title{Generates an object of class \linkS4class{StQ} from the input
\linkS4class{rawStQ} object.}
\usage{
rawStQToStQ(rawQ)

\S4method{rawStQToStQ}{rawStQ}(rawQ)
}
\arguments{
\item{rawQ}{Object of class \linkS4class{rawStQ} whose slot Data has the
key-value pair structure.}
}
\value{
Object of class \linkS4class{StQ} whose slot Data has at least the 
columns \code{IDDD} and \code{Value}.
}
\description{
\code{rawStQToStQ} returns an object of class \linkS4class{StQ}
from the input\linkS4class{rawStQ} object.
}
\examples{
data(ExampleDD)
key <- new(Class = 'rawKey', 
           c('IDDD:Turnover_NumIdEst:001_IsNatMarket:0', 
             'IDDD:Turnover_NumIdEst:001_IsNatMarket:1', 
             'IDDD:Turnover_NumIdEst:001_Province:04', 
             'IDDD:Turnover_NumIdEst:001_NACE09:0512'))
rawData <- new(Class = 'rawDatadt', 
               data.table(Key = key, Value = c('625000', '23154', '25004', '10512')))
rawQ <- new(Class = 'rawStQ', Data = Data, DD = ExampleDD)

newQ <- rawStQToStQ(rawQ)
}

