% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawStQToStQ.R
\name{rawStQToStQ}
\alias{rawStQToStQ}
\alias{rawStQToStQ,rawStQ-method}
\alias{rawStQToStQ,rawStQList-method}
\title{Generates an object of class \link{StQ} from the input \linkS4class{rawStQ} object.}
\usage{
rawStQToStQ(rawQ)

\S4method{rawStQToStQ}{rawStQ}(rawQ)

\S4method{rawStQToStQ}{rawStQList}(rawQ)
}
\arguments{
\item{rawQ}{Object of class \linkS4class{rawStQ} whose slot Data has the  key-value pair
structure.}
}
\value{
Object of class \link{StQ} whose slot Data has at least the columns \code{IDDD}
and \code{Value}.
}
\description{
\code{rawStQToStQ} returns an object of class \link{StQ} from the input
\linkS4class{rawStQ} object.
}
\examples{
data(ExamplerawStQ)
StQ <- rawStQToStQ(ExamplerawStQ)
str(StQ)

}
