% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renVar.R
\docType{methods}
\name{renVar}
\alias{renVar}
\alias{renVar,StQ,character,character-method}
\title{Rename statistical variable names (IDDD) in an object}
\usage{
renVar(object, VarNames, NewVarNames)

\S4method{renVar}{StQ,character,character}(object, VarNames, NewVarNames)
}
\arguments{
\item{object}{Input object of class \linkS4class{StQ}.}

\item{VarNames}{Character vector with the variable names to be renamed.}

\item{NewVarNames}{Character vector with the new variable names.}
}
\value{
Object with the same class as the input object and all specified variable names duly
renamed in all slots.
}
\description{
\code{renVar} rename variables in a object keeping the rest of the object intact.

This method returns the input object with the specified variable names in the input parameter
\code{VarNames} renamed with the names in the input parameter \code{NewVarNames}. Variables are
renamed in all slots. Old names are dropped out.
}
\examples{
data(ExampleStQ)
renVar(ExampleStQ, 'Stocks', 'NewStocks')

}
