% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAggr.R
\docType{methods}
\name{setAggr<-}
\alias{setAggr<-}
\alias{setAggr<-,DD,DDdt-method}
\alias{setAggr<-,StQ,DDdt-method}
\title{Set value of slot \code{Aggregates}.}
\usage{
setAggr(object) <- value

\S4method{setAggr}{DD,DDdt}(object) <- value

\S4method{setAggr}{StQ,DDdt}(object) <- value
}
\arguments{
\item{object}{Object whose slot \code{Aggregates} is to be assigned.}

\item{value}{\linkS4class{DDdt} to be assigned to the slot \code{Aggregates}.}
}
\value{
Object with slot Aggregates updated.
}
\description{
\code{setAggr} assigns a \linkS4class{DDdt} to the slot
\code{Aggregates} of the input object.
}
\examples{
# An example:
library(data.table)
MicroDataDD <- data.table(Variable = 'IEPEntradaPed', Sort = 'IDDD', Class = 'numeric',
                          Length = '8',
                          Qual1 = 'NumIdEst', Qual2 = 'Market', ValueRegExp = '')
MicroDataDD <- new(Class = 'DDdt', MicroDataDD)
VarList <- list(MicroData = new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst','','',''),
                                                          NonIDQual = c('', 'Market', 'Cod', ''),
                                                          IDDD = c('', '', '' ,'IEPEntradaPed'),
                                                          NumIdEst = c('', '', '', '.'),
                                                          Market = c('', '', '', '1.'),
                                                          Cod = rep('', 4),
                                                          UnitName = c('', '', '', 'cp02'),
                                                          InFiles = rep('FF', 4))))
VNC <- BuildVNC(VarList)
DD <- new(Class = 'DD', VarNameCorresp = VNC, MicroData = MicroDataDD)
          
Aggdt <- data.table(Variable = c('NACE09', 'Turnover'),
                    Sort = c('IDQual', 'IDDD'),
                    Class = c('character', 'character'),
                    Length = c('4', '10'),
                    Qual1 = c('', 'NACE09'),
                    ValueRegExp = c('([0-4][0-9])|(5[0-2])', '([0-9]{1, 15}| )'))  
Aggdt <- new(Class = 'DDdt', Aggdt)          
setAggr(DD) <- Aggdt
DD

}

