% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAggWeights.R
\docType{methods}
\name{setAggWeights<-}
\alias{setAggWeights<-}
\alias{setAggWeights<-,DD,DDdt-method}
\alias{setAggWeights<-,StQ,DDdt-method}
\title{Set value of slot \code{AggWeights}.}
\usage{
setAggWeights(object) <- value

\S4method{setAggWeights}{DD,DDdt}(object) <- value

\S4method{setAggWeights}{StQ,DDdt}(object) <- value
}
\arguments{
\item{object}{Object whose slot \code{AggWeights} is to be assigned.}

\item{value}{\linkS4class{DDdt} to be assigned to the slot \code{AggWeights}.}
}
\value{
Object with slot AggWeights updated.
}
\description{
\code{setAggrWeights} assigns a \linkS4class{DDdt} to the slot
\code{AggrWeights} of the input object.
}
\examples{
library(data.table)
MicroDataDD <- data.table(Variable = 'IEPEntradaPed', Sort = 'IDDD', Class = 'numeric',
                          QualOrder = c(''),
                          Qual1 = 'NumIdEst', Qual2 = 'EsMercNac', Qual3 = 'EsMercEuro',
                          Qual4 = 'EsMercRM', ValueRegExp = '')
MicroDataDD <- new(Class = 'DDdt', MicroDataDD)
VarList <- list(MicroData = new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst','','','',''),
                                                            NonIDQual = c('EsMercNac', 
                                                                          'EsMercEuro', 
                                                                          'EsMercRM', 'Cod', ''),
                                                            IDDD = c('', '', '' ,'' ,
                                                                     'IEPEntradaPed'),
                                                            NumIdEst = c('', '', '', '', '.'),
                                                            EsMercNac = c('', '', '', '', '0'),
                                                            EsMercEuro = c('', '', '', '', '0'),
                                                            EsMercRM = c('', '', '', '', '1'),
                                                            Cod = rep('', 5),
                                                            Unit1 = c('', '', '', '', 'cp02'))))
VNC <- BuildVNC(VarList)
DD <- new(Class = 'DD', VarNameCorresp = VNC, MicroData = MicroDataDD)

AggWeights <- data.table(Variable = c('ID', 'Pond1'), Sort = c('IDQual', 'IDDD'),
                         Class = c('character', 'character'),
                         QualOrder = c('1', ''),
                         Qual1 = c('', 'ID'), ValueRegExp = c('', ''))
AggWeights <- new(Class = 'DDdt', AggWeights)

setAggWeights(DD) <- AggWeights
DD

}

