% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setData.R
\docType{methods}
\name{setData<-}
\alias{setData<-}
\alias{setData<-,StQ,data.table-method}
\title{Set value of slot \code{Data}}
\usage{
setData(object) <- value

\S4method{setData}{StQ,data.table}(object) <- value
}
\arguments{
\item{object}{Object whose slot \code{Data} is to be assigned.}

\item{value}{\linkS4class{data.table} to be assigned to the slot \code{Data}.}
}
\value{
Object with slot Data updated.
}
\description{
\code{setData} assigns a \linkS4class{data.table} to the slot
\code{Data} of the input object.
}
\examples{
# We build an empty data.table:
library(data.table)
Data <- data.table(IDDD = character(0), Value = numeric(0))

# We assign this data.table to the slot Data of object ExampleQ:
data(ExampleQ)
setData(ExampleQ) <- Data
ExampleQ

}

