% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setData.R
\docType{methods}
\name{setData<-}
\alias{setData<-}
\alias{setData<-,StQ,Datadt-method}
\title{Set value of slot \code{Data}.}
\usage{
setData(object) <- value

\S4method{setData}{StQ,Datadt}(object) <- value
}
\arguments{
\item{object}{Object whose slot \code{Data} is to be assigned.}

\item{value}{\linkS4class{Datadt} to be assigned to the slot \code{Data}.}
}
\value{
Object with slot Data updated.
}
\description{
\code{setData} assigns a \linkS4class{Datadt} to the slot
\code{Data} of the input object.
}
\examples{
# We build an empty data.table:
library(data.table)
Data <- new(Class = 'Datadt', data.table(IDDD = character(0), Value = numeric(0)))

# We assign this data.table to the slot Data of object NewExampleQ:
data(ExampleQ)
NewExampleQ <- ExampleQ
setData(NewExampleQ) <- Data
NewExampleQ

}

