% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDD.R
\docType{methods}
\name{setDD<-}
\alias{setDD<-}
\alias{setDD<-,StQ,DD-method}
\title{Set value of slot \code{DD}}
\usage{
setDD(object) <- value

\S4method{setDD}{StQ,DD}(object) <- value
}
\arguments{
\item{object}{Object whose slot \code{DD} is to be set.}

\item{value}{Object of class \link{DD} to be assigned to the input object.}
}
\value{
Object of class \link{StQ} with slot DD updated.
}
\description{
\code{setDD} assigns an object of class \link{DD} to the slot \code{DD} of 
the input object.
}
\examples{
# We build an empty StQ object:
library(data.table)
Q <- StQ()

# We use an DD object previously created:
data(ExampleDD)
# We assign this DD object to slot DD of the preceding StQ object:
setDD(Q) <- ExampleDD
str(Q)

}
