% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setID.R
\docType{methods}
\name{setID<-}
\alias{setID<-}
\alias{setID<-,DD,DDdt-method}
\alias{setID<-,StQ,DDdt-method}
\title{Set value of slot \code{ID} of an object \linkS4class{DD}}
\usage{
setID(object) <- value

\S4method{setID}{DD,DDdt}(object) <- value

\S4method{setID}{StQ,DDdt}(object) <- value
}
\arguments{
\item{object}{Object containing slot \code{ID} to be assigned.}

\item{value}{\linkS4class{DDdt} to be assigned to the slot \code{ID}.}
}
\value{
Object \linkS4class{DD} with slot \code{ID} updated.
}
\description{
\code{setID} assigns a \linkS4class{DDdt} to the
 slot \code{ID} of the input object.
}
\examples{
# An example:
library(data.table)
MicroDataDD <- data.table(Variable = 'IEPEntradaPed', Sort = 'IDDD', Class = 'numeric',
                          Qual1 = 'NumIdEst', Qual2 = 'EsMercNac', Qual3 = 'EsMercEuro',
                          Qual4 = 'EsMercRM', ValueRegExp = '')
MicroDataDD <- new(Class = 'DDdt', MicroDataDD)
VarList <- list(MicroData = new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst','','','',''),
                                                            NonIDQual = c('EsMercNac',
                                                                          'EsMercEuro',
                                                                          'EsMercRM', 'Cod', ''),
                                                            IDDD = c('', '', '' ,'' ,
                                                                     'IEPEntradaPed'),
                                                            NumIdEst = c('', '', '', '', '.'),
                                                            EsMercNac = c('', '', '', '', '0'),
                                                            EsMercEuro = c('', '', '', '', '0'),
                                                            EsMercRM = c('', '', '', '', '1'),
                                                            Cod = rep('', 5),
                                                            Unit1 = c('', '', '', '', 'cp02'))))
VNC <- BuildVNC(VarList)
DD <- new(Class = 'DD', VarNameCorresp = VNC, MicroData = MicroDataDD)

IDdt <- data.table(Variable = c('NumIdEst', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                   Sort = c('IDQual', rep('IDDD', 4)),
                   Class = rep('character', 5),
                   Qual1 = c('', rep('NumIdEst', 4)),
                   ValueRegExp = c('[0-9]{9}PP', '.+', '.+', '.+', '(6|9)[0-9]{8}'))
IDdt <- new(Class = 'DDdt', IDdt)

setID(DD) <- IDdt
DD
}

