% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setID.R
\docType{methods}
\name{setID<-}
\alias{setID<-}
\alias{setID<-,VNC,data.table-method}
\alias{setID<-,DD,data.table-method}
\title{Set value of slot \code{ID} of an object \linkS4class{DD}}
\usage{
setID(object) <- value

\S4method{setID}{VNC,data.table}(object) <- value

\S4method{setID}{DD,data.table}(object) <- value
}
\arguments{
\item{object}{Object containing slot \code{ID} to be assigned.}

\item{value}{\linkS4class{data.table} to be assigned to the slot \code{ID}.}
}
\value{
Object \linkS4class{DD} with slot \code{ID} updated.
}
\description{
\code{setID} assigns a \linkS4class{data.table} to the slot \code{ID} of the input
object.
}
\examples{
library(data.table)
MicroDataDD <- data.table(Variable = 'NewOrders', Sort = 'IDDD', Class = 'numeric',
                          Length = '8',
                          Qual1 = 'NumIdEst', Qual2 = 'Market', ValueRegExp = '')
VarList <- list(MicroData = data.table(IDQual = c('NumIdEst','','',''),
                                       NonIDQual = c('', 'Market', 'Cod', ''),
                                       IDDD = c('', '', '' ,'NewOrders'),
                                       NumIdEst = c('', '', '', '.'),
                                       Market = c('', '', '', '1.'),
                                       Cod = rep('', 4),
                                       UnitName = c('', '', '', 'cp02'),
                                       InFiles = rep('FF', 4)))
VNC <- BuildVNC(VarList)
DD <- DD(VNC = VNC, MicroData = MicroDataDD)

IDdt <- data.table(Variable = c('NumIdEst', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                   Sort = c('IDQual', rep('IDDD', 4)),
                   Class = rep('character', 5),
                   Length = c('11', '20', '20', '20', '9'),
                   Qual1 = c('', rep('NumIdEst', 4)),
                   ValueRegExp = c('[0-9]{9}PP', '.+', '.+', '.+', '(6|9)[0-9]{8}'))
setID(DD) <- IDdt
DD

}
