% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMicroData.R
\docType{methods}
\name{setMicroData<-}
\alias{setMicroData<-}
\alias{setMicroData<-,DD,data.table-method}
\alias{setMicroData<-,VNC,data.table-method}
\title{Set value of slot \code{MicroData} of an object \linkS4class{DD}}
\usage{
setMicroData(object) <- value

\S4method{setMicroData}{VNC,data.table}(object) <- value

\S4method{setMicroData}{DD,data.table}(object) <- value
}
\arguments{
\item{object}{Object containing slot \code{ID} to be assigned.}

\item{value}{\linkS4class{data.table} to be assigned to the slot \code{ID}.}
}
\value{
Object \linkS4class{DD} with slot \code{ID} updated.
}
\description{
\code{setMicroData} assigns a \linkS4class{data.table} to the slot \code{ID} of the  
input object.
}
\examples{
library(data.table)
MicroDataDD <- data.table(Variable = 'NewOrders', Sort = 'IDDD', Class = 'numeric',
                          Length = '8',
                          Qual1 = 'NumIdEst', Qual2 = 'Market', ValueRegExp = '')
VarList <- list(MicroData = data.table(IDQual = c('NumIdEst','','',''),
                                       NonIDQual = c('', 'Market', 'Cod', ''),
                                       IDDD = c('', '', '' , 'NewOrders'),
                                       NumIdEst = c('', '', '', '.'),
                                       Market = c('', '', '', '1.'),
                                       Cod = rep('', 4),
                                       UnitName = c('', '', '', 'cp02'),
                                       InFiles = rep('FF', 4)))
VNC <- BuildVNC(VarList)
DD <- DD(VNC = VNC, MicroData = MicroDataDD)

Microdt <- data.table(Variable = c('NumIdEst', 'NewOrders', 'Turnover'),
                      Sort = c('IDQual', 'IDDD', 'IDDD'),
                      Class = c('character', 'numeric', 'numeric'),
                      Length = c('11', '8', '8'),
                      Qual1 = c('', 'NumIdEst', 'NumIdEst'),
                      ValueRegExp = c('[0-9]{9}PP', '([0-9]{1, 10}| )', '([0-9]{1, 10}| )'))
setMicroData(DD) <- Microdt
DD

}

