% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOtherDD.R
\docType{methods}
\name{setOtherDD<-}
\alias{setOtherDD<-}
\alias{setOtherDD<-,DD,DDdt-method}
\alias{setOtherDD<-,StQ,DDdt-method}
\title{Set value of slot \code{Other}}
\usage{
setOtherDD(object) <- value

\S4method{setOtherDD}{DD,DDdt}(object) <- value

\S4method{setOtherDD}{StQ,DDdt}(object) <- value
}
\arguments{
\item{object}{Object whose slot \code{Other} is to be assigned.}

\item{value}{\linkS4class{DDdt} to be assigned to the slot \code{Other}.}
}
\value{
Object with slot \code{Other} updated.
}
\description{
\code{setOtherDD} assigns a \linkS4class{DDdt} to the slot \code{Other} of the input
 object.
}
\examples{
library(data.table)
MicroDataDD <- data.table(Variable = 'NewOrders', Sort = 'IDDD', Class = 'numeric',
                          Length = '8',
                          Qual1 = 'NumIdEst', Qual2 = 'Market', ValueRegExp = '')
MicroDataDD <- new(Class = 'DDdt', MicroDataDD)
VarList <- list(MicroData = new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst','','',''),
                                                          NonIDQual = c('', 'Market', 'Cod', ''),
                                                          IDDD = c('', '', '' ,'NewOrders'),
                                                          NumIdEst = c('', '', '', '.'),
                                                          Market = c('', '', '', '1.'),
                                                          Cod = rep('', 4),
                                                          UnitName = c('', '', '', 'cp02'),
                                                          InFiles = rep('FF', 4))))
VNC <- BuildVNC(VarList)
DD <- new(Class = 'DD', VarNameCorresp = VNC, MicroData = MicroDataDD)
          
Otherdt <- data.table(Variable = c('NACE08'), Sort = c('NonIDQual'), Class = c('character'),
                      Length = '4',
                      Qual1 = c(''), ValueRegExp = c('([0-4][0-9])|(5[0-2])'))
Otherdt <- new(Class = 'DDdt', Otherdt)             

setOtherDD(DD) <- Otherdt
DD

}

