% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setParaData.R
\docType{methods}
\name{setParaData<-}
\alias{setParaData<-}
\alias{setParaData<-,DD,data.table-method}
\title{Set value of slot \code{ParaData}}
\usage{
setParaData(object) <- value

\S4method{setParaData}{DD,data.table}(object) <- value
}
\arguments{
\item{object}{Object whose slot \code{ParaData} is to be assigned.}

\item{value}{\linkS4class{data.table} to be assigned to the slot \code{ParaData}.}
}
\value{
Object with slot ParaData updated.
}
\description{
\code{setParaData} assigns a \linkS4class{data.table} to the slot
\code{ParaData} of the input object.
}
\examples{
# We build an empty data.table:
library(data.table)
ParaData <- data.table(IDDD = character(0), Value = numeric(0))

# We assign this data.table to the slot ParaData of object NewExampleDD:
data(ExampleDD)
NewExampleDD <- ExampleDD
setParaData(NewExampleDD) <- ParaData
NewExampleDD

}

