% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setParaData.R
\docType{methods}
\name{setParaData<-}
\alias{setParaData<-}
\alias{setParaData<-,DD,DDdt-method}
\alias{setParaData<-,StQ,DDdt-method}
\title{Set value of slot \code{ParaData}.}
\usage{
setParaData(object) <- value

\S4method{setParaData}{DD,DDdt}(object) <- value

\S4method{setParaData}{StQ,DDdt}(object) <- value
}
\arguments{
\item{object}{Object whose slot \code{ParaData} is to be assigned.}

\item{value}{\linkS4class{DDdt} to be assigned to the slot \code{ParaData}.}
}
\value{
Object with slot ParaData updated.
}
\description{
\code{setParaData} assigns a \linkS4class{DDdt} to the slot
\code{ParaData} of the input object.
}
\examples{
# An example:
library(data.table)
MicroDataDD <- data.table(Variable = 'IEPEntradaPed', Sort = 'IDDD', Class = 'numeric',
                          Length = '8',
                          Qual1 = 'NumIdEst', Qual2 = 'Market', ValueRegExp = '')
MicroDataDD <- new(Class = 'DDdt', MicroDataDD)
VarList <- list(MicroData = new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst','','',''),
                                                          NonIDQual = c('', 'Market', 'Cod', ''),
                                                          IDDD = c('', '', '' ,'IEPEntradaPed'),
                                                          NumIdEst = c('', '', '', '.'),
                                                          Market = c('', '', '', '1.'),
                                                          Cod = rep('', 4),
                                                          UnitName = c('', '', '', 'cp02'))))
VNC <- BuildVNC(VarList)
DD <- new(Class = 'DD', VarNameCorresp = VNC, MicroData = MicroDataDD)
          
Paradt <- data.table(
             Variable = c('NumIdEst', 'Date'),
             Sort = c('IDQual', 'IDDD'),
             Class = rep('character', 2),
             Length = c('11', '8'),
             Qual1 = c('', 'NumIdEst'),
             ValueRegExp = c('[0-9]{9}PP', 
                             '(([0-9]{2}-(0[1-9]|1(0-2))-[0-9]{4})| )'))       
Paradt <- new(Class = 'DDdt', Paradt)          
setParaData(DD) <- Paradt
DD

}

