% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUnits.R
\docType{methods}
\name{setUnits<-}
\alias{setUnits<-}
\alias{setUnits<-,StQ,data.table-method}
\title{Restrict slot \code{Data} of the input object to specified units}
\usage{
setUnits(object) <- value

\S4method{setUnits}{StQ,data.table}(object) <- value
}
\arguments{
\item{object}{Objeto whose slot \code{Data} is to be restricted.}

\item{value}{\linkS4class{data.table} with the values of the unit qualifiers
identifying the restricted units.}
}
\value{
Object with the same class as the input object with slot Data
restricted to the specified units.
}
\description{
\code{setUnits} restricts slot \code{Data} to those units
specified as input parameter.
}
\examples{
library(data.table)
data(ExampleQ)
NewExampleQ <- ExampleQ
Units <- data.table(NOrden = 1:5)
setUnits(NewExampleQ) <- Units
str(NewExampleQ)

}

