% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setUnits.R
\docType{methods}
\name{setUnits<-}
\alias{setUnits<-}
\alias{setUnits<-,StQ,data.table-method}
\title{Return an object as the input object but with slot \code{Data} restricted to the specified
units.}
\usage{
setUnits(object) <- value

\S4method{setUnits}{StQ,data.table}(object) <- value
}
\arguments{
\item{object}{Objeto whose slot \code{Data} is to be restricted.}

\item{value}{\linkS4class{data.table} with the values of the unit qualifiers identifying the
restricted units.}
}
\value{
Object with the same class as the input object with slot Data restricted to the specified
units.
}
\description{
\code{setUnits} returns an object as the input object but with slot \code{Data}
restricted to the specified units.
}
\examples{
library(data.table)
library(stringr)
data(ExampleStQ)
NewExampleStQ <- ExampleStQ
Units <- data.table(ID = str_pad(1:5, 5, 'left', '0'))
setUnits(NewExampleStQ) <- Units
NewExampleStQ
}

