% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setVar.R
\docType{methods}
\name{setVar}
\alias{setVar}
\alias{setVar,StQ,character,ANY,DD-method}
\title{Method to set values of a variable}
\usage{
setVar(object, VarName, Value, DDnl, lag = NULL, by = NULL)

\S4method{setVar}{StQ,character,ANY,DD}(object, VarName, Value, DDnl,
  lag = NULL, by = NULL)
}
\arguments{
\item{object}{Object in which to include the new variable.}

\item{VarName}{Character vector of length 1 with the name of the new
variable.}

\item{Value}{Vector of length the number of statistical units in the input
object with the values of the new variable for each of these units or
alternatively an object of class \code{\link{expression}} with a mathematical
 formula for the computation of these values.}

\item{DDnl}{\linkS4class{DD} object with a row of columns \code{Variable},
\code{Sort}, \code{Class} and \code{Qual1} to \code{Qualn} with the
information of the new variable needed for the slot \code{DD} of the input
object.}

\item{lag}{Integer vector of length 1 with the lag to use when the input
object is of class \linkS4class{StQList}.}

\item{by}{Character vector with the names of the variables specifying those
statistical units groups under which the mathematical formula will be
applied.}
}
\value{
Object of the same class as the input object with the new variable
included.
}
\description{
\code{setVar} sets the values of a variable in the input object.

This method creates a variable with the name \code{VarName} with values
specified in the input \code{Value}. If the variable is already present in
the input variable, the method overwrites it.

It is also necessary to provide as input parameter an object of class
\linkS4class{DD} named \code{DDnl} with the row corresponding to the variable
set with this method to be included in the slot \code{DD}. This row is
specified as a data.table with a single row and columns Variable, Sort, Class
and Qual1-Qualn (as many as so many qualifiers are necessary).

Input parameters \code{by} and \code{lag} are optional and specify how to
compute the values of the new variable in some circunstamces.
}
\examples{
library(data.table)
data(ExampleQ)
newDD <- new(Class = 'DD', Data = data.table(Variable = 'lCN',
                                             Sort = 'IDDD',
                                             Class = 'numeric',
                                             Qual1 = 'NOrden'))
NewQ <- setVar(object = ExampleQ, VarName = 'lCN',
               Value = expression(log(1 + IASSCifraNeg)), DDnl = newDD)
getVar(NewQ, 'lCN')

}

