% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setVNC.R
\docType{methods}
\name{setVNC<-}
\alias{setVNC<-}
\alias{setVNC<-,DD,list-method}
\alias{setVNC<-,StQ,list-method}
\alias{setVNC<-,VarNameCorresp,list-method}
\title{Set value of slot \code{VarNameCorresp}}
\usage{
setVNC(object) <- value

\S4method{setVNC}{VarNameCorresp,list}(object) <- value

\S4method{setVNC}{DD,list}(object) <- value

\S4method{setVNC}{StQ,list}(object) <- value
}
\arguments{
\item{object}{Object containing slot \code{VarNameCorresp} to be assigned.}

\item{value}{List of \linkS4class{data.table}s to be assigned to the slot
\code{VarNameCorresp}.}
}
\value{
Object with slot \code{VarNameCorresp} updated.
}
\description{
\code{setVNC} assigns a list of \linkS4class{data.table}s to the
 slot \code{VarNameCorresp} of the input object.
}
\examples{
# On an empty object of class VarNameCorresp:
VNC <- new(Class = 'VarNameCorresp')
VarList <- list(MicroData = data.table(IDQual = c('NumIdEst','','','',''),
                                       NonIDQual = c('EsMercNac',
                                                     'EsMercEuro',
                                                     'EsMercRM',
                                                     'Cod',
                                                     ''),
                                       IDDD = c('', '', '' ,'' ,
                                                'IEPEntradaPed'),
                                       Unit1 = c('', '', '', '', 'cp02')))
setVNC(VNC) <- VarList
VNC
str(VNC)

# On an object of class DD:
MicroDataDD <- data.table(Variable = 'IEPEntradaPed',
                          Sort = 'IDDD',
                          Class = 'numeric',
                          Qual1 = 'NOrden')
DD <- new(Class = 'DD', MicroData = MicroDataDD)
VarList <- list(MicroData = data.table(IDQual = c('NumIdEst','','','',''),
                                       NonIDQual = c('EsMercNac',
                                                     'EsMercEuro',
                                                     'EsMercRM',
                                                     'Cod',
                                                     ''),
                                       IDDD = c('', '', '' ,'' ,
                                                'IEPEntradaPed'),
                                       Unit1 = c('', '', '', '', 'cp02')))
setVNC(DD) <- VarList
DD
str(DD)

# On an object of class StQ:
MicroDataDD <- data.table(Variable = 'IEPEntradaPed',
                          Sort = 'IDDD',
                          Class = 'numeric',
                          Qual1 = 'NOrden')
DD <- new(Class = 'DD', MicroData = MicroDataDD)
StQ <- new(Class = 'StQ',
           Data = data.table(IDDD = character(0), Value = character(0)),
           DD = DD)
VarList <- list(MicroData = data.table(IDQual = c('NumIdEst','','','',''),
                                       NonIDQual = c('EsMercNac',
                                                     'EsMercEuro',
                                                     'EsMercRM',
                                                     'Cod',
                                                     ''),
                                       IDDD = c('', '', '' ,'' ,
                                                'IEPEntradaPed'),
                                       Unit1 = c('', '', '', '', 'cp02')))
setVNC(StQ) <- VarList
StQ
str(StQ)
}

