% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setVNC.R
\docType{methods}
\name{setVNC<-}
\alias{setVNC<-}
\alias{setVNC<-,DD,VarNameCorresp-method}
\alias{setVNC<-,StQ,VarNameCorresp-method}
\title{Set value of slot \code{VarNameCorresp}}
\usage{
setVNC(object) <- value

\S4method{setVNC}{DD,VarNameCorresp}(object) <- value

\S4method{setVNC}{StQ,VarNameCorresp}(object) <- value
}
\arguments{
\item{object}{Object containing slot \code{VarNameCorresp} to be assigned.}

\item{value}{List of \linkS4class{data.table}s to be assigned to the slot 
\code{VarNameCorresp}.}
}
\value{
Object with slot \code{VarNameCorresp} updated.
}
\description{
\code{setVNC} assigns a list of \linkS4class{data.table}s to the
 slot \code{VarNameCorresp} of the input object.
}
\examples{
# On an object of class DD:
library(data.table)
MicroDataDD <- data.table(Variable = 'IEPEntradaPed', Sort = 'IDDD', Class = 'numeric',
                          Length = '8',
                          Qual1 = 'NumIdEst', Qual2 = 'Market', ValueRegExp = '')
MicroDataDD <- new(Class = 'DDdt', MicroDataDD)
VarList <- list(MicroData = new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst','','',''),
                                                          NonIDQual = c('', 'Market', 'Cod', ''),
                                                          IDDD = c('', '', '' ,'IEPEntradaPed'),
                                                          NumIdEst = c('', '', '', '.'),
                                                          Market = c('', '', '', '1.'),
                                                          Cod = rep('', 4),
                                                          Unit1 = c('', '', '', 'cp02'))))
VNC <- BuildVNC(VarList)
DD <- new(Class = 'DD', VarNameCorresp = VNC, MicroData = MicroDataDD)

VarListAdd <- list(MicroData = new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst','','',''),
                                                          NonIDQual = c('','Market','Cod', ''),
                                                          IDDD = c('', '', '' , 'IEPEntradaPed'),
                                                          NumIdEst = c('', '', '', '.'),
                                                          Market = c('', '', '', '2.'),
                                                          Cod = rep('', 4),
                                                          Unit1 = c('', '', '', 'cp02'))))
VNCAdd <- BuildVNC(VarListAdd)
setVNC(DD) <- VNCAdd
DD

# On an object of class StQ:
library(data.table)
MicroDataDD <- data.table(Variable = 'IEPEntradaPed', Sort = 'IDDD', Class = 'numeric',
                          Length = '8', Qual1 = 'NOrden', ValueRegExp = '')
MicroDataDD <- new(Class = 'DDdt', MicroDataDD)
VarList <- list(MicroData = new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst','','',''),
                                                          NonIDQual = c('', 'Market', 'Cod', ''),
                                                          IDDD = c('', '', '' ,'IEPEntradaPed'),
                                                          NumIdEst = c('', '', '', '.'),
                                                          Market = c('', '', '', '1.'),
                                                          Cod = rep('', 4),
                                                          Unit1 = c('', '', '', 'cp02'))))
VNC <- BuildVNC(VarList)
DD <- new(Class = 'DD', VarNameCorresp = VNC, MicroData = MicroDataDD)  
StQ <- new(Class = 'StQ', Data = new(Class = 'Datadt',
                                     data.table(IDDD = character(0), Value = character(0))),
                          DD = DD)
             
VarListAdd <- list(MicroData = new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst','','',''),
                                                          NonIDQual = c('','Market','Cod', ''),
                                                          IDDD = c('', '', '' , 'IEPEntradaPed'),
                                                          NumIdEst = c('', '', '', '.'),
                                                          Market = c('', '', '', '2.'),
                                                          Cod = rep('', 4),
                                                          Unit1 = c('', '', '', 'cp02'))))
VNCAdd <- BuildVNC(VarListAdd)                                                    
setVNC(StQ) <- VNCAdd
getDD(StQ)

}

