% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.rawStQ.R
\docType{methods}
\name{show,rawStQ-method}
\alias{show,rawStQ-method}
<<<<<<< HEAD
\title{Show an object of class \linkS4class{StQ}}
=======
\title{Show an object of class \linkS4class{rawStQ}}
>>>>>>> 5034523f22c62817420f2f5687369d62b4523cd8
\usage{
\S4method{show}{rawStQ}(object)
}
\arguments{
<<<<<<< HEAD
\item{object}{Object of class \linkS4class{StQ}.}
=======
\item{object}{Object of class \linkS4class{rawStQ}.}
>>>>>>> 5034523f22c62817420f2f5687369d62b4523cd8
}
\value{
Invisible \code{\link{NULL}}.
}
\description{
<<<<<<< HEAD
\code{show} displays the slot \code{Data} of the input \linkS4class{StQ} object limiting the
number of columns on screen up to 8.
}
\details{
This method displays only the content of slot \code{Data} from the input \linkS4class{StQ} 
object. It is indeed the method \link[methods]{show} adapted to class \linkS4class{StQ}.
}
\examples{
# A trivial example
show(new(Class = 'StQ'))
data(ExampleStQ)
show(ExampleStQ)
ExampleStQ
=======
\code{show} displays the slot \code{Data} of the input \linkS4class{rawStQ}.
}
\details{
This method displays only the content of slot \code{Data} from the input \linkS4class{rawStQ}
object. It is indeed the method \link[methods]{show} adapted to class \linkS4class{rawStQ}.
}
\examples{
# A trivial example
show(new(Class = 'rawStQ'))

show(ExamplerawStQ)
>>>>>>> 5034523f22c62817420f2f5687369d62b4523cd8

}

