% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.rawStQ.R
\docType{methods}
\name{show,rawStQ-method}
\alias{show,rawStQ-method}
\title{Show an object of class \linkS4class{StQ}}
\usage{
\S4method{show}{rawStQ}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{StQ}.}
}
\value{
Invisible \code{\link{NULL}}.
}
\description{
\code{show} displays the slot \code{Data} of the input \linkS4class{StQ} object limiting the
number of columns on screen up to 8.
}
\details{
This method displays only the content of slot \code{Data} from the input \linkS4class{StQ} 
object. It is indeed the method \link[methods]{show} adapted to class \linkS4class{StQ}.
}
\examples{
# A trivial example
show(new(Class = 'StQ'))
data(ExampleStQ)
show(ExampleStQ)
ExampleStQ

}
