% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.VarNameCorresp.R
\docType{methods}
\name{show,VarNameCorresp-method}
\alias{show,VarNameCorresp-method}
\title{Show an object of class \linkS4class{VarNameCorresp}}
\usage{
\S4method{show}{VarNameCorresp}(object)
}
\arguments{
\item{Object}{of class \linkS4class{VarNameCorresp}.}
}
\value{
Invisible object of class \code{\link{NULL}}.
}
\description{
The method \code{show} shows the slots of an object
\linkS4class{VarNameCorresp} limiting the number of columns on screen up to 
10.

It is indeed the method \link[methods]{show} adapted to the class
\linkS4class{VarNameCorresp}.
}
\examples{
# A trivial example
show(new(Class = 'VarNameCorresp'))

# A more elaborate example
VarList <- list(MicroData = data.table(IDQual = c('NumIdEst','','', '', '', '', ''),
                NonIDQual = c('', 'IsNatMarket', 'IsEuroMarket', 'IsRWMarket', '', '', ''),
                IDDD = c('','','','','Turnover', 'Turnover', 'Turnover'),
                NumIdEst = c('', '', '', '', '.', '.', '.'),
                IsNatMarket = c('', '', '', '', '1', '0', '0'),
                IsEuroMarket = c('', '', '', '', '', '1', '0'),
                IsRWMarket = c('', '', '', '', '', '', '1'),
                Unit1 = c('','','','','cn01', 'cn02', 'cn03')))
Example <- new(Class = 'VarNameCorresp', VarList)
show(Example)

}

