% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/show.VarNameCorresp.R
\docType{methods}
\name{show,VarNameCorresp-method}
\alias{show,VarNameCorresp-method}
\title{Show an object of class \linkS4class{VarNameCorresp}}
\usage{
\S4method{show}{VarNameCorresp}(object)
}
\arguments{
\item{Object}{of class \linkS4class{VarNameCorresp}.}
}
\value{
Invisible object of class \code{\link{NULL}}.
}
\description{
The method \code{show} shows the slots of an object
\linkS4class{VarNameCorresp} limiting the number of columns on screen up to 8.

It is indeed the method \link[methods]{show} adapted to the class
\linkS4class{VarNameCorresp}.
}
\examples{
# A trivial example
show(new(Class = 'VarNameCorresp'))

# A more elaborate example. See StQ-class and method getDD for details
VarList <- list(MicroData = data.table(IDQual = c('NumIdEst','','','',''),
               NonIDQual = c('EsMercNac', 'EsMercEuro', 'EsMercRM','Cod',''),
               IDDD = c('','','','','IEPEntradaPed'),
               Unit1 = c('','','','','')))
Example <- new(Class = 'VarNameCorresp', VarNameCorresp = VarList)
show(Example)
}

