% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.VNCdt.R
\docType{methods}
\name{show,VNCdt-method}
\alias{show,VNCdt-method}
\title{Show an object of class \linkS4class{VNCdt}}
\usage{
\S4method{show}{VNCdt}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{VNCdt}.}
}
\value{
Invisible object of class \code{\link{NULL}}.
}
\description{
The method \code{show} shows the slots of an object \linkS4class{VNCdt} limiting the
number of columns on screen up to 8.

It is indeed the method \link[methods]{show} adapted to the class \linkS4class{VNCdt}.
}
\examples{
# A trivial example
show(new(Class = 'VNCdt'))

library(data.table)
VNCdt1 <- new(Class = 'VNCdt',
              data.table(IDQual = c('NumIdEst', '', '', ''),
                         NonIDQual = c('', 'Market', 'Cod', ''),
                         IDDD = c('', '', '','NewOrders'),
                         NumIdEst = c('', '', '', '.'),
                         Market = c( '', '', '','0'),
                         Cod = c('', '', '', ''),
                         UnitName = c( '', '', '','cp09'),
                         InFiles = rep('FF', 4)))
show(VNCdt1)

VNCdt2 <- new(Class = 'VNCdt', 
              data.table(IDQual = c('NOrden', '', ''), 
                         NonIDQual = c('', 'Market', ''), 
                         IDDD = c('', '', 'Turnover'),
                         NOrden = c('', '', '.'),
                         Market = c('', '', '1.'),
                         Import = c('', '', '0'),
                         Geo = c('', '', '2.'),
                         UnitName = c('', '', 'cn01'),
                         InFiles = rep('FF', 3)))
show(VNCdt2)

}

