% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/show.VNCdt.R
\docType{methods}
\name{show,VNCdt-method}
\alias{show,VNCdt-method}
\title{Show an object of class \linkS4class{VNCdt}}
\usage{
\S4method{show}{VNCdt}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{VNCdt}.}
}
\value{
Invisible object of class \code{\link{NULL}}.
}
\description{
The method \code{show} shows the slots of an object
\linkS4class{VNCdt} limiting the number of columns on screen up to
10.

It is indeed the method \link[methods]{show} adapted to the class
\linkS4class{VNCdt}.
}
\examples{
# A trivial example
show(new(Class = 'VNCdt'))

library(data.table)
VNCdt1 <- new(Class = 'VNCdt',
              data.table(IDQual = c('NumIdEst', '', '', '', '', '', '', ''),
                         NonIDQual = c('', 'EsMercNac', 'EsMercEuro', 'EsMercRM',
                                                     'Cod','Var1', 'Var2', ''),
                         IDDD = c('', '', '', '', '', '', '','IEPEntradaPed'),
                         NumIdEst = c('', '', '', '', '', '', '', '.'),
                         EsMercNac = c('', '', '', '', '', '', '','0'),
                         EsMercEuro = c('', '', '', '', '', '', '','0'),
                         EsMercRM = c('', '', '', '', '', '', '','1'),
                         Cod = c('', '', '', '', '', '', '', ''),
                         Var1 = c('', '', '', '', '', '', '', ''),
                         Var2 = c('', '', '', '', '', '', '', ''),
                         Unit1 = c('', '', '', '', '', '', '','cp09')))


VNCdt2 <- new(Class = 'VNCdt',
              data.table(IDQual = c('NOrden', '', '', ''),
                         NonIDQual = c('', 'IsNatMarket', '', ''),
                         IDDD = c('', '', 'Turnover', 'Turnover'),
                         NOrden = c('', '', '.', '.'),
                         IsNatMarket = c('', '', '0', '1'),
                         Unit1 = c('', '', 'cn01', 'cn02')))
}

