% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StQ-class.R
\docType{class}
\name{StQ-class}
\alias{StQ-class}
\title{S4 class for sets of \emph{St}andarized \emph{Q}uestionnaires.}
\description{
Definition of an S4 class named \code{StQ} for sets of
standardized questionnaires.

The structure of the class \code{StQ} comprises 2 attribute:
\itemize{
\item The attribute \code{Data}, which is an object of class
\linkS4class{Datadt}, that is, a \linkS4class{data.table} with at least the
columns \code{IDDD} and \code{Value}.

\item The attribute \code{DD}, which is an object of class \linkS4class{DD}.
It basically contains the definition and properties of each variable.
}

Every variable name in the attribute \code{Data} must be present in the
attribute \code{DD}.
}
\section{Slots}{

\describe{
\item{\code{Data}}{Object of class \linkS4class{Datadt}. It must have at least two
columns: \code{IDDD} and \code{Value}. It contains all statistical variables
(including some metadata) together with their corresponding values. If
\code{Data} is not specified as an input parameter, an empty
\linkS4class{Datadt} object with columns \code{IDDD} and \code{Value} will be
initiated.}

\item{\code{DD}}{Object of class \linkS4class{DD} with the definition and properties
of all variables. If \code{DD} is not specified as an input parameter, an
empty \linkS4class{DD} object with columns \code{Variable}, \code{Sort},
\code{Class}, \code{Qual1} and \ code{ValueRegExp} will be initiated.}
}}
\examples{
# An empty standardized questionnaire set:
new(Class = 'StQ')


library(data.table)
data(ExampleDD)
data(ExampleDatadt)
Q <- new(Class = 'StQ', Data = ExampleDatadt, DD = ExampleDD)
Q
# Notice that only the slot Data appears on screen, but the object is not a Datadt data.table:
str(Q)
}

