% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StQListToStQ.R
\docType{methods}
\name{StQListToStQ}
\alias{StQListToStQ}
\alias{StQListToStQ,StQList-method}
\title{Method to transform an StQList object into a list of StQ's objects.}
\usage{
StQListToStQ(object)

\S4method{StQListToStQ}{StQList}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{StQList} to be transformed.}
}
\value{
a list of objects of class \linkS4class{StQ}.
}
\description{
\code{StQListToStQ} transform an StQList object into a list of StQ's objects with a
new variable "Period" to take in account the interval times related to StQList object.

This method creates a variable with the name \code{Period} in Data slots with the period related
to each Data and adds this variable to each DD slot of the new StQ's objects.
}

