% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.VNCdt.R
\docType{methods}
\name{[,VNCdt-method}
\alias{[,VNCdt-method}
\title{Extract parts of an object of class \linkS4class{VNCdt}}
\usage{
\S4method{[}{VNCdt}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{Object of class \linkS4class{VNCdt}.}

\item{i, j, ...}{Indices corresponding to elements to be extracted. The indices are numeric or
character vectors, \code{\link{missing}} or \code{\link{NULL}}. Numeric values are coerced to
\code{integer} with \code{\link{as.integer}} (thus truncated to zero).}

\item{drop}{Included by coherence.}
}
\value{
Object of class \linkS4class{VNCdt} with the subsetted input object.
}
\description{
\code{[} extracts parts of an object of class \linkS4class{VNCdt}.

It is indeed the method \code{[} for the class \linkS4class{VNCdt}.
}
\examples{
library(data.table)
MicroData = new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst', rep('', 2)),
                                            NonIDQual = c('', 'Market', ''),
                                            IDDD = c(rep('', 2), 'NewOrders'),
                                            NumIdEst = c(rep('', 2), '.'),
                                            Market = c(rep('', 2), '1.'),
                                            UnitName = c('numidest', '', 'cp09'),
                                            InFiles = rep('FF', 3)))
MicroData[IDDD == 'NewOrders']


}

