% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarNameCorresp-class.R
\docType{class}
\name{VarNameCorresp-class}
\alias{VarNameCorresp-class}
\title{S4 class for the correspondence between variable names}
\description{
Definition of an S4 class named \linkS4class{VarNameCorresp} with the correspondence 
between variable names used by the different production unit for a given statistical variable and
variable names according to the key-value pair data model with qualifiers given by statistical 
metadata.

The class \linkS4class{VarNameCorresp} comprises a \linkS4class{list} whose components are 
\linkS4class{data.table}s with a row per each variable and the following columns:

\itemize{
 \item \code{IDQual}: Names of unit qualifiers.
 \item \code{NonIDQual}: Names of variable names qualifiers.
 \item \code{IDDD}: Names of the variables.
 \item \emph{One column per each name of the \code{IDQual} qualifiers}: 
 Values of the corresponding \code{IDQual} Qualifier for each variable (row).
 \item \emph{One column per each name of the \code{NonIDQual} qualifiers}: 
 Values of the corresponding \code{NonIDQual} Qualifier for each variable 
 (row).
 \item \code{UnitName}: Variable name used in production.
 \item \code{InFiles}: code of the files where each variable is included.
}
}
\examples{
library(data.table)
VarList <- list(ID = new(Class = 'VNCdt',
                data.table(IDQual = c('NumIdEst', rep('', 4)),
                           NonIDQual = c('','','','',''),
                           IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                           NumIdEst = c('', rep('.', 4)),
                           UnitName = c('numidest', 'nombre', 'apellidos', 'direccion', 'telefono'),
                           InFiles = rep('FI', 5))),
MicroData =new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst', rep('', 2)),
                                           NonIDQual = c('', 'Market', ''),
                                           IDDD = c(rep('', 2), 'NewOrders'),
                                           NumIdEst = c(rep('', 2), '.'),
                                           Market = c(rep('', 2), '1'),
                                           UnitName = c('numidest', '', 'cp09'),
                                           InFiles = rep('FF, FD, FG', 3))),
ParaData = new(Class = 'VNCdt', data.table(IDQual = c('NumIdEst', rep('', 2)),
                                           NonIDQual = c('', 'Action', ''),
                                           IDDD = c(rep('', 2), 'Date'),
                                           NumIdEst = c(rep('', 2), '.'),
                                           Action = c(rep('', 2), 'Imputation'),
                                           UnitName = c('numidest', '', 'FechaImput'),
                                           InFiles = rep('FP', 3))))

VNC <- new(Class = 'VarNameCorresp', VarList)
VNC

}

