% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarNameCorresp-class.R
\docType{class}
\name{VarNameCorresp-class}
\alias{VarNameCorresp-class}
\title{S4 class for the variable names correspondence}
\description{
Definition of an S4 class named \code{VarNameCorresp} with the 
correspondence between variable names used by different production units for
a given statistical variable.

The class \code{VarNameCorresp} comprises a \linkS4class{list} whose 
components are \code{data.table}s with a row per each variable and the
following columns:

\itemize{
 \item \code{IDQual}: Names of unit qualifiers.
 \item \code{NonIDQual}: Names of variable names qualifiers.
 \item \code{IDDD}: Names of the variables.
 \item \emph{One column per each name of the \code{IDQual} qualifiers}: 
 Values of the corresponding \code{IDQual} Qualifier for each variable (row).
 \item \emph{One column per each name of the \code{NonIDQual} qualifiers}: 
 Values of the corresponding \code{NonIDQual} Qualifier for each variable 
 (row).
 \item \code{UnitName}: Variable name used in production.
 \item \code{InFiles}: code of the files where each variable is included.
}
}
\examples{
library(data.table)
VarList <- list(
  ID = new(Class = 'VNCdt', 
           .Data = data.table(
                 IDQual = c('NumIdEst', rep('', 4)),
                 NonIDQual = rep('', 5),
                 IDDD = c('', 'Name', 'Surname', 'PostalAddr', 'PhoneNo'),
                 NumIdEst = c('', rep('.', 4)),
                 UnitName = c('numidest', 'nombre', 'apellidos', 'direccion', 'telefono'),
                 InFiles = rep('FI', 5))),
  MicroData = new(Class = 'VNCdt', 
                  .Data = data.table(
                      IDQual = c('NumIdEst', rep('', 4)),
                      NonIDQual = c('', 'IsNatMarket', 'IsEuroMarket', 'IsRWMarket', ''),
                      IDDD = c(rep('', 4), 'NewOrders'),
                      NumIdEst = c(rep('', 4), '.'),
                      IsNatMarket = c(rep('', 4), '0'),
                      IsEuroMarket = c(rep('', 4), '0'),
                      IsRWMarket = c(rep('', 4), '1'),
                      UnitName = c('numidest', rep('', 3), 'cp09'),
                      InFiles = rep('FF, FD, FG', 5))),
 ParaData = new(Class = 'VNCdt'),
 Aggregates = new(Class = 'VNCdt', 
                  .Data = data.table(
                     IDQual = c('Province', 'NACE', 'IsNatMarket', ''),
                     NonIDQual = rep('', 4),
                     IDDD = c('', '', '', 'TotalTurnover'),
                     Province = c('', '', '', '.'),
                     NACE = c('', '', '', '.'),
                     IsNatMarket = c('', '', '', '1'),
                     UnitName = c('provincia', 'actividad', '', 'cn01'),
                     InFiles = rep('FA', 4))))
new(Class = 'VarNameCorresp', .Data = VarList)

}

