% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarNameCorresp-class.R
\docType{class}
\name{VarNameCorresp-class}
\alias{VarNameCorresp-class}
\title{S4 class for the variable names correspondence}
\description{
Definition of an S4 class named \code{VarNameCorresp} with the 
correspondence between variable names used by different production units for
a given statistical variable.

The class \code{VarNames} comprises a slot of class \linkS4class{list} whose
components are data.tables with a row per each variable and the following 
columns:

\itemize{
 \item \code{IDQual}: Names of unit qualifiers.
 \item \code{NonIDQual}: Names of variable names qualifiers.
 \item \code{IDDD}: Names of the variables.
 \item \emph{One column per each name of the \code{IDQual} qualifiers}: 
 Values of the corresponding \code{IDQual} Qualifier for each variable (row).
 \item \emph{One column per each name of the \code{NonIDQual} qualifiers}: 
 Values of the corresponding \code{NonIDQual} Qualifier for each variable 
 (row).
 \item \code{ProdUnit}\emph{j}: Variable name used by production Unit\emph{j}.
}
}
\section{Slots}{

\describe{
\item{\code{VarNameCorresp}}{\linkS4class{list} with as many data.tables as tables of
variable names correspondences.}
}}
\examples{
library(data.table)
VarList <- list(data.table(IDQual = c('NumIdEst','','','','',''),
                     NonIDQual = c('', 'EsMercNac', 'EsMercEuro', 'EsMercRM',
                                   'Cod',''),
                     IDDD = c('','','','','', 'IEPEntradaPed'),
                     Unit1 = c('','','','','','')))
new(Class = 'VarNameCorresp', VarNameCorresp = VarList)

}

