% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarNamesToDD.R
\name{VarNamesToDD}
\alias{VarNamesToDD}
\title{Return a reduced \link{DD} involving the input variable names}
\usage{
VarNamesToDD(VarNames, DD)
}
\arguments{
\item{VarNames}{Character vector with the compound variable names (IDDD names).}

\item{DD}{Object of class \link{DD} with the definition and properties of the variables.}
}
\value{
An object of class \link{DD} involving only the compound variable names specified in the input 
parameter \code{VarNames}.
}
\description{
\code{VarNamesToDD} returns a \link{DD} object including only the variable 
names specified in the input parameter \code{VarNames}.
}
\examples{
VarNames <- c('Turnover')
data(ExampleDD)          
VarNamesToDD(VarNames, ExampleDD)
 
VarNames <- c('Turnover', 'Employees', 'NACE09', 'Stocks')
VarNamesToDD(VarNames, ExampleDD)                     

}
